% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{approved_list}
\alias{approved_list}
\title{List of verified records to exempted from future quality check}
\format{
A list of 4 data frames:
\itemize{
   \item{Brood_approved_list}
   \item{Capture_approved_list}
   \item{Individual_approved_list}
   \item{Location_approved_list}
}

Each data frame has 3 variables:
\itemize{
   \item{PopID, unique identifier for population.}
   \item{A unique identifier for records in each particular data frame (i.e., BroodID, CaptureID, IndvID, LocationID).}
   \item{CheckID, unique identifier for individual quality check.}
}
}
\description{
List of records that were flagged as warning/error but have been verified by the data owner to be correct.
}
