% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\title{Archive SPI-Birds raw data}
\usage{
archive(
  data_folder = choose_directory(),
  update_type = "major",
  institutionID,
  new_data_path,
  new_data_date = NULL,
  initial = FALSE
)
}
\arguments{
\item{data_folder}{Path. Folder containing all current SPI-Birds data (N drive)}

\item{update_type}{Character string. Have we received a new year of data ("major")
or data with bug/error fixes ("minor")?}

\item{institutionID}{Character string. 3-4 letter-code for data owner.
Should be the same as the folder name where data are saved.}

\item{new_data_path}{File path. Updated data to use as current data in the system.
Where a site uses multiple files provide a list of file paths.}

\item{new_data_date}{Date at which archiving should be recorded (i.e. the name of the folder
where archived data will be stored.)}

\item{initial}{Logical. Are we creating the initial archiving information? If TRUE, will create
a MetaData.txt file and copy all information into an archive folder.}
}
\value{
Returns nothing.
}
\description{
Archive SPI-Birds raw data
}
