% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{brood_check}
\alias{brood_check}
\title{A wrapper function to perform quality checks on brood data}
\usage{
brood_check(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
A list of 3 items: a summary data frame of all checks, a record-by-record list of warnings, and a record-by-record list of errors.
}
\description{
A wrapper that runs all single checks related to `Brood_data`.
}
\details{
\strong{Brood check 1}: Check if the formats of each column in `Brood_data` match with the standard format using \code{\link{check_format_brood}}.
\strong{Brood check 2}: Compare clutch size and brood size per brood using \code{\link{compare_clutch_brood}}.
\strong{Brood check 3}: Compare brood size and fledgling number per brood using \code{\link{compare_brood_fledglings}}.
}
