% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\docType{data}
\name{Brood_data}
\alias{Brood_data}
\title{Quality check dummy data}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1 rows and 31 columns.}
\usage{
Brood_data
}
\description{
Create dummy pipeline output (\code{dummy_data}) to be tested in \code{\link{quality_check}}.
Each row in \code{dummy_data} corresponds to a single check in \code{\link{quality_check}}.
}
\keyword{datasets}
