% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{brood_data_template}
\alias{brood_data_template}
\title{Brood data template in standard format (version 1.1.0)}
\format{
Data frame with 1 row and 36 variables:
}
\description{
Brood data template in standard format (version 1.1.0)
}
