% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{brood_ref_values_list}
\alias{brood_ref_values_list}
\title{List of expected values for brood data.}
\format{A list with 2 items.
\describe{
 \item{PARMAJ}{Expected values of clutch size, brood size and number of fledglings for great tits.}
 \item{FICALB}{Expected values of clutch size, brood size and number of fledglings for collared flycatchers.}}}
\description{
Plausible brood data values that can be used to detect outliers.
}
