% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Calculate age based on when an individual was first captured}
\usage{
calc_age(data, ID, Age, Date, Year)
}
\arguments{
\item{data}{Data frame. Data frame with capture information.}

\item{ID}{Unquoted expression (i.e. character without quotation marks). Name
of column with individual identity.}

\item{Age}{Unquoted expression (i.e. character without quotation marks). Name
of column with observed age of individual in each capture. Must be in
EURING codes.}

\item{Date}{Unquoted expression (i.e. character without quotation marks).
Name of column with CaptureDate information. Must be in format dd/mm/yyyy.}

\item{Year}{Unquoted expression (i.e. character without quotation marks).
Name of column with year information. N.B. This could be different to
CaptureDate if we are dealing with species that breed over two year (e.g.
Southern Hemisphere species).}
}
\value{
A data frame with calculated age.
}
\description{
Arrange data by individual and year and then determine potential age in each
capture.
}
\details{
We only consider whether an individual was captured as a chick or not. We
don't consider the adult age when an individual was first captured (e.g.
EURING 4 or 5) This prevents any cases where an individual might be wrongly
aged at first capture.
}
\examples{
set.seed(666)
bird_data <- tibble::tibble(IndvID = LETTERS[sample(1:26, 100, replace = TRUE)],
SampleYear = sample(2012:2016, 100, replace = TRUE),
CaptureDate = as.Date(paste(SampleYear, 3, 31, sep = "-"), format = "\%Y-\%m-\%d") + sample(1:30, 100, replace = TRUE),
Age_obsv = sample(c(1, 4), 100, replace = TRUE)) \%>\%
calc_age(ID = IndvID, Age = Age_obsv, Date = CaptureDate, Year = SampleYear)
}
