% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_birdNBuse.R
\name{calc_birdNBuse}
\alias{calc_birdNBuse}
\title{Calculate first and last breeding season when a nestbox was used by birds}
\usage{
calc_birdNBuse(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Data frame with standardized brood data.}
}
\value{
A data frame with the NestBoxID and corresponding FirstSeason and LastSeason.
}
\description{
Currently, the variables FirstSeason and LastSeason in Location data are
defined as first and last breeding seasons data collectors put out a nestbox.
However, this data may often not be available, in which case it may be useful
to know instead in which breeding season a nestbox was first and last used by
breeding birds.
}
\details{
This function uses information from the standardized Brood data table to
calculate the first and last breeding season when each nestbox was used by birds

For the moment, if a nestbox was used in the latest year of data available, this
year is returned in LastSeason (which differs from the definition in the protocol,
which states that LastSeason for still actively used nestboxes should be NA).

The function assumes that NestBoxID = LocationID.
}
\examples{
library(dplyr)
set.seed(666)
Brood_data <- tibble::tibble(BroodID = c(1:10), LocationID = LETTERS[c(1,3,1,1,2,3,1,2,1,2)],
                             BreedingSeason = sort(sample(2012:2016, 10, replace = TRUE)))
}
