% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clutchtype.R
\name{calc_clutchtype}
\alias{calc_clutchtype}
\title{Calculate clutch type (when laying date is in date format)}
\usage{
calc_clutchtype(data, na.rm = T)
}
\arguments{
\item{data}{Brood data to use for conversion}

\item{na.rm}{Should NA's be treated as 0s or true unknowns?}
}
\value{
A character vector with either 'first', 'replacement', 'second', or
  NA
}
\description{
Use info on laying date and fledge data to determine the clutch type of a
given brood
}
\examples{
#Create fake dataset
set.seed(666)
dat <- tibble::tibble(PopID = "TEST", Species = "PARMAJ",
FemaleID = sample(LETTERS[1:7], size = 100, replace = TRUE),
NumberFledged = rpois(n = 100, lambda = 1),
#Create 100 fake broods
SampleYear = sample(c(seq(2000, 2012, 1)), size = 100, replace = TRUE),
LayingDate = as.Date(paste(SampleYear,
                           sample(c(4, 5, 6), size = 100, replace = TRUE),
                           sample(seq(1, 31, 1), size = 100, replace = TRUE), sep = "-"),
                           format = "\%Y-\%m-\%d"))
calc_clutchtype(data = dat, na.rm = TRUE)
}
