% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_cumfledge}
\alias{calc_cumfledge}
\title{Calculate cumulative number of fledgings}
\usage{
calc_cumfledge(x, na.rm = TRUE)
}
\arguments{
\item{x}{Column observedNumberFledged in the Brood_data table}

\item{na.rm}{Logical. If TRUE, returns cumulative number of fledglings
where NA is assumed to be 0. If FALSE, returns a vector of logicals
showing whether any nests before the current nest were unmeasured (NA).}
}
\value{
A vector of numeric values (if na.rm = TRUE) or logical values (na.rm = FALSE).
}
\description{
For a given nest, determine the cumulative number of fledglings in all nests
before this. This is used to calculate calculatedClutchType using \code{\link{calc_clutchtype}}.
The function also includes functionality to return whether number of fledglings in previous
nests was measured or not (i.e., were there NAs in any nests before the
current one). This is necessary in populations where the number of fledglings
is not measured consistently.
}
\examples{
#Assuming NA is 0
#Return vector of numerics.
calc_cumfledge(x = c(1, 3, NA, 4), na.rm = TRUE)

#Do not assume NA is 0.
#Return a vector of logicals showing whether an NA occurred before
#the current record.
calc_cumfledge(x = c(1, 3, NA, 4), na.rm = FALSE)
}
