% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_nestattempt}
\alias{calc_nestattempt}
\title{Calculate sequence of nest attempts by a breeding pair in a season}
\usage{
calc_nestattempt(data, season)
}
\arguments{
\item{data}{Data frame with brood information.}

\item{season}{Unquoted expression (i.e. character without quotation marks). Name of the variable in primary data that marks the
breeding seasons according to the data owner. This information is necessary if a single breeding season spreads over
two calendar years, or if multiple seasons occur in a single calendar year.}
}
\value{
A data frame with nestAttemptNumber.
}
\description{
Arrange data by breeding pairs (femaleID & maleID) and laying date to
calculate the sequence of nest attempts in a single season
}
\details{
If either female or male was not identified, the nest is assumed to be the pair's first attempt.
}
\examples{
library(dplyr)

# Single season in a single calendar year (no season information needed)
bird_data <- tibble::tibble(broodID = 1:7,
                            femaleID = c("F1", "F1", "F2", "F2", "F3", "F3", NA),
                            maleID = c("M1", "M1", "M2", "M2", "M3", NA, "M3"),
                            observedLayYear = c(2020, 2020, 2020, 2021, 2022, 2022, 2022),
                            observedLayMonth = rep(5, 7),
                            observedLayDay = rep(1, 7)) \%>\%
  calc_nestattempt()

# Single season over two calendar years (season information needed)
bird_data <- tibble::tibble(broodID = 1:7,
                            femaleID = c("F1", "F1", "F2", "F2", "F3", "F3", NA),
                            maleID = c("M1", "M1", "M2", "M2", "M3", NA, "M3"),
                            season = c(1, 1, 1, 2, 3, 3, 3),
                            observedLayYear = c(2020, 2021, 2020, 2021, 2022, 2022, 2022),
                            observedLayMonth = c(12, 1, rep(12, 5)),
                            observedLayDay = rep(1, 7)) \%>\%
  calc_nestattempt(season = season)
}
