% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_season}
\alias{calc_season}
\title{Determine the breeding season in which a brood was laid}
\usage{
calc_season(data, season)
}
\arguments{
\item{data}{Data frame with brood information.}

\item{season}{Unquoted expression (i.e. character without quotation marks). Name of the variable in primary data that marks the
breeding seasons according to the data owner. This information is necessary if a single breeding season spreads over
two calendar years, or if multiple seasons occur in a single calendar year.}
}
\value{
A data frame with breedingSeason.
}
\description{
Determine the breeding season based on the data owner's expertise on the breeding biology of the species at the study site.
}
\details{
The breeding season is determined as follows:
\itemize{
  \item{Species with 1 breeding season per calendar year} \itemize{
    \item{Breeding season falls completely within a single calendar year: \code{breedingSeason} = calendar year. For example, the breeding season from April 2022 to June 2022 is marked "2022".}
    \item{Breeding season is spread over two calendar years: \code{breedingSeason} = calendar year in which the season \emph{starts}. For example, the breeding season from December 2021 to February 2022 is marked "2021".}
  }
  \item{Species with >1 season per calendar year} \itemize{
    \item{All breeding seasons fall completely within a single calendar year: \code{breedingSeason} = <calendar year>_<season number>. For example, the first breeding season (from April 2022 to May 2022) is marked "2022_1", and the second breeding season (from August 2022 to September 2022) is marked "2022_2".}
    \item{Not all breeding seasons fall completely within a single calendar year: \code{breedingSeason} = <calendar year>_<season number>, where calendar year is the year in which the \emph{first} season \emph{starts}. For example, the first breeding season (from October 2021 to November 2021) is marked "2021_1", and the second breeding season (from December 2021 to January 2022) is marked "2021_2".}
  }
}
}
\examples{
# 1 breeding season per calendar year
dat <- tibble::tibble(siteID = "AAA",
                      speciesID = "PARMAJ",
                      femaleID = LETTERS[1:10],
                      observedLayYear = c(rep(2020, 5), rep(2021, 5)))

calc_season(data = dat)

# 1 breeding season over 2 calendar years
dat <- tibble::tibble(siteID = "AAA",
                      speciesID = "PARMAJ",
                      femaleID = LETTERS[1:10],
                      observedLayYear = c(rep(2020, 3), rep(2021, 5), rep(2022, 2)),
                      season = rep(1:2, each = 5))

calc_season(data = dat, season = "season")

# 2 breeding seasons per calendar year
dat <- tibble::tibble(siteID = "AAA",
                      speciesID = "PARMAJ",
                      femaleID = LETTERS[1:12],
                      observedLayYear = c(rep(2020, 6), rep(2021, 6)),
                      season = rep(1:4, each = 3))

calc_season(data = dat, season = "season")
}
