% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_sex}
\alias{calc_sex}
\title{Determine sex of an individual}
\usage{
calc_sex(individual_data, capture_data)
}
\arguments{
\item{individual_data}{Data frame with individual information.}

\item{capture_data}{Data frame with capture information.}
}
\value{
A data frame (individual information) with calculatedSex, which takes either 'F', 'M', 'C', or NA.
}
\description{
Conclude the sex of an individual based on sex scores during captures (observedSex).
}
\details{
An individual is recorded as having a conflicting sex ('C') when it has been sexed differently during
different captures.
}
\examples{
library(dplyr)

capture_data <- tibble::tibble(individualID = rep(paste0("I", 1:4), each = 2),
                               observedSex = c("F", "F", "M", NA, "F", "M", NA, NA))

individual_data <- tibble::tibble(individualID = paste0("I", 1:4)) \%>\%
  calc_sex(., capture_data)
}
