% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chick_mass_cutoffs.R
\name{calculate_chick_mass_cutoffs}
\alias{calculate_chick_mass_cutoffs}
\title{Determine chick growth curves}
\usage{
calculate_chick_mass_cutoffs(HOG_chick_data)
}
\arguments{
\item{HOG_chick_data}{Data frame with chick age and mass of Hoge Veluwe great and blue tits}
}
\value{
A data frame with cutoffs values at each age
}
\description{
Fit logistic and gompertz growth curves
to chick mass data for great and blue tits.
}
\details{
We fit this using data from Hoge Veluwe. See example for data extraction.
}
\examples{
\dontrun{
#run pipeline for Hoge Veluwe great and blue tits
HOG <- run_pipelines(PopID = "HOG", Species = c("PARMAJ", "CYACAE"), output_type = "R")

Subset relevant columns
HOG_chick_data <- HOG$Capture_data \%>\%
dplyr::filter(!is.na(ChickAge) & !is.na(Mass), BreedingSeason <= 2019) \%>\%
dplyr::select(ChickAge, Mass, Species)

calculate_chick_mass_cutoffs(HOG_chick_data)
}
}
