% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{capture_check}
\alias{capture_check}
\title{Perform quality checks on capture data}
\usage{
capture_check(
  Capture_data,
  Location_data,
  Brood_data,
  check_format = TRUE,
  approved_list
)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{Location_data}{Data frame. Location data output from pipeline.}

\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{check_format}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the check on variable format (i.e. \code{\link{check_format_capture}}) is included in the quality check. Default: \code{TRUE}.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
A wrapper that runs all single checks related to \code{Capture_data}.
}
\details{
The following capture data checks are performed:
\itemize{
\item \strong{C1}: Check if the formats of each column in \code{Capture_data} match with the standard format using \code{\link{check_format_capture}}.
\item \strong{C2a-b}: Check capture variable values against reference values using \code{\link{check_values_capture}}. Capture variables checked for adults and chicks: Mass and Tarsus.
\item \strong{C3}: Check chick age (in numbers of days since hatching) using \code{\link{check_chick_age}}.
\item \strong{C4}: Check that adults caught on nest are listed as parents using \code{\link{check_adult_parent_nest}}.
}
}
