% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{capture_data_template}
\alias{capture_data_template}
\title{Capture data template in standard format (version 1.1.0)}
\format{
Data frame with 1 row and 23 variables:
}
\description{
Capture data template in standard format (version 1.1.0)
}
