% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{capture_ref_values_list}
\alias{capture_ref_values_list}
\title{List of expected values for capture data.}
\format{A list with 2 items.
\describe{
 \item{PARMAJ}{Expected values of mass and tarsus for great tits.}
 \item{FICALB}{Expected values of mass and tarsus for collared flycatchers.}}}
\description{
Plausible capture data values that can be used to detect outliers.
}
