% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{check_adult_parent_nest}
\alias{check_adult_parent_nest}
\title{Check that adults captured on a nest are listed as parents of that nest}
\usage{
check_adult_parent_nest(Capture_data, Location_data, Brood_data)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{Location_data}{Data frame. Location data output from pipeline.}

\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that adults captured on a nest are are listed as the parents of that nest in Brood_data. If not, records will be flagged as a warning. Adults can be caught near a nest box
}
\details{
Check ID: C4.
}
