% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_brood_locations}
\alias{check_brood_locations}
\title{Check that all nest locations in Brood_data appear in Location_data}
\usage{
check_brood_locations(Brood_data, Location_data, approved_list, output)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Location_data}{Data frame. Location data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that all nest locations recorded in Brood_data appear in Location_data. Missing locations will be flagged as a potential error. This check is the opposite of check L2 (\code{\link{check_locations_brood_capture}}).
}
\details{
Check ID: B15.
}
