% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{check_captures_individuals}
\alias{check_captures_individuals}
\title{Check that all individuals captured in Capture_data appear in Individual_data}
\usage{
check_captures_individuals(
  Capture_data,
  Individual_data,
  approved_list,
  output,
  skip
)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that all individuals recorded in Capture_data have a record in Individual_data. Missing individuals should never occur, because Individual_data is usually a direct product of Capture_data (i.e., all unique individuals from Capture_data). Missing individuals will be flagged as a potential error. If there are any missing individuals, the SPI-Birds team needs to check the pipeline code. This check is the opposite of check I5 (\code{\link{check_individuals_captures}}).
}
\details{
Check ID: C5.
}
