% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{check_chick_age}
\alias{check_chick_age}
\title{Check chick age}
\usage{
check_chick_age(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check whether chick ages (in number of days since hatching) are within the range of 0 and 30 days since hatching. Values outside this range will result in an error.
}
