% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_clutch_type_order}
\alias{check_clutch_type_order}
\title{Check clutch type order}
\usage{
check_clutch_type_order(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the order of calculated clutch types per breeding female per season is correct; "first" should always come before "replacement" and "second".
}
\details{
Check ID: B9.
}
