% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_conflicting_species}
\alias{check_conflicting_species}
\title{Check conflicting species}
\usage{
check_conflicting_species(Individual_data, approved_list)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the species of individuals in Individual_data is recorded consistently. Individuals who have been recorded as two different species will have conflicting species ('CONFLICTED') in Individual_data.
}
\details{
Check ID: I5.
}
