% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_format_brood}
\alias{check_format_brood}
\title{Check format of brood data}
\usage{
check_format_brood(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
Check list, warning output, error output.
}
\description{
Check if the formats of each column in the brood data match with the standard format
}
