% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{check_format_capture}
\alias{check_format_capture}
\title{Check format of capture data}
\usage{
check_format_capture(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the format of each column in the capture data match with the standard format.
}
\details{
Check ID: C1.
}
