% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_format_individual}
\alias{check_format_individual}
\title{Check format of individual data}
\usage{
check_format_individual(Individual_data, approved_list)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the format of each column in the individual data match with the standard format.
}
\details{
Check ID: I1.
}
