% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{check_format_location}
\alias{check_format_location}
\title{Check format of location data}
\usage{
check_format_location(Location_data)
}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the format of each column in the location data match with the standard format.
}
\details{
Check ID: L1.
}
