% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_parent_species}
\alias{check_parent_species}
\title{Check parent species}
\usage{
check_parent_species(Brood_data, Individual_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the parents of broods are the same species.
}
\details{
Check ID: B7.
}
