% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funcs.R
\name{check_pop}
\alias{check_pop}
\title{Check that a species entry is correct}
\usage{
check_pop(PopID)
}
\arguments{
\item{PopID}{Vector of 6-letter species codes as recorded in `Species_codes`.}
}
\value{
Vector of 3-letter population codes.
}
\description{
Check that a species entry is correct
}
\examples{
#Codes are correct and are returned
check_pop("AMM")

#Codes are incorrect and throws error
#check_pop("WRONG")
}
