% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_sex_mothers}
\alias{check_sex_mothers}
\title{Check sex of mothers}
\usage{
check_sex_mothers(Brood_data, Individual_data, approved_list)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the individuals listed under FemaleID are female. Individuals with conflicted sex ("C") are ignored and checked in check I3 \code{\link{check_conflicting_sex}}.
}
\details{
Check ID: B12.
}
