% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funcs.R
\name{check_site}
\alias{check_site}
\title{Check that a site entry is correct}
\usage{
check_site(Site)
}
\arguments{
\item{Site}{Vector of 3-letter site codes as recorded in `site_codes`.}
}
\value{
Vector of 3-letter site codes.
}
\description{
Check that a site entry is correct
}
\examples{
#Codes are correct and are returned
check_site("AMM")

#Codes are incorrect and throws error
#check_site("WRONG")
}
