% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_funcs.R
\name{check_species}
\alias{check_species}
\title{Check that a species entry is correct}
\usage{
check_species(Species)
}
\arguments{
\item{Species}{Vector of 6-letter species codes as recorded in `species_codes`.}
}
\value{
Vector of 6-letter species codes.
}
\description{
Check that a species entry is correct
}
\examples{
#Codes are correct and are returned
check_species("PARMAJ")

#Codes are incorrect and throws error
#check_species("WRONG")
}
