% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_unique_IndvID}
\alias{check_unique_IndvID}
\title{Check unique individual identifiers}
\usage{
check_unique_IndvID(Individual_data, approved_list, output, skip)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the individual identifiers (IndvID) are unique. Records with individual identifiers that are not unique within populations will be flagged as a potential error.
}
\details{
Check ID: I1.
}
