% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{check_values_capture}
\alias{check_values_capture}
\title{Check capture variable values against reference values}
\usage{
check_values_capture(Capture_data, var, approved_list)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{var}{Character. Variable to check against reference values.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check variable values against population-species-specific reference values in capture data. Reference values are based on the data if the number of observations is sufficiently large. Records for population-species combinations that are low in number are not evaluated by this check.
}
\details{
\strong{Mass, adults}: Records are considered unusual if they are smaller than the 1st percentile or larger than the 99th percentile, and will be flagged as a warning. Records are considered impossible if they are negative or larger than 4 times the 99th percentile, and will be flagged as an error. Check ID: C2a

\strong{Mass, chicks}: Reference values for chicks are calculated for each age (in days). If the number of observations is sufficiently large (n > 500), this function tries to fit a logistic growth model to determine reference values to each day. If this model fails, reference values are determined per age if the number of observations for chicks of that age is sufficiently large (n > 50). Records are considered unusual if they are smaller than the 1st percentile or larger than the 99th percentile, and will be flagged as a warning. Records are considered impossible if they are negative or larger than 4 times the 99th percentile, and will be flagged as an error. Check ID: C2a

\strong{Tarsus}: : Records are considered unusual if they are smaller than the 1st percentile or larger than the 99th percentile, and will be flagged as a warning. Records are considered impossible if they are negative or larger than 4 times the 99th percentile, and will be flagged as an error. Check ID: C2b
}
