% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{checkID_var}
\alias{checkID_var}
\title{CheckIDs and variable names}
\format{
Data frame with 2 variables:
\describe{
\item{CheckID}{Unique identifier for individual quality check.}
\item{Var}{Variable name.}
}
}
\description{
Data frame to link variable names and CheckIDs to be used in whitelisting of previously flagged records.
}
