% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checks_capture_params}
\alias{checks_capture_params}
\title{Parameter documentation for capture data checks}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}
}
\description{
Parameter documentation for capture data checks
}
