% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checks_individual_params}
\alias{checks_individual_params}
\title{Parameter documentation for individual data checks}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\description{
Parameter documentation for individual data checks
}
