% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checks_return}
\alias{checks_return}
\title{Return documentation for data checks}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Return documentation for data checks
}
