% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{chick_mass_cutoffs}
\alias{chick_mass_cutoffs}
\title{List of cutoffs used to define errors and warnings for chick mass}
\format{
Data frame with 4 variables and 248 rows:
\describe{
\item{Species}{6 letter code to identify species}
\item{ChickAge}{Age at which cutoff is applied}
\item{Reference}{In what context should the cutoff be used (e.g. warning, error)}
\item{Value}{Value of the cutoff}
}
}
\description{
List of cutoffs used to define errors and warnings for chick mass
}
