% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_directory.R
\name{choose_directory}
\alias{choose_directory}
\title{Choose folder interactively.}
\usage{
choose_directory()
}
\value{
A pathname.
}
\description{
On Windows, will use 'choose.dir' on Mac will use 'tk_choose.dir'.
}
