% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WHZ.R
\name{clean_query_WHZ}
\alias{clean_query_WHZ}
\title{Clean syntax in SQL files}
\usage{
clean_query_WHZ(path)
}
\arguments{
\item{path}{Location of the SQL file.}
}
\value{
A list of characters with one element for each SQl statement.
}
\description{
WHZ primary data are stored in a series of SQL files, which contain redundant lines and syntax that is not correctly interpreted when reading into R. To execute the SQL statements, we first clean the SQL syntax.
}
