% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{column_names_v1.1}
\alias{column_names_v1.1}
\title{List of standard format column names (protocol version 1.1)}
\format{
A list with 4 elements. Each element is a character vector.
\describe{
  \item{Brood}{Column names in `Brood_data`.}
  \item{Capture}{Column names in `Capture_data`.}
  \item{Individual}{Column names in `Individual_data`.}
  \item{Location}{Column names in `Location_data`.}
  }
}
\description{
All columns names appearing in the 4 tables of the standard format, organized
in a list.
}
