% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{common_hybrids}
\alias{common_hybrids}
\title{Common hybrids}
\format{
Data frame with 2 variables:
\describe{
\item{Species1}{6 letter code for species of record 1}
\item{Species2}{6 letter code for species of record 2}
}
}
\usage{
common_hybrids
}
\description{
Data frame with common hybrid broods. Can be used in quality checks to compare species across parents, between parents and brood, and between brood and nestlings.
}
