% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_broodsize_chicknumber}
\alias{compare_broodsize_chicknumber}
\title{Compare brood size with number of chicks in Individual_data}
\usage{
compare_broodsize_chicknumber(
  Brood_data,
  Individual_data,
  approved_list,
  output,
  skip
)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Compare BroodSize in Brood_data with the number of chicks recorded in Individual_data. We expect these numbers to be equal. Records where BroodSize is larger than the number of chicks recorded will be flagged as a warning, because chicks might have died before ringing and measuring. In experimentally manipulated broods, BroodSize may be smaller than the number of chicks in Individual_data. If so, the record will be flagged as a warning. In non-manipulated broods, BroodSize should never be smaller than the number of chicks in Individual_data. If so, the records will be flagged as a potential error.
}
\details{
Check ID: B6.
}
