% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_clutch_brood}
\alias{compare_clutch_brood}
\title{Compare clutch and brood sizes}
\usage{
compare_clutch_brood(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
Check list, warning output, error output.
}
\description{
Compare clutch size and brood size per brood. In non-manipulated broods, clutch size should be larger or equal to brood size. If not, the record will result in an error. In broods with clutch manipulation, clutch size might be smaller than brood size. If so, the record will result in a warning.
}
