% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_species_brood_chicks}
\alias{compare_species_brood_chicks}
\title{Compare species of brood and species of chicks}
\usage{
compare_species_brood_chicks(Brood_data, Individual_data, approved_list)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the species of broods is the same as species of the chicks in that brood. Records that violate this assumption are flagged as 'warning' because brood fostering is known to exist.
}
\details{
Check ID: B12.
}
