% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_tarsus.R
\name{convert_tarsus}
\alias{convert_tarsus}
\title{Convert tarsus measurements to Svennson's Alternative}
\usage{
convert_tarsus(x, method)
}
\arguments{
\item{x}{A numeric vector. Tarsus length measurements (mm).}

\item{method}{Character string. The measurement method of the original data.
Can be: "Standard" (Svennson's Standard) or "Oxford" (Oxford maximum).}
}
\description{
Convert tarsus lenght measurements in Svennson's Standard and Oxford Maximum
to Svensson's Alternative.
}
\details{
We consider Svensson's alternative to be the most reliable measurement
method.
}
\examples{

convert_tarsus(x = c(15, 20, 22), method = "Standard")
}
