% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approved_list.R
\name{create_approved_list}
\alias{create_approved_list}
\title{Create approved_list}
\usage{
create_approved_list()
}
\value{
List of 4 dataframes:
\item{Brood_approved_list}{Approved records in Brood_data.}
\item{Capture_approved_list}{Approved records in Capture_data.}
\item{Individual_approved_list}{Approved records in Individual_data.}
\item{Location_approved_list}{Approved records in Location_data.}
}
\description{
A function to create a list of values that are verified as correct by data owners but previously marked as warning/error.
}
\details{
When a data owner is sent a quality check report, they can determine whether the flagged records are a) unverifiable, b) verifiable and inccorect, or c) verifiable and correct. If a), it cannot be known for certain whether the information should be kept or removed. They are continued to be flagged as improbable/impossible, and it is up to the user to decide how they treat them in analysis. If b), it will continue to be flagged until the data owner changes the value in the raw data. If c), the records will be disregarded for future quality checks.

To prevent verified records from flagged in future quality checks, there are two pieces of information we use: the unique identifier of the record (e.g., BroodID in Brood_data) and the unique identifier of the quality check that resulted in the original warning/error (e.g., B2).
}
