% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BAN.R
\name{create_brood_BAN}
\alias{create_brood_BAN}
\title{Create brood data table for Bandon Valley.}
\usage{
create_brood_BAN(data)
}
\arguments{
\item{data}{Data frame. Primary data from Bandon Valley.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Bandon Valley.
}
