% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BRG.R
\name{create_brood_BRG}
\alias{create_brood_BRG}
\title{Create brood data table in Bergen, Norway.}
\usage{
create_brood_BRG(nest_data, chick_data, adult_data)
}
\arguments{
\item{nest_data}{Data frame of nest data from Bergen, Norway.}

\item{chick_data}{Data frame of chick ringing records from Bergen, Norway.}

\item{adult_data}{Data frame of adult ringing records from Bergen, Norway.}
}
\value{
A data frame.
}
\description{
Create brood data table in Bergen, Norway.
}
