% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{create_brood_CHO}
\alias{create_brood_CHO}
\title{Create brood data table for Choupal, Portugal.}
\usage{
create_brood_CHO(data, optional_variables = NULL)
}
\arguments{
\item{data}{Data frame. Primary data from Choupal.}

\item{optional_variables}{A character vector of names of optional variables (generated by standard utility functions) to be included in the pipeline output.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Choupal,
Portugal.
}
