% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_brood_EDM}
\alias{create_brood_EDM}
\title{Create brood data table for EastDartmoor.}
\usage{
create_brood_EDM(Primary_data, ReRingTable)
}
\arguments{
\item{Primary_data}{Primary data from EastDartmoor.}

\item{ReRingTable}{Table containing ring numbers and unique identifiers for all individuals}
}
\value{
A data frame with Brood data
}
\description{
Create brood data table for EastDartmoor.
}
