% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_FOR.R
\name{create_brood_FOR}
\alias{create_brood_FOR}
\title{Create brood data table for Forstenrieder park, Germany.}
\usage{
create_brood_FOR(table_dir)
}
\arguments{
\item{table_dir}{Path to directory containing exported CSV tables.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Forstenrieder park, Germany.
}
