% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HAR.R
\name{create_brood_HAR}
\alias{create_brood_HAR}
\title{Create brood data table for Harjavalta.}
\usage{
create_brood_HAR(db, species_filter)
}
\arguments{
\item{db}{Location of primary data from Harjavalta.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Harjavalta, Finland.
}
