% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HOC.R
\name{create_brood_HOC}
\alias{create_brood_HOC}
\title{Create brood data table for Hochstadt.}
\usage{
create_brood_HOC(db)
}
\arguments{
\item{db}{Location of primary data from Hochstadt.}
}
\value{
A data frame with Brood data
}
\description{
Create brood data table for Hochstadt.
}
