% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KEV.R
\name{create_brood_KEV}
\alias{create_brood_KEV}
\title{Create brood data table for Kevo, Finland.}
\usage{
create_brood_KEV(db, species_filter)
}
\arguments{
\item{db}{Location of primary data from Kevo}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Kevo, Finland.
}
