% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KIL.R
\name{create_brood_KIL}
\alias{create_brood_KIL}
\title{Create brood data table for Kilingi Nomme, Estonia.}
\usage{
create_brood_KIL(brood_data_til92, kil_data_95)
}
\arguments{
\item{brood_data_til92}{Data frame. Primary brood data from Kilingi Nomme, Estonia from 1971 to 1992.}

\item{kil_data_95}{Data frame. Primary brood data from Kilingi Nomme, Estonia from 1995 onwards.}
}
\value{
A data frame.
}
\description{
Create brood data table in the standard format (v1.1.0) for data from Kilingi Nomme, Estonia.
}
