% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAL.R
\name{create_brood_MAL}
\alias{create_brood_MAL}
\title{Create brood data table for great tits and blue tits in Malmo, Sweden.}
\usage{
create_brood_MAL(nest_data, rr_data, protocol_version)
}
\arguments{
\item{nest_data}{Data frame of nest data from Malmo, Sweden.}

\item{rr_data}{Data frame of ringing records from Malmo, Sweden.}
}
\value{
A data frame.
}
\description{
Create brood data table for great tits and blue tits in Malmo, Sweden.
}
