% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAR.R
\name{create_brood_MAR}
\alias{create_brood_MAR}
\title{Create brood data table for Mariola, Spain.}
\usage{
create_brood_MAR(brood_data, rr_data)
}
\arguments{
\item{brood_data}{Brood data compiled from primary data from Mariola, Spain.}

\item{rr_data}{Ringing data compiled from primary data from Mariola, Spain.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Mariola, Spain.
}
