% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{create_brood_MAY}
\alias{create_brood_MAY}
\title{Create brood data table for Mayachino, Russia.}
\usage{
create_brood_MAY(gt_data, pf_data, species_filter)
}
\arguments{
\item{gt_data}{Data frame. Great tit data from Mayachino, Russia.}

\item{pf_data}{Data frame. Pied flycatcher data from Mayachino, Russia.}

\item{species_filter}{Species six-letter codes from the standard protocol.
Used to filter the data.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Mayachino, Russia.
}
