% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{create_brood_MON}
\alias{create_brood_MON}
\title{Create brood data table for Montpellier}
\usage{
create_brood_MON(db, species_filter, pop_filter)
}
\arguments{
\item{db}{Location of primary data from Montpellier.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}
}
\value{
A data frame with Brood data
}
\description{
Create brood data table for Montpellier
}
