% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MTA.R
\name{create_brood_MTA}
\alias{create_brood_MTA}
\title{Create brood data table for great tits in MTA-PE Evolutionary Ecology Group, Hungary.}
\usage{
create_brood_MTA(mta_data)
}
\arguments{
\item{mta_data}{Data frame. Primary data from MTA-PE Evolutionary Ecology Group, Hungary.}
}
\value{
A data frame.
}
\description{
Create a capture data table in standard format for great tits in MTA-PE Evolutionary Ecology Group, Hungary.
}
