% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PEW.R
\name{create_brood_PEW}
\alias{create_brood_PEW}
\title{Create brood data table for Peerdsbos West, Belgium.}
\usage{
create_brood_PEW(data)
}
\arguments{
\item{data}{Data frame. Primary data from Peerdsbos West, Belgium.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for blue tit data from Peerdsbos West, Belgium.
}
