% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_brood_PFN}
\alias{create_brood_PFN}
\title{Create brood data table for EastDartmoor.}
\usage{
create_brood_PFN(db)
}
\arguments{
\item{db}{Location of primary data from EastDartmoor.}
}
\value{
A data frame with Brood data
}
\description{
Create brood data table for EastDartmoor.
}
