% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{create_brood_UAN}
\alias{create_brood_UAN}
\title{Create brood data table for data from University of Antwerp, Belgium.}
\usage{
create_brood_UAN(data, CAPTURE_info, species_filter)
}
\arguments{
\item{data}{Data frame. Primary data from University of Antwerp.}

\item{CAPTURE_info}{Capture data table from the raw data}

\item{species_filter}{6 letter species codes for filtering data.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from University of
Antwerp, Belgium.
}
