% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VEL.R
\name{create_brood_VEL}
\alias{create_brood_VEL}
\title{Create brood data table for Velky Kosir}
\usage{
create_brood_VEL(FICALB_data, TIT_data)
}
\arguments{
\item{FICALB_data}{Data frame. Flycatcher data from Velky Kosir.}

\item{TIT_data}{Data frame. Tit data from Velky Kosir.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for flycatcher and tit data from Velky
Kosir.
}
