% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WRS.R
\name{create_brood_WRS}
\alias{create_brood_WRS}
\title{Create brood data table for great tits and blue tits in Warsaw, Poland.}
\usage{
create_brood_WRS(nest_data, chick_data, adult_data)
}
\arguments{
\item{nest_data}{Data frame of nest data from Warsaw, Poland.}

\item{chick_data}{Data frame of chick ringing records from Warsaw, Poland.}

\item{adult_data}{Data frame of adult ringing records from Warsaw, Poland.}
}
\value{
A data frame.
}
\description{
Create brood data table for great tits and blue tits in Warsaw, Poland.
}
