% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WYT.R
\name{create_brood_WYT}
\alias{create_brood_WYT}
\title{Create brood data table for Wytham Woods}
\usage{
create_brood_WYT(db, species_filter)
}
\arguments{
\item{db}{Location of primary data from Whytham Woods.}

\item{species_filter}{Species of interest.
The 6 letter codes of all the species of
 interest as listed in the
 \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{standard protocol}.}
}
\value{
A data frame with Brood data
}
\description{
Create brood data table for Wytham Woods
}
