% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BRG.R
\name{create_capture_BRG}
\alias{create_capture_BRG}
\title{Create capture data table for Bergen, Norway.}
\usage{
create_capture_BRG(chick_data, adult_data, Brood_data_temp)
}
\arguments{
\item{chick_data, }{Data frame of chick ringing records from Bergen, Norway.}

\item{adult_data, }{Data frame of adult ringing records from Bergen, Norway.}

\item{Brood_data_temp, }{Data frame of brood data created from primary data from Bergen, Norway.}
}
\value{
A data frame.
}
\description{
Create capture data table for Bergen, Norway.
}
