% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{create_capture_CHO}
\alias{create_capture_CHO}
\title{Create capture data table for Choupal, Portugal.}
\usage{
create_capture_CHO(data)
}
\arguments{
\item{data}{Data frame. Primary data from Choupal.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from Choupal,
Portugal.
}
