% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_capture_EDM}
\alias{create_capture_EDM}
\title{Create capture data table for EastDartmoor from ringing data.}
\usage{
create_capture_EDM(CMR_data, Primary_data, ReRingTable)
}
\arguments{
\item{CMR_data}{Complete raw PiedFlyNet ringing data.}

\item{Primary_data}{Primary brood data from EastDartmoor.}

\item{ReRingTable}{Table containing ring numbers and unique identifiers for re-ringed individuals}
}
\value{
A data frame with Capture data
}
\description{
Create capture data table for EastDartmoor from ringing data.
}
