% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_FOR.R
\name{create_capture_FOR}
\alias{create_capture_FOR}
\title{Create capture data table for Forstenrieder park, Germany}
\usage{
create_capture_FOR(Brood_data, table_dir)
}
\arguments{
\item{Brood_data}{Data frame. Output from \code{\link{create_brood_FOR}}.}

\item{table_dir}{Path to directory containing exported CSV tables.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from Forstenrieder park, Germany.
}
