% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GRO.R
\name{create_capture_GRO}
\alias{create_capture_GRO}
\title{Create capture data table for great tits and blue tits in Grobla, Poland.}
\usage{
create_capture_GRO(gro_data, protocol_version)
}
\arguments{
\item{gro_data}{Data frame of modified primary data from Grobla, Poland.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create capture data table for great tits and blue tits in Grobla, Poland.
}
