% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HAR.R
\name{create_capture_HAR}
\alias{create_capture_HAR}
\title{Create capture table for Harjavalta, Finland.}
\usage{
create_capture_HAR(db, Brood_data, species_filter, return_errors)
}
\arguments{
\item{db}{Location of primary data from Harjavalta.}

\item{Brood_data}{Output of \code{\link{create_brood_HAR}}.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}
}
\value{
A data frame.
}
\description{
Create full capture data table in standard format for data from Harjavalta, Finland.
}
