% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HOC.R
\name{create_capture_HOC}
\alias{create_capture_HOC}
\title{Create capture data table for Hochstadt.}
\usage{
create_capture_HOC(db)
}
\arguments{
\item{db}{Location of primary data from Hochstadt.}
}
\value{
A data frame with Capture data
}
\description{
Create capture data table for Hochstadt.
}
