% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_capture_IPMR_PFN}
\alias{create_capture_IPMR_PFN}
\title{Extract capture data table for PiedFlyNet populations from
IPMR primary data.}
\usage{
create_capture_IPMR_PFN(IPMR_data, Nest_data, ReRingTable)
}
\arguments{
\item{IPMR_data}{Collated IPMR primary data from PiedFlyNet populations}

\item{Nest_data}{Collated Nest primary data from PiedFlyNet populations}

\item{ReRingTable}{Table containing ring numbers and unique identifiers for all individuals}
}
\description{
Extract capture data table for PiedFlyNet populations from
IPMR primary data.
}
