% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KEV.R
\name{create_capture_KEV}
\alias{create_capture_KEV}
\title{Create capture table for Kevo, Finland.}
\usage{
create_capture_KEV(db, Brood_data, species_filter, return_errors)
}
\arguments{
\item{db}{Location of primary data from Kevo.}

\item{Brood_data}{Output of \code{\link{create_brood_KEV}}.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}
}
\value{
A data frame.
}
\description{
Create full capture data table in standard format for data from Kevo, Finland.
}
