% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KIL.R
\name{create_capture_KIL}
\alias{create_capture_KIL}
\title{Create capture data table for Kilingi Nomme, Estonia.}
\usage{
create_capture_KIL(
  brood_data_til92,
  adults_data_til92,
  chicks_data_til92,
  adults_data_95,
  chicks_data_95
)
}
\arguments{
\item{brood_data_til92}{Data frame. Primary brood data from Kilingi Nomme, Estonia from 1971 to 1992.}

\item{adults_data_til92}{Data frame. Primary adult capture data from Kilingi Nomme, Estonia from 1971 to 1992.}

\item{chicks_data_til92}{Data frame. Primary chick capture data from Kilingi Nomme, Estonia from 1971 to 1992.}

\item{adults_data_95}{Data frame. Primary adult capture data from Kilingi Nomme, Estonia from 1995 onwards.}

\item{chicks_data_95}{Data frame. Primary chick capture data from Kilingi Nomme, Estonia from 1995 onwards.}
}
\value{
A data frame.
}
\description{
Create capture data table in the standard format (v1.1.0) for data from Kilingi Nomme, Estonia.
}
