% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAL.R
\name{create_capture_MAL}
\alias{create_capture_MAL}
\title{Create capture data table for great tits and blue tits in Malmo, Sweden.}
\usage{
create_capture_MAL(rr_data, protocol_version)
}
\arguments{
\item{rr_data}{Data frame of ringing records from Malmo, Sweden.}
}
\value{
A data frame.
}
\description{
Create capture data table for great tits and blue tits in Malmo, Sweden.
}
