% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAR.R
\name{create_capture_MAR}
\alias{create_capture_MAR}
\title{Create capture data table for Mariola, Spain}
\usage{
create_capture_MAR(Brood_data, connection)
}
\arguments{
\item{Brood_data}{Data frame. Output from \code{\link{create_brood_MAR}}.}

\item{connection}{Connection the SQL database.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from Mariola, Spain.
}
