% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{create_capture_MON}
\alias{create_capture_MON}
\title{Create capture data table for Montpellier}
\usage{
create_capture_MON(db, species_filter, pop_filter)
}
\arguments{
\item{db}{Location of primary data from Montpellier.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}

\item{pop_filter}{Population three letter codes from the standard protocol.
Used to filter the data.}
}
\value{
A data frame with capture data
}
\description{
Create capture data table for Montpellier
}
