% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MTA.R
\name{create_capture_MTA}
\alias{create_capture_MTA}
\title{Create capture data table for great tits in MTA-PE Evolutionary Ecology Group, Hungary.}
\usage{
create_capture_MTA(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood_data from MTA-PE Evolutionary Ecology Group, Hungary.}
}
\value{
A data frame.
}
\description{
Create a capture data table in standard format for great tits in MTA-PE Evolutionary Ecology Group, Hungary.
}
