% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_capture_Nest_PFN}
\alias{create_capture_Nest_PFN}
\title{Extract capture data table for PiedFlyNet populations from
Nest primary data.}
\usage{
create_capture_Nest_PFN(Nest_data, ReRingTable, badIDs)
}
\arguments{
\item{Nest_data}{Collated Nest primary data from PiedFlyNet populations}

\item{ReRingTable}{Table containing ring numbers and unique identifiers for all individuals}

\item{badIDs}{Vector containing non-conslusive IDs (extracted from Nest data)}
}
\description{
Extract capture data table for PiedFlyNet populations from
Nest primary data.
}
