% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PET.R
\name{create_capture_PET}
\alias{create_capture_PET}
\title{Create capture data table for the populations from Institute of Biology, Karelian Research Centre, Russian Academy of Sciences, Petrozavodsk, Russia.}
\usage{
create_capture_PET(parmaj_data, fichyp_data, jyntor_data)
}
\arguments{
\item{parmaj_data}{Data frame. Primary data from PARMAJ records from Institute of Biology, Russia.}

\item{fichyp_data}{Data frame. Primary data from FICHYP records from Institute of Biology, Russia.}

\item{jyntor_data}{Data frame. Primary data from JYNTOR records from Institute of Biology, Russia.}
}
\value{
A data frame.
}
\description{
Create a capture data table in standard format for great tits, pied flycatchers and Wryneck in Institute of Biology, Russia.
}
