% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PEW.R
\name{create_capture_PEW}
\alias{create_capture_PEW}
\title{Create capture data table for blue tits in Peerdsbos West, Belgium.}
\usage{
create_capture_PEW(pew_data, Brood_data)
}
\arguments{
\item{pew_data}{Data frame. Data frame pew_data with primary data from Peerdsbos West, Belgium.}

\item{Brood_data}{Data frame. Brood_data from Peerdsbos West, Belgium.}
}
\value{
A data frame.
}
\description{
Create a capture data table in standard format for blue tits in Peerdsbos West, Belgium.
}
