% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_capture_PFN}
\alias{create_capture_PFN}
\title{Create capture data table for PiedFlyNet populations from
consolidated Nest and IPMR primary data sources.}
\usage{
create_capture_PFN(Nest_data, IPMR_data, ReRingTable, badIDs)
}
\arguments{
\item{Nest_data}{Collated Nest primary data from PiedFlyNet populations}

\item{IPMR_data}{Collated IPMR primary data from PiedFlyNet populations}

\item{ReRingTable}{Table containing ring numbers and unique identifiers for all individuals}

\item{badIDs}{Vector containing non-conslusive IDs (extracted from Nest data)}
}
\value{
A data frame with Capture data
}
\description{
Consolidation of information is necessary when Nest and IPMR primary data
contain conflicting information about what is likely the same capture
Captures are matched via `PopID`, `IndvID` and `CaptureDate` when possible.
When `CaptureDate` is missing or incomplete, matching is attempted using
`BreedingSeason` and `LocationID`. If matching is impossible based on available
information, data is not consolidated.
When one of two matched capture records is missing information, equivalent
information from the other record is used. When two matched capture records
contain conflicting information, the entry is either noted as
conflicting (`Species`, `Sex_observed`) or information from the more reliable
data source is used. With the exception of `CaptureAlive` and `ReleaseAlive`,
Nest data is considered as the more reliable data source.
}
