% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PIL.R
\name{create_capture_PIL}
\alias{create_capture_PIL}
\title{Create capture data table for Pilis-Visegrád Mountains, Hungary.}
\usage{
create_capture_PIL(PIL_data, species_filter)
}
\arguments{
\item{PIL_data}{Data frame with primary data from Pilis-Visegrád Mountains}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from Pilis-Visegrád Mountains, Hungary.
}
