% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VEL.R
\name{create_capture_VEL_FICALB}
\alias{create_capture_VEL_FICALB}
\title{Create capture data table for flycatchers in Velky Kosir.}
\usage{
create_capture_VEL_FICALB(FICALB_data)
}
\arguments{
\item{FICALB_data}{Data frame. Flycatcher data from Velky Kosir.}
}
\value{
A data frame.
}
\description{
Create a capture data table in standard format for only flycatchers in Velky Kosir.
Tit data is created with separate function \code{\link{create_capture_VEL_TIT}}.
}
