% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VEL.R
\name{create_capture_VEL_TIT}
\alias{create_capture_VEL_TIT}
\title{Create capture data table for tits in Velky Kosir.}
\usage{
create_capture_VEL_TIT(TIT_data)
}
\arguments{
\item{TIT_data}{Data frame. Tit data from Velky Kosir.}
}
\value{
A data frame.
}
\description{
Create a capture data table in standard format for great and blue tits in Velky Kosir.
Tit data is created with separate function \code{\link{create_capture_VEL_FICALB}}.
}
