% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WHZ.R
\name{create_capture_WHZ}
\alias{create_capture_WHZ}
\title{Create capture data table for Westerholz, Germany.}
\usage{
create_capture_WHZ(connection, optional_variables)
}
\arguments{
\item{connection}{Connection the SQL database.}

\item{optional_variables}{A character vector of names of optional variables (generated by standard utility functions) to be included in the pipeline output.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from Westerholz, Germany.
}
