% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WRS.R
\name{create_capture_WRS}
\alias{create_capture_WRS}
\title{Create capture data table for Warsaw, Poland.}
\usage{
create_capture_WRS(chick_data, adult_data)
}
\arguments{
\item{chick_data, }{Data frame of chick ringing records from Warsaw, Poland.}

\item{adult_data, }{Data frame of adult ringing records from Warsaw, Poland.}
}
\value{
A data frame.
}
\description{
Create capture data table for Warsaw, Poland.
}
