% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WYT.R
\name{create_capture_WYT}
\alias{create_capture_WYT}
\title{Create capture data table for Wytham Woods}
\usage{
create_capture_WYT(db, Brood_data, species_filter)
}
\arguments{
\item{db}{Location of primary data from Whytham Woods.}

\item{Brood_data}{Brood data generated by \code{\link{create_brood_WYT}}.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}.}
}
\value{
A data frame with Capture data
}
\description{
Create capture data table for Wytham Woods
}
