% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{create_experiment_MAY}
\alias{create_experiment_MAY}
\title{Create experiment data table for Mayachino, Russia.}
\usage{
create_experiment_MAY(brood_data)
}
\arguments{
\item{brood_data}{Data frame. Output from \code{\link{create_brood_MAY}}.}
}
\value{
A data frame.
}
\description{
Create experiment data table in standard format for data from Mayachino, Russia.
}
