% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WHZ.R
\name{create_experiment_WHZ}
\alias{create_experiment_WHZ}
\title{Create experiment data table for Westerholz, Germany.}
\usage{
create_experiment_WHZ(brood_data, capture_data)
}
\arguments{
\item{brood_data}{Data frame. Output from \code{\link{create_brood_WHZ}}.}

\item{capture_data}{Data frame. Output from \code{\link{create_capture_WHZ}}.}
}
\value{
A data frame.
}
\description{
Create experiment data table in standard format for data from Westerholz, Germany.
}
