% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{create_individual_AMM}
\alias{create_individual_AMM}
\title{Create individual data table for Ammersee, Germany}
\usage{
create_individual_AMM(Capture_data, Brood_data, dir)
}
\arguments{
\item{Capture_data}{Data frame. Output from \code{\link{create_capture_AMM}}.}

\item{Brood_data}{Data frame. Output from \code{\link{create_brood_AMM}}.}

\item{dir}{Path to directory containing the relevant table exports from the AMM Access database.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Ammersee, Germany.
}
