% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BRG.R
\name{create_individual_BRG}
\alias{create_individual_BRG}
\title{Create individual table for Bergen, Norway.}
\usage{
create_individual_BRG(Capture_data_temp)
}
\arguments{
\item{Capture_data_temp}{Capture data output from Bergen, Norway}
}
\value{
A data frame.
}
\description{
Create individual table for Bergen, Norway.
}
