% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_DLO.R
\name{create_individual_DLO}
\alias{create_individual_DLO}
\title{Create individual data table for Dlouhá Loučka, Czechia.}
\usage{
create_individual_DLO(capture_data, species_filter, optional_variables = NULL)
}
\arguments{
\item{capture_data}{Data frame. Output from \code{\link{create_capture_DLO}}.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}.}

\item{optional_variables}{A character vector of names of optional variables (generated by standard utility functions) to be included in the pipeline output.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Dlouhá Loučka, Czechia.
}
