% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GLA.R
\name{create_individual_GLA}
\alias{create_individual_GLA}
\title{Create individual table for great tits and blue tits in Glasgow, Scotland.}
\usage{
create_individual_GLA(Capture_data, Brood_data)
}
\arguments{
\item{Capture_data}{Capture data output from Glasgow, Scotland}

\item{Brood_data}{Brood data output from Glasgow, Scotland}
}
\value{
A data frame.
}
\description{
Create full individual data table in standard format for great tits and blue tits in Glasgow, Scotland.
}
