% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HOC.R
\name{create_individual_HOC}
\alias{create_individual_HOC}
\title{Create individual data table for Hochstadt.}
\usage{
create_individual_HOC(db)
}
\arguments{
\item{db}{Location of individual data from Hochstadt.}
}
\value{
A data frame with Individual data
}
\description{
Create individual data table for Hochstadt.
}
