% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KEV.R
\name{create_individual_KEV}
\alias{create_individual_KEV}
\title{Create individual table for Kevo, Finland.}
\usage{
create_individual_KEV(Capture_data, protocol_version)
}
\arguments{
\item{Capture_data}{Output of \code{\link{create_capture_HAR}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create full individual data table in standard format for data from Kevo, Finland.
}
