% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KIL.R
\name{create_individual_KIL}
\alias{create_individual_KIL}
\title{Create individual data table for Kilingi Nomme, Estonia.}
\usage{
create_individual_KIL(
  adults_data_til92,
  chicks_data_til92,
  adults_data_95,
  chicks_data_95,
  protocol_version
)
}
\arguments{
\item{adults_data_til92}{Data frame. Primary adult capture data from Kilingi Nomme, Estonia from 1971 to 1992.}

\item{chicks_data_til92}{Data frame. Primary chick capture data from Kilingi Nomme, Estonia from 1971 to 1992.}

\item{adults_data_95}{Data frame. Primary adult capture data from Kilingi Nomme, Estonia from 1995 onwards.}

\item{chicks_data_95}{Data frame. Primary chick capture data from Kilingi Nomme, Estonia from 1995 onwards.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create individual data table in the standard format (v1.1.0) for data from Kilingi Nomme, Estonia.
}
