% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAL.R
\name{create_individual_MAL}
\alias{create_individual_MAL}
\title{Create individual table for great tits and blue tits in Malmo, Sweden.}
\usage{
create_individual_MAL(Capture_data, Brood_data, protocol_version)
}
\arguments{
\item{Capture_data}{Capture data output from Malmo, Sweden.}

\item{Brood_data}{Brood data output from Malmo, Sweden.}
}
\value{
A data frame.
}
\description{
Create individual table for great tits and blue tits in Malmo, Sweden.
}
