% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAR.R
\name{create_individual_MAR}
\alias{create_individual_MAR}
\title{Create individual data table for Mariola, Spain}
\usage{
create_individual_MAR(Capture_data, Brood_data, connection)
}
\arguments{
\item{Capture_data}{Data frame. Output from \code{\link{create_capture_MAR}}.}

\item{Brood_data}{Data frame. Output from \code{\link{create_brood_MAR}}.}

\item{connection}{Connection the SQL database.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Mariola, Spain.
}
