% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{create_individual_MAY}
\alias{create_individual_MAY}
\title{Create individual data table for Mayachino, Russia.}
\usage{
create_individual_MAY(capture_data, species_filter)
}
\arguments{
\item{capture_data}{Data frame. Output from \code{\link{create_capture_MAY}}.}

\item{species_filter}{Species six-letter codes from the standard protocol.
Used to filter the data.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Mayachino, Russia.
}
