% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{create_individual_MON}
\alias{create_individual_MON}
\title{Create individual data table for Montpellier}
\usage{
create_individual_MON(Capture_data, Brood_data, protocol_version, verbose)
}
\arguments{
\item{Capture_data}{Capture data generated by \code{\link{create_capture_MON}}.}

\item{Brood_data}{Capture data generated by \code{\link{create_brood_MON}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}

\item{verbose}{When chicks with duplicate nests are found, should a message be printed?}
}
\value{
A data frame with Individual data
}
\description{
Create individual data table for Montpellier
}
