% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MTA.R
\name{create_individual_MTA}
\alias{create_individual_MTA}
\title{Create individual table for MTA-PE Evolutionary Ecology Group, Hungary.}
\usage{
create_individual_MTA(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame, output of create_capture_MTA function.}
}
\value{
A data frame.
}
\description{
Create full individual data table in standard format for data from MTA-PE Evolutionary Ecology Group, Hungary.
}
