% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_NIOO.R
\name{create_individual_NIOO}
\alias{create_individual_NIOO}
\title{Create individual data table for NIOO pipeline.}
\usage{
create_individual_NIOO(
  dir,
  Capture_data,
  location_data,
  species_filter,
  pop_filter,
  protocol_version
)
}
\arguments{
\item{dir}{Path to directory containing the relevant table exports from the NIOO Access database.}

\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{location_data}{Data frame with location codes and corresponding PopID.}

\item{species_filter}{Species six letter codes from the standard protocol.
Used to filter the data.}

\item{pop_filter}{Population three letter codes from the standard protocol.
Used to filter the data.}

\item{protocol_version}{Version of the standard protocol.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from NIOO.
}
