% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PET.R
\name{create_individual_PET}
\alias{create_individual_PET}
\title{Create individual table for Institute of Biology, Russia.}
\usage{
create_individual_PET(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame, output of create_capture_PET function.}
}
\value{
A data frame.
}
\description{
Create full individual data table in standard format for data from Institute of Biology, Russia.
}
