% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PEW.R
\name{create_individual_PEW}
\alias{create_individual_PEW}
\title{Create individual table for Peerdsbos West, Belgium.}
\usage{
create_individual_PEW(data)
}
\arguments{
\item{data}{Capture_data, output of create_capture_PEW function.}
}
\value{
A data frame.
}
\description{
Create full individual data table in standard format for data from Peerdsbos West, Belgium.
}
