% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_individual_PFN}
\alias{create_individual_PFN}
\title{Create individual data table for EastDartmoor.}
\usage{
create_individual_PFN(Capture_data)
}
\arguments{
\item{Capture_data}{Capture data table generated by create_capture_PFN}
}
\value{
A data frame with Individual data
}
\description{
Create individual data table for EastDartmoor.
}
