% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PIL.R
\name{create_individual_PIL}
\alias{create_individual_PIL}
\title{Create individual table for Pilis-Visegrád Mountains, Hungary.}
\usage{
create_individual_PIL(Capture_data, protocol_version)
}
\arguments{
\item{Capture_data}{Output of \code{\link{create_capture_PIL}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create full individual data table in standard format for data from Pilis-Visegrád Mountains, Hungary.
}
