% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VOM.R
\name{create_individual_VOM}
\alias{create_individual_VOM}
\title{Create individual table for great tits and blue tits in Vomb Fure, Sweden.}
\usage{
create_individual_VOM(Capture_data_temp)
}
\arguments{
\item{Capture_data_temp}{Data frame of Capture data from Vomb Fure, Sweden.}
}
\value{
A data frame.
}
\description{
Create individual table for great tits and blue tits in Vomb Fure, Sweden.
}
