% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WHZ.R
\name{create_individual_WHZ}
\alias{create_individual_WHZ}
\title{Create individual data table for Westerholz, Germany.}
\usage{
create_individual_WHZ(capture_data, brood_data, optional_variables)
}
\arguments{
\item{capture_data}{Data frame. Output from \code{\link{create_capture_WHZ}}.}

\item{brood_data}{Data frame. Output from \code{\link{create_brood_WHZ}}.}

\item{optional_variables}{A character vector of names of optional variables (generated by standard utility functions) to be included in the pipeline output.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Westerholz, Germany.
}
