% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WRS.R
\name{create_individual_WRS}
\alias{create_individual_WRS}
\title{Create individual table for great tits and blue tits in Warsaw, Poland.}
\usage{
create_individual_WRS(Capture_data_temp, Brood_data_temp)
}
\arguments{
\item{Capture_data_temp}{Capture data output from Warsaw, Poland}

\item{Brood_data_temp}{Brood data output from Warsaw, Poland}
}
\value{
A data frame.
}
\description{
Create individual table for great tits and blue tits in Warsaw, Poland.
}
