% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WYT.R
\name{create_individual_WYT}
\alias{create_individual_WYT}
\title{Create individual data table for Wytham Woods}
\usage{
create_individual_WYT(Capture_data)
}
\arguments{
\item{Capture_data}{Capture data generated by \code{\link{create_capture_WYT}}.}
}
\value{
A data frame with Individual data
}
\description{
Create individual data table for Wytham Woods
}
