% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{create_location_AMM}
\alias{create_location_AMM}
\title{Create location data table for Ammersee, Germany.}
\usage{
create_location_AMM(Capture_data, connection)
}
\arguments{
\item{data}{Data frame. Primary data from Ammersee.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Ammersee, Germany.
}
