% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BAN.R
\name{create_location_BAN}
\alias{create_location_BAN}
\title{Create location data table for Bandon Valley, Ireland.}
\usage{
create_location_BAN(data)
}
\arguments{
\item{data}{Data frame. Primary data from Bandon Valley.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Bandon Valley,
Ireland.
}
