% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BRG.R
\name{create_location_BRG}
\alias{create_location_BRG}
\title{Create location data table for Bergen, Norway.}
\usage{
create_location_BRG(nest_data)
}
\arguments{
\item{nest_data}{Data frame of nest data from Bergen, Norway.}
}
\value{
A data frame.
}
\description{
Create location data table for Bergen, Norway.
}
