% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_DLO.R
\name{create_location_DLO}
\alias{create_location_DLO}
\title{Create location data table for Dlouhá Loučka, Czechia.}
\usage{
create_location_DLO(data)
}
\arguments{
\item{data}{Data frame. Primary data from Dlouhá Loučka, Czechia.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Dlouhá Loučka, Czechia.
}
