% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_location_EDM}
\alias{create_location_EDM}
\title{Create location data table for EastDartmoor.}
\usage{
create_location_EDM(Brood_data, Capture_data)
}
\arguments{
\item{Brood_data}{Brood data table generated by create_brood_EDM}

\item{Capture_data}{Capture data table generated by create_capture_EDM}
}
\value{
A data frame with Location data
}
\description{
Create location data table for EastDartmoor.
}
