% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_FOR.R
\name{create_location_FOR}
\alias{create_location_FOR}
\title{Create location data table for Forstenrieder Park, Germany.}
\usage{
create_location_FOR(Capture_data, table_dir)
}
\arguments{
\item{Capture_data}{Data frame. Output from \code{\link{create_capture_FOR}}.}

\item{table_dir}{Path to directory containing exported CSV tables.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Forstenrieder Park, Germany.
}
