% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GLA.R
\name{create_location_GLA}
\alias{create_location_GLA}
\title{Create location data table for great tits and blue tits in Glasgow, Scotland.}
\usage{
create_location_GLA(nest_data, rr_data, protocol_version)
}
\arguments{
\item{nest_data}{Data frame of nest data from Glasgow, Scotland.}

\item{rr_data}{Data frame of ringing records from Glasgow, Scotland.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create location data table for great tits and blue tits in Glasgow, Scotland.
}
