% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HAR.R
\name{create_location_HAR}
\alias{create_location_HAR}
\title{Create location table for Harjavalta, Finland.}
\usage{
create_location_HAR(db, protocol_version)
}
\arguments{
\item{db}{Location of primary data from Harjavalta.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create full location data table in standard format for data from Harjavalta, Finland.
}
