% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HOC.R
\name{create_location_HOC}
\alias{create_location_HOC}
\title{Create location data table for Hochstadt.}
\usage{
create_location_HOC(db)
}
\arguments{
\item{db}{Location of location data from Hochstadt.}
}
\value{
A data frame with Location data
}
\description{
Create location data table for Hochstadt.
}
