% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KIL.R
\name{create_location_KIL}
\alias{create_location_KIL}
\title{Create location data table for Kilingi Nomme, Estonia.}
\usage{
create_location_KIL(kil_data_95, Brood_data)
}
\arguments{
\item{kil_data_95}{Data frame. Primary brood data from Kilingi Nomme, Estonia from 1995 onwards.}

\item{Brood_data}{Data frame. Brood data generated by \code{\link{create_brood_KIL}}.}
}
\value{
A data frame.
}
\description{
Create location data table in the standard format (v1.1.0) for data from Kilingi Nomme, Estonia.
}
