% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAL.R
\name{create_location_MAL}
\alias{create_location_MAL}
\title{Create location data table for great tits and blue tits in Malmo, Sweden.}
\usage{
create_location_MAL(loc_data, protocol_version)
}
\arguments{
\item{loc_data}{Data frame of location data from Malmo, Sweden.}
}
\value{
A data frame.
}
\description{
Create location data table for great tits and blue tits in Malmo, Sweden.
}
