% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{create_location_MAY}
\alias{create_location_MAY}
\title{Create location data table for Mayachino, Russia.}
\usage{
create_location_MAY(gt_data, pf_data)
}
\arguments{
\item{gt_data}{Data frame. Great tit data from Mayachino, Russia.}

\item{pf_data}{Data frame. Pied flycatcher data from Mayachino, Russia.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Mayachino, Russia.
}
