% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{create_location_MON}
\alias{create_location_MON}
\title{Create location data table for Montpellier}
\usage{
create_location_MON(db, capture_data, brood_data)
}
\arguments{
\item{db}{Location of primary data from Montpellier.}

\item{capture_data}{Capture data generated by \code{\link{create_capture_MON}}.}

\item{brood_data}{Capture data generated by \code{\link{create_brood_MON}}.}
}
\value{
A data frame with Individual data
}
\description{
Create location data table for Montpellier
}
