% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MTA.R
\name{create_location_MTA}
\alias{create_location_MTA}
\title{Create location data table for MTA-PE, Hungary.}
\usage{
create_location_MTA(mta_data)
}
\arguments{
\item{mta_data}{Data frame mta_data with primary data from MTA-PE, Hungary.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from MTA-PE, Hungary.
}
