% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PET.R
\name{create_location_PET}
\alias{create_location_PET}
\title{Create location data table for for the populations from Institute of Biology, Karelian Research Centre, Russian Academy of Sciences, Petrozavodsk, Russia.}
\usage{
create_location_PET(parmaj_data, fichyp_data, jyntor_data)
}
\arguments{
\item{parmaj_data}{Data frame. Primary data from PARMAJ records from Institute of Biology, Russia.}

\item{fichyp_data}{Data frame. Primary data from FICHYP records from Institute of Biology, Russia.}

\item{jyntor_data}{Data frame. Primary data from JYNTOR records from Institute of Biology, Russia.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Institute of Biology, Russia.
}
