% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_location_PFN}
\alias{create_location_PFN}
\title{Create location data table for PiedFlyNet populations.}
\usage{
create_location_PFN(Brood_data, Capture_data, Location_details, pop_filter)
}
\arguments{
\item{Brood_data}{Brood data table generated by `create_brood_PFN`}

\item{Capture_data}{Capture data table generated by `create_capture_PFN`}

\item{Location_details}{Table containing detailed information on nest boxes.
Currently only contains information for EDM. Information for other populations
will be added later.}

\item{pop_filter}{Population three-letter codes from the standard protocol. Used to filter the data.}
}
\value{
A data frame with Location data
}
\description{
Create location data table for PiedFlyNet populations.
}
