% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SSQ.R
\name{create_location_SSQ}
\alias{create_location_SSQ}
\title{Create location data table for Santo Stefano Quisquina, Italy.}
\usage{
create_location_SSQ(data)
}
\arguments{
\item{data}{Data frame. Primary data from Santo Stefano Quisquina.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Santo Stefano
Quisquina, Italy
}
