% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{create_location_UAN}
\alias{create_location_UAN}
\title{Create location data table for data from University of Antwerp, Belgium.}
\usage{
create_location_UAN(BOX_info, protocol_version)
}
\arguments{
\item{BOX_info}{Data frame. Primary location data from University of Antwerp.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from University of
Antwerp, Belgium.
}
