% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VEL.R
\name{create_location_VEL}
\alias{create_location_VEL}
\title{Create location data table for Velky Kosir.}
\usage{
create_location_VEL(db, Brood_data, TIT_data, protocol_version)
}
\arguments{
\item{db}{Location of database file.}

\item{Brood_data}{Data frame. Output of \code{\link{create_brood_VEL}}}

\item{TIT_data}{Data frame. Data frame. Tit data from Velky Kosir. This is
needed to include habitat type information.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for all nest boxes.
}
