% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VOM.R
\name{create_location_VOM}
\alias{create_location_VOM}
\title{Create location data table for great tits and blue tits in Vomb Fure, Sweden.}
\usage{
create_location_VOM(nest_data)
}
\arguments{
\item{nest_data}{Data frame of nest data from Vomb Fure, Sweden.}
}
\value{
A data frame.
}
\description{
Create location data table for great tits and blue tits in Vomb Fure, Sweden.
}
