% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WHZ.R
\name{create_location_WHZ}
\alias{create_location_WHZ}
\title{Create location data table for Westerholz, Germany.}
\usage{
create_location_WHZ(connection)
}
\arguments{
\item{connection}{Connection the SQL database.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Westerholz, Germany.
}
