% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WRS.R
\name{create_location_WRS}
\alias{create_location_WRS}
\title{Create location data table for Warsaw, Poland.}
\usage{
create_location_WRS(nest_data)
}
\arguments{
\item{nest_data}{Data frame of nest data from Warsaw, Poland.}
}
\value{
A data frame.
}
\description{
Create location data table for Warsaw, Poland.
}
