% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WYT.R
\name{create_location_WYT}
\alias{create_location_WYT}
\title{Create location data table for Wytham Woods}
\usage{
create_location_WYT(Brood_data, Capture_data, protocol_version)
}
\arguments{
\item{Brood_data}{Brood data generated by \code{\link{create_brood_WYT}}.}

\item{Capture_data}{Capture data generated by \code{\link{create_capture_WYT}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame with Location data
}
\description{
Create location data table for Wytham Woods
}
