% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{create_measurement_CHO}
\alias{create_measurement_CHO}
\title{Create measurement data table for Choupal, Portugal.}
\usage{
create_measurement_CHO(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Output from \code{\link{create_capture_CHO}}.}
}
\value{
A data frame.
}
\description{
Create measurement data table in standard format for data from Choupal, Portugal.
}
