% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_DLO.R
\name{create_measurement_DLO}
\alias{create_measurement_DLO}
\title{Create measurement data table for Dlouhá Loučka, Czechia.}
\usage{
create_measurement_DLO(capture_data)
}
\arguments{
\item{capture_data}{Data frame. Output from \code{\link{create_capture_DLO}}.}
}
\value{
A data frame.
}
\description{
Create measurement data table in standard format for data from Dlouhá Loučka, Czechia.
}
