% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{create_measurement_MAY}
\alias{create_measurement_MAY}
\title{Create measurement data table for Mayachino, Russia.}
\usage{
create_measurement_MAY(capture_data)
}
\arguments{
\item{capture_data}{Data frame. Output from \code{\link{create_capture_MAY}}.}
}
\value{
A data frame.
}
\description{
Create measurement data table in standard format for data from Mayachino, Russia.
}
