% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WHZ.R
\name{create_measurement_WHZ}
\alias{create_measurement_WHZ}
\title{Create measurement data table for Westerholz, Germany.}
\usage{
create_measurement_WHZ(capture_data)
}
\arguments{
\item{capture_data}{Data frame. Output from \code{\link{create_capture_WHZ}}.}
}
\value{
A data frame.
}
\description{
Create measurement data table in standard format for data from Westerholz, Germany.
}
