% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_multiple.R
\name{create_multiple}
\alias{create_multiple}
\title{Run multiple pipelines}
\usage{
create_multiple(combined = FALSE, torun_path = NULL)
}
\arguments{
\item{combined}{Logical. If TRUE, all pipeline outputs are appended to create
four .csv files. If FALSE, individual .csv files are made for each
population.}

\item{torun_path}{File path. Data frame with two columns: ID, identifier for each
pipeline to run. N.B. These are NOT the same as PopID because in some cases
(e.g. NIOO), one pipeline is used for many PopIDs. Instead, this should be
the letter code at the end of the format function (e.g. NIOO, VEL). path,
directory where raw data is stored. If left blank, `choose.dir()` will be
called to identify the directory. If the data frame is stored as a csv,
`torun_path` must be a file path. If the data frame is stored in the global
environment, no value is needed for `torun_path`.}
}
\value{
Generate 4 .csv files.
}
\description{
Run multiple pipelines as specified by the user. If requested, pipeline
outputs are combined.
}
