% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HAR.R
\name{create_nestling_HAR}
\alias{create_nestling_HAR}
\title{Create nestling data capture table for Harjavalta, Finland.}
\usage{
create_nestling_HAR(db, Brood_data)
}
\arguments{
\item{db}{Location of primary data from Harjavalta, Finland.}

\item{Brood_data}{Output of \code{\link{create_brood_HAR}}.}
}
\value{
A data frame.
}
\description{
Create nestling data capture table for data from Harjavalta, Finland. This is
used inside \code{\link{create_capture_HAR}}.
}
