% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KEV.R
\name{create_nestling_KEV}
\alias{create_nestling_KEV}
\title{Create nestling data capture table for Kevo, Finland.}
\usage{
create_nestling_KEV(db, Brood_data)
}
\arguments{
\item{db}{Location of primary data from Kevo, Finland.}

\item{Brood_data}{Output of \code{\link{create_brood_KEV}}.}
}
\value{
A data frame.
}
\description{
Create nestling data capture table for data from Kevo, Finland. This is
used inside \code{\link{create_capture_HAR}}.
}
