% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whitelist.R
\name{create_whitelist}
\alias{create_whitelist}
\title{Create whitelist}
\usage{
create_whitelist(db = paste0(getwd(), "/data/"))
}
\arguments{
\item{db}{File path. Location to save whitelist. Default pipelines/data folder.}
}
\value{
List of 4 dataframes:
\item{Brood_whitelist}{Whitelisted records in Brood_data.}
\item{Capture_whitelist}{Whitelisted records in Capture_data.}
\item{Individual_whitelist}{Whitelisted records in Individual_data.}
\item{Location_whitelist}{Whitelisted records in Location_data.}
}
\description{
A function to create a whitelist of values that are verified as correct by data owners but previously marked as warning/error.
}
\details{
When a data owner is sent a quality check report, they can determine whether the flagged records are a) unverifiable, b) verifiable to be inccorect, or c) verifiable to be correct. If a), it cannot be known for certain whether the information should be kept or removed. They are continued to be flagged as improbable/impossible, and it is up to the user to decide how they treat them in analysis. If b), the data owner is expected to fix this in the raw data, and it will automatically be no longer flagged in future reports. If c), the records will be whitelisted for future quality checks.

To whitelist records, there are to pieces of information we use: the unique identifier of the record (e.g., BroodID in Brood_data) and the unique identifier of the quality check that resulted in the original warning/error (e.g., B2).
}
