% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_paradox_db.R
\name{extract_paradox_db}
\alias{extract_paradox_db}
\title{Convert paradox data file to data frame.}
\usage{
extract_paradox_db(path, file_name)
}
\arguments{
\item{path}{Directory where paradox files are located.}

\item{file_name}{File name of paradox file (.db) that will be converted.}
}
\value{
A data frame
}
\description{
Converts .db paradox data file into R data frame.
}
