% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_NIOO.R
\name{extract_species}
\alias{extract_species}
\title{Determine major species studied for each population.}
\usage{
extract_species(db = NULL)
}
\arguments{
\item{db}{Location of database file.}
}
\value{
A data frame with all major species for each population
}
\description{
For each of the main populations, return the names
of all species where >100 broods have been recorded
over the study period.
}
