% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{find_box}
\alias{find_box}
\title{Cut origin/destination box information into plot/box number}
\usage{
find_box(string, position = 1)
}
\arguments{
\item{string}{A vector. The origin/destination column.}

\item{position}{Start position to search along the character string.}
}
\value{
A vector of length 1 (for box number only) or 2 (for plot and box
  number respectively).
}
\description{
In MON primary data, the origin/destination of a cross-fostering or
translocation can have different formats:
}
\details{
- Box number only. This can be numeric (11) or alpha-numeric (11bis). - Plot
and box number. This is alpha numeric (ava13) - No information (NA)

We need to identify when the box number starts (the first numeric) and
separate plot and box number if both are present.
}
\examples{
#Return box number if only box number is provided
find_box("11")

#Return separate plot and box number if both provided
find_box("ava11")

#Still return box number of it is alpha-numeric
find_box("ava11bis")

#Return NA if no information is provided
find_box(NA)
}
