% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{format_AMM}
\alias{format_AMM}
\title{Construct standard format for data from Ammersee, Germany}
\usage{
format_AMM(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
4 data tables in the standard format (version 1.1.0). When `output_type = "R"`, a list of 4 data frames corresponding to the 4 standard data tables and 1 character vector indicating the protocol version on which the pipeline is based. When `output_type = "csv"`, 4 .csv files corresponding to the 4 standard data tables and 1 text file indicating the protocol version on which the pipeline is based.
}
\description{
A pipeline to produce the standard format for the nest box population in Ammersee,
Germany, administered by Niels Dingemanse.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{Species}: Species code 5 refers to Willow tit/Marsh tit. As we cannot
confidently identify species, these are considered Unknown and removed. This
currently only occurs once in the data. Similarly, species code 0 refers to
'unknown' and is treated as NA and removed.

\strong{ClutchType_observed}: Code 6 'Replacement First Replacement' is considered to be
a 2nd replacement after the first (i.e. not successful clutch) and is defined as 'replacement'.
Code 4 'Replacement second' is counted as 'second' in our method because it has come after at least
one successful clutch. Code 5 'Probably replacement first or second' is considered unknown and treated
as NA.

\strong{ClutchSize}: In a small number of cases (13 as of 2019 data) there are eggs from multiple species.
In this case, we are giving the clutch size as the number of eggs of both species. We are thinking
about ways to more effectively deal with multi-species clutches.

\strong{BroodSize}: Use NumberHatched to determine observed BroodSize. BroodSize_min is equal to BroodSize
observed. BroodSize_max is NumberHatched + ErrorHatched (i.e. number of eggs never observed either alive
or dead, so status was unknown).

\strong{NumberFledged}: Use NumberFledged to determine observed number of fledglings.
NumberFledged_min is equal to BroodSize_observed.
NumberFledged_max is NumberFledged + ErrorFledged (i.e. number of chicks never observed either alive
or dead, so status was unknown).

\strong{HatchDate}: The day a clutch hatched.
}
