% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BRG.R
\name{format_BRG}
\alias{format_BRG}
\title{Construct standard format for data from Bergen, Norway}
\usage{
format_BRG(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the nest box population in Bergen, Norway, administered by Adele Mennerat.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{Species}: Primarily Great tits and Blue tits.

\strong{IndvID}: Should be a character string of length 7 where the first two characters are either a letter or number and the last five characters are all numbers.

\strong{CaptureAlive}: A individual is assumed to be captured alive,
if the comment cell in the primary data file is empty.
If that cell contains the words "prior" and "died" or "dead",
individual is assumed to be captured dead.

\strong{ReleaseAlive}: We apply the same rules as for CaptureAlive.
}
