% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{format_CHO}
\alias{format_CHO}
\title{Construct standard summary for data from Choupal, Portugal.}
\usage{
format_CHO(db = file.choose(), Species = NULL, path = ".",
  debug = FALSE)
}
\arguments{
\item{db}{Location of database file.}

\item{Species}{A numeric vector. Which species should be included (EUring codes)? If blank will return all major species (see details below).}

\item{path}{Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This
includes: a) Histogram of continuous variables with mean/SD b) unique
values of all categorical variables.}
}
\value{
Generates 5 .csv files with data in a standard format.
}
\description{
A pipeline to produce a standard output for the great tit population
in Choupal, Portugal, administered by the University of Coimbra.
Output follows the HNB standard breeding data format.
}
\details{
This section provides details on data management choices that are unique to this data.
For a general description of the standard format please see XXXXX PLACE HOLDER!

\strong{NumberFledged}: This population has no estimation of actual fledgling numbers.
The last time nests are counted is 14 days post hatching. We use this as an estimation of fledling numbers.
This also affects \emph{ClutchType_calc} as \emph{NumberFledged} is used to estimated second clutches.

\strong{MinAge, RingAge}: Any individual caught as a chick was assumed to
have a EUring code of 1: 'Pullus: nestling or chick, unable to fly freely, still able to be caught by hand.'
Any individual not caught as a chick was assumed to have a EUring code of 4:
'After first full year: year of hatching unknown'.

\strong{ClutchType_observed}: In the raw data, there is no distinction between 'second' and 'replacement' clutches.
Any clutch recorded as '2nd' is assumed to be a 'second' clutch under our definitions.
'ClutchType_calc' may give a more appropriate estimate of clutch type for this data.
}
