% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_DLO.R
\name{format_DLO}
\alias{format_DLO}
\title{Construct standard format for data from Dlouhá Loučka, Czechia.}
\usage{
format_DLO(
  db = choose_directory(),
  species = NULL,
  site = NULL,
  optional_variables = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{site}{The three-letter code of the study site as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. For data owners with multiple study sites (e.g., NIOO, UAN) where a single
pipeline is used for many study sites, this argument is used to extract data from
individual study sites. For other pipelines that contain only one study site
this argument can be ignored.}

\item{optional_variables}{Optional variables (generated by standard utility functions) to be included in the pipeline output. Either a character vector of variable names or 'all' (if all optional variables should be included).}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 6 .csv files or 6 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the study site at
Dlouhá Loučka, Czechia, administered by the Palacky University.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{here}.

\strong{speciesID}: Minority species (i.e., short-toed treecreeper, common redstart)
and species with imprecise identifiers (e.g., Parus, FX) are excluded.
There is a small number of hybrid flycatcher broods. For parents, we know their exact species (i.e., FICALB or FICHYP).
For broods and chicks, we know these are mixed broods, and marked as FICHIB. Check with data owner.
Further, in cases where the species of the father is not specified, it is assumed to be equal to the brood/mother.

\strong{individualID}: Individual IDs start with 1 or 2 letters, followed by 4 to 8 digits. Check with data owner what the expected format is. Other IDs are removed.

\strong{chicks}: Except for individualIDs, little information is known about chicks. No capture/ringing date and/or chick age, which hampers the calculation of an exact age. Also speciesID is not known for mixed/hybrid broods.

\strong{captureAlive, releaseAlive}: All individuals are assumed to be captured and released alive.

\strong{captureRingNumber}: First captures of all individuals are assumed to be ringing events, and thus captureRingNumber is set to NA.

\strong{startYear}: Assume all boxes were placed in the first year of the study.

\strong{habitatID}: Assume that habitat type is 1.4: Forest - Temperate. Check with data owner.

\strong{Experiment data}: No experiments were conducted, resulting in an empty Experiment data table.
}
