% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_FOR.R
\name{format_FOR}
\alias{format_FOR}
\title{Construct standard format for data from Forstenrieder park, Germany}
\usage{
format_FOR(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format and 1 character vector indicating the protocol version on which the pipeline is based. When `output_type = "csv"`, 4 .csv files corresponding to the 4 standard data tables and 1 text file indicating the protocol version on which the pipeline is based.
}
\description{
A pipeline to produce the standard format for the nest box population in Forstenrieder park,
Germany, administered by Niels Dingemanse.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{Species}: Species code 0 refers to 'unknown' and is treated as NA and removed.

\strong{ClutchType_observed}: Code 4 'Replacement second' is counted as 'second' in our method because it has come after at least
one successful clutch. Code 5 'Probably replacement first or second' and Code 0 'unknown' are considered unknown and treated as NA.

\strong{BroodSize}: BroodSize is the number of chicks that hatched (maximum number of chicks between d1 and d3/4)
BroodSize_min is equal to BroodSize observed.
BroodSize_max is ClutchSize_observed - the number of unhatched eggs.

\strong{NumberFledged}: Use NumberFledged to determine observed number of fledglings.
NumberFledged_min is equal to BroodSize_observed.
NumberFledged_max is the number of chicks measured on d14(PARMAJ)/15(CYACAE)

#'\strong{FledgeDate}: Use NumberFledged to determine observed number of fledglings.
FledgeDate_min is equal to FledgeDate_observed - FledgedCheckInterval (time since the last fledge check)
FledgeDate_max is equal to FledgeDate_observed

\strong{LocationID}: Due to an ongoing density manipulation experiment (since 2020), nest box coordinates and entrance sizes can change (GT, BT or closed) between years. Therefore each location ID is unique every year.
\strong{Capture_data}: Capture records are excluded if they have missing values in key columns (IndvID, Species, or CaptureDate).
This removes individuals that escaped before being ringed, or records with incomplete capture information.
}
