% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GLA.R
\name{format_GLA}
\alias{format_GLA}
\title{Construct standard format for data from Glasgow, Scotland}
\usage{
format_GLA(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the nest box population in Glasgow, Scotland, administered by Davide Dominoni.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{Species}: Primarily great tits and blue tits.

\strong{BroodID}: A concatentation of PopID and the row number of the brood record (e.g. SAL-1).

\strong{IndvID}: Should be a 7 digit alphanumeric string. IndvIDs with a different number of characters are likely errors.
These are set to NA and removed.

\strong{CaptureDate}: Some individuals were not recorded in the ringing records, but were observed breeding at a monitored nest.
For these individuals, the CaptureDate is set as May 15 of the breeding year.
}
