% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GRO.R
\name{format_GRO}
\alias{format_GRO}
\title{Construct standard format for data from Grobla, Poland}
\usage{
format_GRO(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
4 data tables in the standard format (version 1.1.0). When `output_type = "R"`, a list of 4 data frames corresponding to the 4 standard data tables and 1 character vector indicating the protocol version on which the pipeline is based. When `output_type = "csv"`, 4 .csv files corresponding to the 4 standard data tables and 1 text file indicating the protocol version on which the pipeline is based.
}
\description{
A pipeline to produce the standard format for the nest box population in Grobla, Poland, administered by Mariusz Cichoń
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

#'\strong{CaptureDate}: Adults are typically captured two weeks into the nestling period.
If hatch date is not known for the nest, CaptureDate is set to two weeks after May 14th, which is the average hatch date.
}
