% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KIL.R
\name{format_KIL}
\alias{format_KIL}
\title{Construct standard format for data from Kilingi Nomme, Estonia.}
\usage{
format_KIL(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the great tit population
in Kilingi Nomme, Estonia, administered by Agu Leivits (until 1995) and
Raivo Mänd, Vallo Tilgar, Marko Mägi and Jaanis Lodjak (from 1995).
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.


\strong{BroodID}: BroodID has two formats in the final Brood_data. In the original primary data
until year 1992, the brood has individual numerical ID. In the original primary data from data
after the year 1995, there is no BroodID - therefore we create one using the "Breeding season_NestboxID_BreedingAttempt".

\strong{CaptureDate}: Date of capture for the primary data until the year 1992 are not available in the digitized format,
therefore the CaptureDate was artificially created as the 1st of May of the correspondent breeding year ("BreedingYear-05-01").

\strong{LayDate}: In the primary data, laying date is the date on which the first egg of a clutch is laid.
Expressed as days after 31st of March: 1 = April 1st, 31 = May 1st, etc.Use negative numbers for laying dates in March.

\strong{HatchDate_observed}: In the primary data (only data after 1995), hatch date is expressed as days after 30 of April: 1 = May 1st, 31 = June 1st, etc.

\strong{AvgTarsus}: Only available for data after 1995, where the measurements at 15 days old chicks are used.

\strong{ChickAge}: Note from data owner for the data until the 1992: Chicks ringed from age of 7-6 day up to fledging, most frequently in age of 8-10 day.
We used 8 days.

\strong{Location_data}: Coordinates of nestboxes can be provided by the data owner, we do not have them at this point.
}
