% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAL.R
\name{format_MAL}
\alias{format_MAL}
\title{Construct standard format for data from Malmo, Sweden}
\usage{
format_MAL(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the nest box population in Malmo, Sweden, administered by Caroline Isaksson
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{Species}: Data includes great tit  and blue tit.
Records with species RS or NA are excluded.

\strong{BroodID}: Generated as PopID-row_number (e.g. "MAL-1"). Only broods
with a known species receive a BroodID.

\strong{IndvID}: Taken directly from the Id column in the ringing records.
Some IDs may not be standard ring numbers.

\strong{CaptureDate}: Taken from the Date column in the ringing records.

\strong{ClutchType_observed}: Clutch type is derived from the 'Nesting
attempt' column in the primary data. Attempt 1 is classified as 'first' and
attempt 2 is classified as 'second'. No distinction between
'second' and 'replacement'.
}
