% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{format_MON}
\alias{format_MON}
\title{Construct standard format for data from Montpellier, France.}
\usage{
format_MON(db = utils::choose.dir(), species = NULL, pop = NULL,
  path = ".", debug = FALSE, output_type = "csv")
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This includes: a) Histogram of
continuous variables with mean/SD b) unique values of all categorical
variables.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the hole nesting bird population
in Montpellier, France, administered by CNRS Montpellier (Anne Charmantier and colleagues).
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard protocl please see
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong(Sex:) We ignore uncertainty in sex (e.g. category 4 = Male probable but not certain)

\strong(LocationID:) For birds caught in boxes. Location is Plot_NextboxNumber.

\strong(BroodID:) Broods are BreedingSeason_LocationID_day_month = BreedingSeason_Plot_NestboxNumber_day_month.

\strong(ClutchTypeObserved:) No clutch type recorded, only calculated clutch type is given.

\strong(Tarsus:) Left and right tarsus are measured. Right generally has more data, so we use this as our
measure of tarsus length. Currently, we assume everything is in Svensson's alternative, but there is supposedly
some change from before 1989. Need to ask Anne about this.

\strong(Age:) We translate observed age codes into EURING codes as follows:
\itemize{

\item P0 (Poussin bagué au nichoir ou en cavité naturelle/chick banded in box or natural cavity):
We give EURING code 1: Nestling or chick unable to fly freely.

\item J0
(Juvénile (oiseau de l’année) capturé entre le moment où il s’est envolé du nichoir et le 31 décembre)
(Juvenile (of this year) captured between fledging and Dec 31st of that year)
We give EURING code 3: First-year: full-grown bird hatched in the breeding season of this calendar year.

\item J1
(Juvénile (oiseau de l’année dernière) capturé après le 1er janvier et avant la mue post reproduction (juillet-août))
(Juvenile (of last year) capture between Jan 1st and July-Aug)
We give EURING code 5: 2nd year: a bird hatched last calendar year and now in its second calendar year.

\item A0
(Adulte capturé entre la mue post reproduction (environ juillet-août) et le 31 dec)
(Adult captured between July-Aug and Dec 31st)
We give EURING code 4: Afer first-year: full-grown bird hatched before this calendar year; year of hatching otherwise unknown.

\item A1
(Adulte capturé après le 1er janvier et avant la mue post reproduction (environ juillet-août).)
(Adult captured between Jan 1st and July-Aug)
We give EURING code 6: Afer 2nd year: full-grown bird hatched before last calendar year; year of hatching otherwise unknown.
This one is at least 2nd year because if it already had adult feathers before reproduction
it can't have been born last year.
}
}
