% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MTA.R
\name{format_MTA}
\alias{format_MTA}
\title{Construct standard format data from MTA-PE Evolutionary Ecology Group, University of Pannonia, Hungary (MTA)}
\usage{
format_MTA(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for bird study population
at the MTA-PE Evolutionary Ecology Group, Hungary, administered by
Gabor Seres & András Liker.
The data include 5 sites, we treat them as separate populations:
Veszprém VES, Balatonfüred BAL, Szentgál SZE, Vilma-puszta VIL and Gulya-domb GUL.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

This is a temporary pipeline, as the data owners are updating their storage system.

\strong{Individual_data}: Currently, only the information about adults is available, there is no
data included regarding chicks ringed.

\strong{Capture_date}: No capture date is available, we created artificial date as 1st of April in the format "BreedingSeason-04-01"
}
