% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PET.R
\name{format_PET}
\alias{format_PET}
\title{Construct standard format data from Institute of Biology, Karelian Research Centre, Russian Academy of Sciences, Petrozavodsk, Russia}
\usage{
format_PET(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for bird study population
at the Institute of Biology, Karelian Research Centre, Russian Academy of Sciences,
Petrozavodsk, Russia, administered by
Tolstoguzov Andrey.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.


\strong{Species}: Great tit (Parus major), Pied Flycatcher (Ficedula hypoleuca), Wryneck (Jynx torquilla).

\strong{ID format}: Should be a 7 character string.

\strong{CaptureDate}: When unknown, it is set to a default value of April 1st of that year.
}
