% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PEW.R
\name{format_PEW}
\alias{format_PEW}
\title{Construct standard format data from Peerdsbos West, Belgium (PEW)}
\usage{
format_PEW(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for bird study population
at the Peerdsbos West, Belgium, administered by Arne Iserbyt.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{IndvID}: Generally, the unique ID is an 8-character digit number reflecting
the metal ring number. In 4 cases, this information was not provided (NA), therefore
those observations were removed from the dataset (correspond to chicks ringed in
the year 2017 in the nestboxes 97, 98, 102, kk12).
Several individuals were not identified by the unique ring number, only by
the information reflecting the sex, nestbox and year (i.e. "Unknown_M2019NestK97").
Those IndvID were unified as "unringed" and are included in the Brood_data,
but not included in the Individual_data.

\strong{NumberFledged} Information provided by the data owner: The number of
fledged nestlings were not consistently monitored. However, nestlings were
ringed at day 14, so very close to fledging. In almost all cases, the number
of ringed chicks will be the same as number of fledged chicks.

\strong{BroodSize_observed} We used the information provided in the raw data
in the column indicating the number of chicks at day 3.

\strong{HatchDate_observed} For several broods the hatch date was not provided.
When neither hatch date nor capture date were provided, we input fake date
("breeding season-06-01", 1st of June of the corresponding breeding season).

\strong{CaptureDate} For several chicks, where the capture date was indicated
as "D14", the capture date was calculated as hatch date + 14 days.
For several adult individuals, where the capture date was missing and the record
corresponded to the only capture (observation) of the individual in the
breeding season, a fake date was input ("breeding season-06-01",
1st of June of the corresponding breeding season).
For several adult individuals, where the capture data was missing and the
capture method indicated that the capture happened during the incubation period,
the fake date was input ("breeding season, 04-01", 1st of April of the corresponding
breeding season).

\strong{Latitude, Longitude} The exact coordinates of the nestboxes are not available.
Data owner provided an map of the location, which can be georeferenced in the case of
interest or necessity of the data user. Currently, only general coordinates
for the location are provided.
}
