% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{format_PFN}
\alias{format_PFN}
\title{Construct standard format for data from East Dartmoor, UK}
\usage{
format_PFN(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the hole-nesting bird population
in East Dartmoor, UK, administered by PiedFlyNet.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard protocl please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{Species}: By default, pipeline outputs will include great tit \emph{Parus
major}; blue tit \emph{Cyanistes caeruleus}; pied flycatcher \emph{Ficedula
hypoleuca}; Eurasian nuthatch \emph{Sitta europaea}; coal tit \emph{Periparus
ater}; marsh tit \emph{Poecile palustris}, and common redstart \emph{Phoenicurus phoenicurus}.
Other minority species are excluded.

\strong{Primary data sources}: Brood data is extracted from data in a
brood table/nest monitoring format. 'Capture data is extracted from Malcolm Burgess'
version of the BTO ringing database for all birds marked within the PiedFlyNet.
The format of this ringing data is IPMR, the 'old' BTO ringing database format, and Malcolm
has his own pipeline to convert the new BTO ringing database format to the old one.
Since all chick measurements are complete in the brood table/nest monitoring data but some have
not been added to the BTO ringing data, the pipeline merges chick trait values from the former
into Capture data. Similarly, parent age information is more detailed in the brood table/nest
monitoring data, and therefore merged into Capture data (and given priority over age information
from the BTO ringing data).

\strong{Unidentified individuals}: Some individuals have not been given unique IDs. They appear
in the primary data as "RUNT", "ringed", "ringed left", "ringed right", "unringed", or "Unringed",
and the pipeline recognizes them as improper IDs as they do not follow the standard format of ring
numbers for this data (defined as a sequences of 1-9 letters followed by a sequence of 1-9 integer numbers).
In the Brood data, we assume unknown identity for individuals with improper IDs and treat ID as NA.
From Capture data and Individual data, these individuals are omitted entirely.

\strong{Re-ringing events}: Some individuals carry multiple rings (up to 3 in this data) throughout
their lives, and re-ringing events are recorded in the BTO ringing database. To correctly link data to
individuals (not just ring numbers), the pipeline assigns a unique individual identifier defined as
the number of the first ring ever given to this individual.

\strong{Resightings}: The capture primary data contains some resightings of birds. At present, these
records are excluded from the outputs because the standard format for capture data cannot distinguish
between recaptures and resightings in its current state. Functionality do do this (i.e. via a "capture type" variable)
may be added to a future update of the standard format.

\strong{Age calculation}: The age of each bird is estimated based on its history of ringing and recaptures.
Since some birds that breed in East Dartmoor may have been ringed in another study sites (= immigrants), the
pipeline calculates age based on capture histories from the entire BTO ringing database, not just ringing/captures
in locations within the East Dartmoor study site.

\strong{LayDate_observed and HatchDate_observed}: Information is provided as date of first egg (DFE) and
date of hatching (DH). These are given as integer number, and represent days after a set
starting date (day 0). Day 0 is defined as 31. March.

\strong{ChickAge}: For every capture, we estimate the age of a chick as the difference between the hatch date
taken from BroodIDFledged (in Individual_data) and the CaptureDate.
}
