% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SSQ.R
\name{format_SSQ}
\alias{format_SSQ}
\title{Construct standard summary for data from Santo Stefano Quisquina, Italy.}
\usage{
format_SSQ(db = NULL, Species = NULL, path = ".", debug = FALSE)
}
\arguments{
\item{db}{Location of database file.}

\item{Species}{A numeric vector. Which species should be included (EUring codes)? If blank will return all major species (see details below).}

\item{path}{Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This
includes: a) Histogram of continuous variables with mean/SD b) unique
values of all categorical variables.}
}
\value{
Generates 5 .csv files with data in a standard format.
}
\description{
A pipeline to produce a standard output for the great and blue tit population
in Santo Stefano Quisquina, Sicly, Italy, administered by Camillo Cusimano
and Daniel Campobello.
}
\details{
This section provides details on data management choices that are unique to this data.
For a general description of the standard format please see XXXXX PLACE HOLDER!
}
