% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{format_UAN}
\alias{format_UAN}
\title{Construct standard summary for data from the Univeristy of Antwerp.}
\usage{
format_UAN(db = choose.dir(), Species = NULL, path = ".",
  debug = FALSE)
}
\arguments{
\item{db}{Directory where raw data files are stored.}

\item{Species}{A numeric vector. Which species should be included (EUring codes)? If blank will return all major species (see details below).}

\item{path}{Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This
includes: a) Histogram of continuous variables with mean/SD b) unique
values of all categorical variables.}
}
\value{
Generates 5 .csv files with data in a standard format.
}
\description{
A pipeline to produce a standard output for 2 hole-nesting bird study populations
administered by the University of Amsterdam.
Output follows the HNB standard breeding data format.
}
\details{
This section provides details on data management choices that are unique to the NIOO database.
For a general description of the standard format please see XXXXX PLACE HOLDER!

\strong{ClutchType_observed}: The raw data distinguishes second and third
nests and first second and third replacements. We group these all as 'second'
and 'replacement'.

\strong{ExperimentID}: Experimental codes are provided in their original
format. These still need to be translated into our experimental groups.

\strong{Tarsus}: Tarsus is measured using Svennson's Standard in early years and
Svennson's Alternative in later years. When Svennson's Alternative is available this is used,
otherwise we use converted Svensson's Standard, using \code{\link[HNBStandFormat]{convert_tarsus}}.
}
