% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VOM.R
\name{format_VOM}
\alias{format_VOM}
\title{Construct standard format for data from Vomb Fure, Sweden}
\usage{
format_VOM(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the nest box population in Vomb Fure, Sweden, administered by Juli Broggi and Jan-Ake Nilsson.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{IndvID}: Should have the form: "2AA11111" - The first character is always a 2, the second two characters are letters,
and the last five characters are numbers for a total of 8 characters.

\strong{CaptureDate}: Set as the lay date for the nest.
}
