% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WHZ.R
\name{format_WHZ}
\alias{format_WHZ}
\title{Construct standard format for data from Westerholz, Germany}
\usage{
format_WHZ(
  db = choose_directory(),
  species = NULL,
  site = NULL,
  optional_variables = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{site}{The three-letter code of the study site as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. For data owners with multiple study sites (e.g., NIOO, UAN) where a single
pipeline is used for many study sites, this argument is used to extract data from
individual study sites. For other pipelines that contain only one study site
this argument can be ignored.}

\item{optional_variables}{Optional variables (generated by standard utility functions) to be included in the pipeline output. Either a character vector of variable names or 'all' (if all optional variables should be included).}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 6 .csv files or 6 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the nest box population in Westerholz,
Germany, administered by Bart Kempenaers, Max Planck Institute for Biological Intelligence, Seewiesen, Germany.
}
\details{
This section provides details on data management choices that are unique to this data. For a general description of the standard format please see \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{here}.

\strong{broodID}: Brood IDs are of the form <year_nest box number_nesting attempt>.

\strong{individualID}: Individual IDs are mostly 7 characters long, where the first and third are capital letters, and the remaining are digits. Other IDs (e.g., starting with X) are used too, but they seem to indicate IDs for individuals that died prematurely. Verify with data owner.

\strong{captureRingNumber, releaseRingNumber}: First captures of all individuals are assumed to be ringing events, and thus captureRingNumber is set to NA. Only when individual IDs follow the expected format (see \strong{individualID}), the ID is considered to be a ring number and stored in captureRingNumber and/or releaseRingNumber. When, for example, the ID is longer than 7 characters, the ID is considered to be a place holder rather than a ring number, and captureRingNumber and releaseRingNumber are set to NA.

\strong{treatmentID}: Some broods, adult captures, and chick captures are recorded as experimental. Treatment IDs are of the form <year_nest box number>. It seems that there is more info on experiments. Ask data owner.

\strong{startYear}: Assume all boxes were placed in the first year of study.

\strong{habitatID}: Assume that habitat type is 1.4: Forest - Temperate. Check with data owner.
}
