% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WYT.R
\name{format_WYT}
\alias{format_WYT}
\title{Construct standard summary for data from Wytham Woods, UK.}
\usage{
format_WYT(db = choose.dir(), path = ".", debug = FALSE)
}
\arguments{
\item{db}{Directory path. Location of primary data.}

\item{path}{File path. Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This includes: a) Histogram of
continuous variables with mean/SD b) unique values of all categorical
variables.}
}
\value{
Generates 4 .csv files with data in a standard format.
}
\description{
A pipeline to produce a standard output for the nest box population in Wytham
Woods, UK, administered by Edward Grey Institute Oxford (Ben Sheldon).
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see XXXXX
PLACE HOLDER!

\strong{ExperimentID}: There are experiment codes given, I have tried to
adapt these to the ExperimentID categories described in the standard
protocol; however, these need to be checked by data owners. For now I assume:
- Egg manipulation (code 2) is clutch size manipulation.
- Chick manipulation (code 3) is cross fostering.
- Alter temperature in nest box (code 7) only affects phenology.
- Feeding manipulation (code 8) affects phenology.
- Altering parasites, predation, competition and territory quality (codes 9, 11, 12, 13) all affect survival.

\strong{Species}: We include nests form blue tits, great tits, coal tits,
marsh tits, and nuthatches. Currently, mixed broods are treated as having no
species (NA), but we will fix this. There is one brood with Species 'w',
which I suspect is willow tit. This is currently ignored.

\strong{AvgEggMass}: There are two columns with mass data, one is a 'legacy'
column. When these overlap, they can differ up to 5g! I assume the 'legacy'
column is less prefered (due to it's name). I only use this data where no
other egg mass data is provided.

\strong{AvgChickMass}: As with AvgChickMass, there is also a 'legacy' column.
This is only used if the regular column is empty.

\strong{HatchDate}: As with AvgEggMass, there is also a 'legacy' column for
hatch date. This is only used if the regular column is empty.
}
