% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_debug_report.R
\name{generate_debug_report}
\alias{generate_debug_report}
\title{Generate a debug report for a pipeline}
\usage{
generate_debug_report(path, Pop, Brood_data, Capture_data, Indv_data)
}
\arguments{
\item{path}{File path where debug report will be created.}

\item{Pop}{Name of the population being debugged.}

\item{Brood_data}{A table containing the brood data.}

\item{Capture_data}{A table containing the capture data.}

\item{Indv_data}{A table containing the individual data.}
}
\value{
Creates a directory 'debug_report' which contains .csv and .pdf files
  with debug information.
}
\description{
Generates a debug report which include histograms of all continuous variables
and unique values for (some) categorical variable.
}
