% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_MON_metadata.R
\name{get_MON_metdata}
\alias{get_MON_metdata}
\title{Generate rough meta-data for Montpellier.}
\usage{
get_MON_metdata()
}
\value{
Nothing
}
\description{
This is a temporary function until we
get full meta-data. This is not intended to be
run, just to record how meta data were
extracted.
}
