% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{habitat_codes}
\alias{habitat_codes}
\title{Habitat types and descriptions}
\format{
A data frame with 125 rows and 3 variables
\describe{
 \item{habitatID}{Identifier for the habitat as provided by IUCN.}
 \item{habitatType}{Name or short description for the habitat as proved by IUCN.}
 \item{habitatDetails}{Detailed description or definition for the habitat as provided by IUCN.}
 }
}
\usage{
habitat_codes
}
\description{
Habitat descriptions according to IUCN Red List's Habitats Classification Scheme (version 3.1).
}
\keyword{datasets}
