% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{identify_PopID_MON}
\alias{identify_PopID_MON}
\title{Translate plots into corresponding study populations.}
\usage{
identify_PopID_MON(variable)
}
\arguments{
\item{variable}{Plot name.}
}
\value{
Three letter PopID or 'aviary'
}
\description{
Plots in MON are translated into one of 7 populations, including
'aviary' (for birds transferred to aviaries) and "MIS" for
miscellaneous populations.
}
\examples{
identify_PopID_MON("ava")
}
