% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{individual_check}
\alias{individual_check}
\title{Perform quality checks on individual data}
\usage{
individual_check(Individual_data, check_format = TRUE)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{check_format}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the check on variable format (i.e. \code{\link{check_format_individual}}) is included in the quality check. Default: \code{TRUE}.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
A wrapper that runs all single checks related to \code{Individual_data}.
}
\details{
The following individual data checks are performed:
\itemize{
\item \strong{I1}: Check if the formats of each column in \code{Individual_data} match with the standard format using \code{\link{check_format_individual}}.
}
}
