% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{location_check}
\alias{location_check}
\title{Perform quality checks on location data}
\usage{
location_check(Location_data, check_format = TRUE)
}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}

\item{check_format}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the check on variable format (i.e. \code{\link{check_format_location}}) is included in the quality check. Default: \code{TRUE}.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
A wrapper that runs all single checks related to \code{Location_data}.
}
\details{
The following location data checks are performed:
\itemize{
\item \strong{L1}: Check if the formats of each column in \code{Location_data} match with the standard format using \code{\link{check_format_location}}.
}
}
