% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{mean_countObs}
\alias{mean_countObs}
\title{Return mean and number of observations of a data vector}
\usage{
mean_countObs(x)
}
\arguments{
\item{x}{A single row of relevant columns in a data frame}
}
\value{
A vector containing a calculated mean and a number of observations
}
\description{
Return mean and number of observations of a data vector
}
