% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{pipeline_params}
\alias{pipeline_params}
\title{Parameter documentation for all pipelines}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{site}{The three-letter code of the study site as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. For data owners with multiple study sites (e.g., NIOO, UAN) where a single
pipeline is used for many study sites, this argument is used to extract data from
individual study sites. For other pipelines that contain only one study site
this argument can be ignored.}

\item{optional_variables}{Optional variables (generated by standard utility functions) to be included in the pipeline output. Either a character vector of variable names or 'all' (if all optional variables should be included).}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\description{
Parameter documentation for all pipelines
}
