% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_debug_hist.R
\name{plot_debug_hist}
\alias{plot_debug_hist}
\title{Generate histograms of continuous data for debugging}
\usage{
plot_debug_hist(table, variable)
}
\arguments{
\item{table}{A table containing data of interest.}

\item{variable}{A character string of the variable name.}
}
\value{
A histogram generated with ggplot.
}
\description{
Generates histograms for any provided continuous variable. When editing
pipelines, these allow the user to check that outputs are as expected. This
functionality will eventually be integrated into the data quality check
pipeline.
}
