% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_popmap.R
\name{plot_popmap}
\alias{plot_popmap}
\title{Create a map of all known populations}
\usage{
plot_popmap(scale = 2, filename = NULL)
}
\arguments{
\item{scale}{Scale of the saved jpeg image.}

\item{filename}{Name of saved file. Also allows for the filetype to be
changed. By default, uses "Population_map.jpeg".}
}
\value{
A jpeg file in the working directory
}
\description{
Plot all known populations. Point colour represents whether data has been
provided.
}
\examples{
#Create jpeg map
plot_popmap()
}
