% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chick_mass_cutoffs.R
\name{predictNLS}
\alias{predictNLS}
\title{Predict MC CIs.}
\usage{
predictNLS(object, newdata, level = 0.95, nsim = 10000, ...)
}
\arguments{
\item{...}{}
}
\description{
Code taken from https://www.r-bloggers.com/predictnls-part-1-monte-carlo-simulation-confidence-intervals-for-nls-models/.
Approach seems reasonable
}
