% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{A wrapper function to perform a quality check on pipeline outputs}
\usage{
quality_check(R_data, output_format = "both")
}
\arguments{
\item{R_data}{Output of pipeline as an R object. Generated using
'output_type = R' in \code{\link{run_pipelines}}.}

\item{output_format}{Format of report: "html", "pdf", or "both" (default).}
}
\value{
A summary dataframe of test warnings and errors for use in testthat,
  and a report of the line-by-line list of warnings and errors.
}
\description{
A wrapper function that performs a quality check and produces a report.
}
\details{
Combines the four data frame-specific wrappers: \code{\link{brood_check}}, \code{\link{capture_check}}, \code{\link{individual_check}} and \code{\link{location_check}}
}
\examples{
\dontrun{

CHO <- run_pipelines(PopID = "CHO", output_type = "R")
quality_check(CHO)

}
}
