% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{site_codes}
\alias{site_codes}
\title{Study site names and locations}
\format{
A data frame with 114 rows and 11 variables
\describe{
 \item{siteID}{Three-letter identifier for the study site.}
 \item{siteName}{Name of the study site.}
 \item{country}{Name of the country in which the study site is located.}
 \item{countryCode}{Standard code for the country, using ISO 3166-1 alpha-2: \url{https://www.iso.org/iso-3166-country-codes.html}}
 \item{institutionID}{Three- or four-letter identifier for data owner/institution. Note, multiple study sites can have
 the same data owner.}
 \item{institutionCode}{Name of owner/institution.}
 \item{decimalLatitude}{Geographic latitude of the geographic center of the study site in decimal degrees.}
 \item{decimalLongitude}{Geographic longitude of the geographic center of the study site in decimal degrees.}
 \item{locationAccordingTo}{Source of location information. Either "data owner" or "data".}
 \item{data}{Logical. Are the data available through SPI-Birds?}
 \item{standardFormat}{Logical. If data are available, are they standardized?}
 }
}
\usage{
site_codes
}
\description{
Information and identifying codes for all study sites and data owners.
}
\keyword{datasets}
