% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{site_species_combos}
\alias{site_species_combos}
\title{Combination of all siteID and speciesID from all pipelines.}
\format{
A data frame with 143 rows and 2 variables.
\describe{
   \item{siteID}{Three-letter site ID.}
   \item{speciesID}{Six-letter species ID.}
}
}
\usage{
site_species_combos
}
\description{
A data frame with each row containing a unique site and species combination.
}
\keyword{datasets}
