% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_general_format.R
\name{test_col_present}
\alias{test_col_present}
\title{Test column presence}
\usage{
test_col_present(pipeline_output, table, protocol_version, debug = FALSE)
}
\arguments{
\item{pipeline_output}{A list of data frames returned from format_X, where X is the pipeline code.}

\item{table}{Character string. Which data table should be tested? One of: "Brood", "Individual", "Capture", "Location", "Measurement" or "Experiment".}

\item{protocol_version}{The protocol version of the SPI-Birds
standard data being used to process the primary data. Either "1.0.0", "1.1.0", or "2.0.0".}

\item{debug}{Logical (TRUE/FALSE) for whether the function returns the test output (FALSE) or a vector of columns that failed this test (TRUE).}
}
\value{
See `debug`.
}
\description{
Test whether expected columns are present
}
\details{
This function is intended to be used in the test for every pipeline in order to ensure consistency between pipelines.
}
