% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_protocol_wrapper.R
\name{test_protocol_compliance}
\alias{test_protocol_compliance}
\title{Test protocol compliance}
\usage{
test_protocol_compliance(pipeline_output, ID_format = FALSE)
}
\arguments{
\item{pipeline_output}{A list of data frames returned from format_X,
where X is the pipeline code.}

\item{ID_format}{Regular expression pattern for validating ID formats.
Default is FALSE, which skips this test.}
}
\value{
Invisible TRUE if all tests pass, errors otherwise.
}
\description{
Runs all protocol validation tests from test_general_format.R
}
