% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_general_format.R
\name{test_time_columns}
\alias{test_time_columns}
\title{Test time columns}
\usage{
test_time_columns(pipeline_output, protocol_version, debug = FALSE)
}
\arguments{
\item{pipeline_output}{A list of data frames returned from format_X,
where X is the pipeline code.}

\item{protocol_version}{The protocol version of the SPI-Birds
standard data being used to process the primary data. Either "1.0.0",
"1.1.0", or "2.0.0".}

\item{debug}{Logical (TRUE/FALSE) for whether the function returns the
test output (FALSE) or a list of columns (and the unexpected values in
them) that failed this test (TRUE).}
}
\value{
See `debug`.
}
\description{
Time columns should only include certain ranges.
For example, 'CaptureTime' should have hours between 0 and 23 and
minutes between 0 and 59.
This test makes sure that there are no unexpected or non-valid
values in these time columns.
}
