% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VLI_twitter.R
\name{VLI_twitter}
\alias{VLI_twitter}
\title{Create plots to introduce the Hoge Veluwe population on twitter}
\usage{
VLI_twitter()
}
\value{
Generate two plots. One .jpeg showing the population on the global map and study species.
One .gif showing number of nests sampled for each species over time.
}
\description{
Create plots to introduce the Hoge Veluwe population on twitter
}
