% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{whitelist}
\alias{whitelist}
\title{Whitelist of records to be used in quality check}
\format{
A list of 4 data frames:
\itemize{
   \item{Brood_whitelist}
   \item{Capture_whitelist}
   \item{Individual_whitelist}
   \item{Location_whitelist}
}

Each data frame has 3 variables:
\itemize{
   \item{PopID, unique identifier for population.}
   \item{A unique identifier for records in each particular data frame (i.e., BroodID, CaptureID, IndvID, LocationID).}
   \item{CheckID, unique identifier for individual quality check.}
}
}
\description{
Whitelist of records that were flagged as warning/error but have been verified by the data owner to be correct.
}
