% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{archive}
\alias{archive}
\title{Archive SPI-Birds raw data}
\usage{
archive(
  data_folder = choose_directory(),
  update_type = "major",
  PopID,
  new_data_path
)
}
\arguments{
\item{data_folder}{Path. Folder containing all current SPI-Birds data (N drive)}

\item{update_type}{Character string. Have we received a new year of data ("major")
or data with bug/error fixes ("minor")?}

\item{PopID}{Character string. 3-4 letter code for population.
Should be the same as the folder name where data are saved.}

\item{new_data_path}{File path. Updated data to use as current data in the system.
Where a population uses multiple files provide a list of file paths.}
}
\value{
Returns nothing.
}
\description{
Archive SPI-Birds raw data
}
