% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{brood_check}
\alias{brood_check}
\title{Perform quality checks on brood data}
\usage{
brood_check(Brood_data, Individual_data, check_format = TRUE)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{check_format}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the check on variable format (i.e. \code{\link{check_format_brood}}) is included in the quality check. Default: \code{TRUE}.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
A wrapper that runs all single checks related to \code{Brood_data}.
}
\details{
The following brood data checks are performed:
\itemize{
\item \strong{B1}: Check if the formats of each column in \code{Brood_data} match with the standard format using \code{\link{check_format_brood}}.
\item \strong{B2}: Compare clutch size and brood size per brood using \code{\link{compare_clutch_brood}}.
\item \strong{B3}: Compare brood size and fledgling number per brood using \code{\link{compare_brood_fledglings}}.
}
}
