% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{brood_data_template}
\alias{brood_data_template}
\title{brood_data_template template}
\format{
Tibble
}
\description{
Tibble giving names and associated classes for all the columns in the Brood data for the standard protocol v1.1.0
}
