% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_clutchtype.R
\name{calc_clutchtype}
\alias{calc_clutchtype}
\title{Calculate clutch type (when laying date is in date format)}
\usage{
calc_clutchtype(data, protocol_version = "1.0", na.rm = FALSE)
}
\arguments{
\item{data}{Data frame. Brood data to use for conversion}

\item{protocol_version}{Character string. Either "1.0" (default) or "1.1".}

\item{na.rm}{Logical. Should NA's be removed and treated as 0s. If FALSE, NAs
are treated as true unknowns. If TRUE, NAs are treated as 0s.}
}
\value{
A character vector with either 'first', 'replacement', 'second', or
  NA
}
\description{
Use info on laying date and fledge data to determine the clutch type of a
given brood
}
\examples{
#Create fake dataset
set.seed(666)
dat <- tibble::tibble(PopID = "TEST", Species = "PARMAJ",
FemaleID = sample(LETTERS[1:7], size = 100, replace = TRUE),
NumberFledged = rpois(n = 100, lambda = 1),
#Create 100 fake broods
BreedingSeason = sample(c(seq(2000, 2012, 1)), size = 100, replace = TRUE),
LayDate = as.Date(paste(BreedingSeason,
                           sample(c(4, 5, 6), size = 100, replace = TRUE),
                           sample(seq(1, 31, 1), size = 100, replace = TRUE), sep = "-"),
                           format = "\%Y-\%m-\%d"))
calc_clutchtype(data = dat, na.rm = FALSE)
}
