% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_functions.R
\name{calc_clutchtype}
\alias{calc_clutchtype}
\title{Calculate clutch type (when laying date is in date format)}
\usage{
calc_clutchtype(data, protocol_version = "1.0", na.rm = FALSE)
}
\arguments{
\item{data}{Data frame with brood information.}

\item{protocol_version}{Character string. The protocol version of the SPI Birds
standard data being used to process the primary data. Either "1.0" (default), "1.1", or "1.2".}

\item{na.rm}{Logical. Should NAs be removed and treated as 0s? If TRUE, NAs are treated as 0s. If FALSE, NAs
are treated as true unknowns.}
}
\value{
A character vector with either 'first', 'replacement', 'second', or NA (v1.0 or v1.1), or a data frame with calculatedClutchType (v1.2), which takes either 'first', 'replacement', 'second', or NA.
}
\description{
Use info on laying date and fledge data to determine the clutch type of a
given brood.
}
\examples{
#Create fake dataset
set.seed(666)

dat <- tibble::tibble(PopID = "TEST", Species = "PARMAJ",
                      FemaleID = sample(LETTERS[1:7], size = 100, replace = TRUE),
                      NumberFledged = rpois(n = 100, lambda = 1),
                      #Create 100 fake broods
                      BreedingSeason = sample(c(seq(2000, 2012, 1)), size = 100, replace = TRUE),
                      LayDate = as.Date(paste(BreedingSeason,
                                              sample(c(4, 5, 6), size = 100, replace = TRUE),
                                              sample(seq(1, 31, 1), size = 100, replace = TRUE), sep = "-"),
                                        format = "\%Y-\%m-\%d"),
                      LayDate_observed = LayDate)

calc_clutchtype(data = dat, na.rm = FALSE)
}
