% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_cumfledge.R
\name{calc_cumfledge}
\alias{calc_cumfledge}
\title{Calculate cumulative number of fledgings}
\usage{
calc_cumfledge(x, na.rm = TRUE)
}
\arguments{
\item{x}{Column NumberFledged in the Brood_data table}

\item{na.rm}{Logical. If TRUE, returns cumulative number of fledglings
where NA is assumed to be 0. If FALSE, returns a vector of logicals
showing whether any nests before the current nest were unmeasured (NA).}
}
\value{
A vector of numerics (if na.rm = TRUE) or logicals (na.rm = FALSE)
}
\description{
For a given nest, determine the cumulative number of fledglings in all nests
before this. This is used to calculate ClutchType_calc. The function also
includes functionality to return whether number of fledglings in previous
nests was measured or not (i.e. were there NAs in any nests before the
current one). This is neccesary in populations where the number of fledglings
is not measured consistently
}
\examples{
#Assuming NA is 0
#Return vector of numerics.
fledge_calc(x = c(1, 3, NA, 4), na.rm = TRUE)

#Do not assume NA is 0.
#Return a vector of logicals showing whether an NA occurred before
#the current record.
fledge_calc(x = c(1, 3, NA, 4), na.rm = FALSE)
}
