% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_chick_mass_cutoffs.R
\name{calculate_chick_mass_cutoffs}
\alias{calculate_chick_mass_cutoffs}
\title{Determine chick growth curves}
\usage{
calculate_chick_mass_cutoffs(Capture_data, plot = FALSE)
}
\arguments{
\item{Capture_data}{Data frame with capture data on chicks of 1 species in 1 population}

\item{plot}{Create plot of mass data and logistic growth curve. Default: FALSE.}
}
\value{
A data frame with reference values for mass at each age
}
\description{
Fit logistic growth curve to chick mass data to calculate age-specific reference values.
}
\examples{
\dontrun{
# Run pipeline for Hoge Veluwe great tits
HOG_PARMAJ_data <- run_pipelines(PopID = "HOG", Species = "PARMAJ", output_type = "R")

# Calculate reference values using logistic growth model
calculate_chick_mass_cutoffs(HOG_PARMAJ_data, plot = TRUE)
}
}
