% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{capture_check}
\alias{capture_check}
\title{A wrapper function to perform quality checks on capture data}
\usage{
capture_check(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}
}
\value{
A list of 3 items: a summary data frame of all checks, a record-by-record list of warnings, and a record-by-record list of errors.
}
\description{
A wrapper that runs all single checks related to `Capture_data`.
}
\details{
\strong{Capture check 1}: Check if the formats of each column in `Capture_data` match with the standard format using \code{\link{check_format_capture}}.
\strong{Capture check 2}: Check capture variable values against reference values using \code{\link{check_values_capture}}.
}
