% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{capture_check}
\alias{capture_check}
\title{Perform quality checks on capture data}
\usage{
capture_check(
  Capture_data,
  Location_data,
  Brood_data,
  Individual_data,
  approved_list,
  output,
  skip
)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{Location_data}{Data frame. Location data output from pipeline.}

\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
A wrapper that runs all single checks related to \code{Capture_data}.
}
\details{
The following capture data checks are performed:
\itemize{
\item \strong{C1a-b}: Check capture variable values against reference values using \code{\link{check_values_capture}}. Capture variables checked for adults and chicks: Mass and Tarsus.
\item \strong{C2}: Check chick age (in numbers of days since hatching) using \code{\link{check_chick_age}}.
\item \strong{C3}: Check that adults caught on nest are listed as parents using \code{\link{check_adult_parent_nest}}.
\item \strong{C4}: Check that the age of subsequent captures of the same individual is correct using \code{\link{check_age_captures}}.
\item \strong{C5}: Check if individuals in Capture_data appear in Individual_data using \code{\link{check_captures_individuals}}.
\item \strong{C6}: Check that capture locations appear in Location_data using \code{\link{check_capture_locations}}.
}
}
