% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{capture_data_template}
\alias{capture_data_template}
\title{capture_data_template template}
\format{
Tibble

Data frame with 1 row and 23 variables:
}
\description{
Tibble giving names and associated classes for all the columns in the Capture data for the standard protocol v1.1.0
}
