% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_BroodID_chicks}
\alias{check_BroodID_chicks}
\title{Check that chicks have BroodID}
\usage{
check_BroodID_chicks(
  Individual_data,
  Capture_data,
  Location_data,
  approved_list
)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}

\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{Location_data}{Data frame. Location data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that all chicks in Individual_data that are caught and ringed in a nest box have a BroodID. Individuals just ringed after fledging are regarded as chicks but are not associated with a BroodID.
}
\details{
Check ID: I3.
}
