% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_conflicting_sex}
\alias{check_conflicting_sex}
\title{Check conflicting sex}
\usage{
check_conflicting_sex(Individual_data)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Check that the sex of individuals in Individual_data is recorded consistently. Individuals who have been recorded as both male ('M') and female ('F') will have conflicting sex ('C') in Individual_data.
}
