% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{check_coordinates}
\alias{check_coordinates}
\title{Check coordinates of locations}
\usage{
check_coordinates(
  Location_data,
  Brood_data,
  Capture_data,
  approved_list,
  output,
  skip,
  map
)
}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}

\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}

\item{map}{Logical. Produce map of capture locations?}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
\item{Maps}{A list of maps with capture locations.}
}
\description{
Check that the coordinates of locations recorded in Brood_data and/or Capture_Data are close to the centre point of the study site. Locations that are farther than 20 km will result in an error. It's optional to print the locations on a map, which is then visualized in the quality check report.
}
\details{
Check ID: L1.
}
