% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_format_brood}
\alias{check_format_brood}
\title{Check format of brood data}
\usage{
check_format_brood(Brood_data, approved_list)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the format of each column in the brood data match with the standard format.
}
\details{
Check ID: B1.
}
