% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{check_format_capture}
\alias{check_format_capture}
\title{Check format of capture data}
\usage{
check_format_capture(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Check if the formats of each column in the capture data match with the standard format.
}
