% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_format_capture}
\alias{check_format_capture}
\title{Check format of capture data}
\usage{
check_format_capture(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}
}
\value{
Check list, warning output, error output.
}
\description{
Check if the formats of each column in the capture data match with the standard format
}
