% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_format_individual}
\alias{check_format_individual}
\title{Check format of individual data}
\usage{
check_format_individual(Individual_data)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
Check list, warning output, error output.
}
\description{
Check if the formats of each column in the individual data match with the standard format
}
