% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_format_individual}
\alias{check_format_individual}
\title{Check format of individual data}
\usage{
check_format_individual(Individual_data)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Check that the format of each column in the individual data match with the standard format
}
