% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{check_format_location}
\alias{check_format_location}
\title{Check format of location data}
\usage{
check_format_location(Location_data)
}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Check if the formats of each column in the location data match with the standard format
}
