% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{check_format_location}
\alias{check_format_location}
\title{Check format of location data}
\usage{
check_format_location(Location_data)
}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}
}
\value{
Check list, warning output, error output.
}
\description{
Check if the formats of each column in the location data match with the standard format
}
