% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_parent_species}
\alias{check_parent_species}
\title{Check parent species}
\usage{
check_parent_species(Brood_data, Individual_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Check that the parents of broods are the same species.
}
