% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual_check.R
\name{check_unique_IndvID}
\alias{check_unique_IndvID}
\title{Check unique individual identifiers}
\usage{
check_unique_IndvID(Individual_data)
}
\arguments{
\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the individual identifiers (IndvID) are unique within populations.
}
