% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_values_brood}
\alias{check_values_brood}
\title{Check brood variable values against reference values}
\usage{
check_values_brood(Brood_data, var)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{var}{Character. Variable to check against reference values.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check variable values against species-specific reference values in brood data. Unusual values will result in a warning. Impossible values will result in an error. Variables that can be checked: LayDate, ClutchSize, HatchDate, BroodSize, FledgeDate, NumberFledged, AvgEggMass, AvgChickMass, AvgTarsus.
}
