% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_values_brood}
\alias{check_values_brood}
\title{Check brood variable values against reference values}
\usage{
check_values_brood(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Check variable values against species-specific reference values in brood data. Implausible values will result in a warning. Impossible values will result in an error. Variables that are checked: LayingDate, ClutchSize, HatchDate, BroodSize, FledgeDate, NumberFledged, AvgEggMass, AvgChickMass, AvgTarsus.
}
