% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_values_brood}
\alias{check_values_brood}
\title{Check brood variable values against reference values}
\usage{
check_values_brood(Brood_data, species)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{species}{Six-letter species ID to select species-specific reference values.}
}
\value{
Check list, warning output, error output.
}
\description{
Check variable values against species-specific reference values in brood data. Implausible values will result in a warning. Impossible values will result in an error. Variables that are checked: LayingDate, ClutchSize, HatchDate, BroodSize, FledgeDate, NumberFledged, AvgEggMass, AvgChickMass, AvgTarsus.
}
