% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{check_values_brood}
\alias{check_values_brood}
\title{Check brood variable values against reference values}
\usage{
check_values_brood(Brood_data, var, approved_list, output, skip)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{var}{Character. Variable to check against reference values.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check variable values against population-species-specific reference values in brood data. Reference values are based on the data if the number of observations is sufficiently large. Records for population-species combinations that are too low in number are only compared to reference values that are not data generated (see Details below).
}
\details{
\strong{ClutchSize_observed, BroodSize_observed, NumberFledged_observed} \cr
Check IDs: B5a-c
\itemize{
\item \emph{n >= 100}\cr
Records are considered impossible if they are negative or larger than 2 times the 99th percentile, and will be flagged as a potential error.
\item \emph{n < 100}\cr
Records are considered impossible if they are negative, and will be flagged as a potential error.
}

\strong{LayDate_observed, HatchDate_observed, FledgeDate_observed} \cr
Check ID: B5d-f
\itemize{
\item \emph{n >= 100}\cr
Date columns are transformed to Julian days to calculate percentiles. Records are considered impossible if they are earlier than January 1st or later than December 31st of the current breeding season, and will be flagged as a potential error.
\item \emph{n < 100}\cr
Date columns are transformed to Julian days to calculate percentiles. Records are considered impossible if they are earlier than January 1st or later than December 31st of the current breeding season, and will be flagged as a potential error.
}

Note: when the number of observations is too low to generate reference values, a message is added to the list of warnings.
}
