% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{check_values_capture}
\alias{check_values_capture}
\title{Check capture variable values against reference values}
\usage{
check_values_capture(Capture_data, species)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{species}{Six-letter species ID to select species-specific reference values.}
}
\value{
Check list, warning output, error output.
}
\description{
Check variable values against species-specific reference values in capture data. Implausible values will result in a warning. Impossible values will result in an error. Variables that are checked: Mass, Tarsus, WingLength, Age_obsv, Age_calc, Chick_age.
}
