% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture_check.R
\name{check_values_capture}
\alias{check_values_capture}
\title{Check capture variable values against reference values}
\usage{
check_values_capture(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Capture data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Check variable values against species-specific reference values in capture data. Unusual values will result in a warning. Impossible values will result in an error. Variables that are checked: Mass, Tarsus, WingLength, Age_obsv, Age_calc, Chick_age.
}
