% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{checkID_variable_combos}
\alias{checkID_variable_combos}
\title{CheckIDs and variable names}
\format{
Data frame with 2 variables:
\describe{
\item{CheckID}{Unique identifier for individual quality check.}
\item{Variable}{Variable name.}
}
}
\usage{
checkID_variable_combos
}
\description{
Data frame to link variable names and CheckIDs to be used in approve-listing of previously flagged records.
}
\keyword{datasets}
