% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checks_brood_params}
\alias{checks_brood_params}
\title{Parameter documentation for brood data checks}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\description{
Parameter documentation for brood data checks
}
