% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checks_brood_params}
\alias{checks_brood_params}
\title{Parameter documentation for brood data checks}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}
}
\description{
Parameter documentation for brood data checks
}
