% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checks_location_params}
\alias{checks_location_params}
\title{Parameter documentation for location data checks}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}
}
\description{
Parameter documentation for location data checks
}
