% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{checks_location_params}
\alias{checks_location_params}
\title{Parameter documentation for location data checks}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}
}
\description{
Parameter documentation for location data checks
}
