% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_brood_fledglings}
\alias{compare_brood_fledglings}
\title{Compare brood sizes and fledgling numbers}
\usage{
compare_brood_fledglings(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
Check list, warning output, error output.
}
\description{
Compare brood size and fledgling number per brood. In non-manipulated broods, brood size should be larger or equal to fledgling number. If not, the record will result in an error. In broods with clutch manipulation, brood size might be smaller than fledgling number. If so, the record will result in a warning.
}
