% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_brood_fledglings}
\alias{compare_brood_fledglings}
\title{Compare brood sizes and fledgling numbers}
\usage{
compare_brood_fledglings(Brood_data, approved_list, output)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Compare brood size and fledgling number per brood. In non-manipulated broods, brood size should be larger or equal to fledgling number. If not, the record will be flagged as a potential error. In broods with clutch manipulation, brood size might be smaller than fledgling number. If so, the record will be flagged as a warning.
}
\details{
Check ID: B2.
}
