% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_hatching_fledging}
\alias{compare_hatching_fledging}
\title{Compare hatching and fledging dates}
\usage{
compare_hatching_fledging(Brood_data, approved_list)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Compare hatching and fledging date per brood. Broods with hatching date later than fledging date will result in an error. Broods with hatching date earlier than fledging date but the difference in number of days is smaller than breeding time will result in a warning.
}
\details{
Check ID: B4.
}
