% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_laying_hatching}
\alias{compare_laying_hatching}
\title{Compare laying and hatching dates}
\usage{
compare_laying_hatching(Brood_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of whether the check resulted in any warnings or errors.}
\item{WarningOutput}{A list of row-by-row warnings.}
\item{ErrorOutput}{A list of row-by-row errors.}
}
\description{
Compare laying and hatching date per brood. Broods with laying date later than hatching date will result in an error. Broods with laying date earlier than hatching date but the difference in number of days is smaller than incubation time will result in a warning.
}
