% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brood_check.R
\name{compare_species_parents}
\alias{compare_species_parents}
\title{Compare species of parents}
\usage{
compare_species_parents(Brood_data, Individual_data)
}
\arguments{
\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Individual_data}{Data frame. Individual data output from pipeline.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
Check that the parents of broods are of the same species. Broods with parents of different species are flagged as 'warning' because multi-species broods are known to exist.
}
\details{
Check ID: B10.
}
