% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{create_brood_AMM}
\alias{create_brood_AMM}
\title{Create brood data table for Ammersee, Germany.}
\usage{
create_brood_AMM(connection, species_filter)
}
\arguments{
\item{connection}{Connection the SQL database.}

\item{species_filter}{Species six-letter codes from the standard protocol. Used to filter the data.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Ammersee, Germany.
}
