% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{create_brood_AMM}
\alias{create_brood_AMM}
\title{Create brood data table for Ammersee, Germany.}
\usage{
create_brood_AMM(connection)
}
\arguments{
\item{data}{Data frame. Primary data from Ammersee.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Ammersee, Germany.
}
