% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{create_brood_CHO}
\alias{create_brood_CHO}
\title{Create brood data table for Choupal, Portugal.}
\usage{
create_brood_CHO(data)
}
\arguments{
\item{data}{Data frame. Primary data from Choupal.}
}
\value{
A data frame.
first'. For ClutchType_observed we are not giving broods the value
replacement' as we don't have enough info.
}
\description{
Create brood data table in standard format for data from Choupal,
Portugal.
}
