% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GLA.R
\name{create_brood_GLA}
\alias{create_brood_GLA}
\title{Create brood data table for great tits and blue tits in Glasgow, Scotland.}
\usage{
create_brood_GLA(nest_data, rr_data)
}
\arguments{
\item{nest_data}{Data frame of nest data from Glasgow, Scotland.}

\item{rr_data}{Data frame of ringing records from Glasgow, Scotland.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for Glasgow, Scotland.
}
