% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GRO.R
\name{create_brood_GRO}
\alias{create_brood_GRO}
\title{Create brood data table for great tits and blue tits in Grobla, Poland.}
\usage{
create_brood_GRO(gro_data)
}
\arguments{
\item{gro_data}{Data frame of modified primary data from Grobla, Poland.}
}
\value{
A data frame.
}
\description{
Create brood data table for great tits and blue tits in Grobla, Poland.
}
