% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{create_brood_MAY}
\alias{create_brood_MAY}
\title{Create brood data table for Mayachino, Russia.}
\usage{
create_brood_MAY(gt_data, pf_data, species_filter, optional_variables = NULL)
}
\arguments{
\item{gt_data}{Data frame. Great tit data from Mayachino, Russia.}

\item{pf_data}{Data frame. Pied flycatcher data from Mayachino, Russia.}

\item{species_filter}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}.}

\item{optional_variables}{A character vector of names of optional variables (generated by standard utility functions) to be included in the pipeline output.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Mayachino, Russia.
}
