% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SSQ.R
\name{create_brood_SSQ}
\alias{create_brood_SSQ}
\title{Create brood data table for Santo Stefano Quisquina, Italy.}
\usage{
create_brood_SSQ(data, protocol_version)
}
\arguments{
\item{data}{Data frame. Primary data from Santo Stefano Quisquina.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from Santo Stefano
Quisquina, Italy
}
