% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{create_brood_UAN}
\alias{create_brood_UAN}
\title{Create brood data table for data from University of Antwerp, Belgium.}
\usage{
create_brood_UAN(
  BROOD_info,
  CAPTURE_info,
  species_filter,
  pop_filter,
  protocol_version
)
}
\arguments{
\item{BROOD_info}{Data frame. Primary brood data from University of Antwerp.}

\item{CAPTURE_info}{Data frame. Primary capture data from University of Antwerp.}

\item{species_filter}{6 letter species codes for filtering data.}

\item{pop_filter}{Population three letter codes from the standard protocol.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create brood data table in standard format for data from University of
Antwerp, Belgium.
}
