% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{create_capture_AMM}
\alias{create_capture_AMM}
\title{Create capture data table for Ammersee, Germany}
\usage{
create_capture_AMM(Brood_data, connection)
}
\arguments{
\item{Brood_data}{Data frame. Output from \code{\link{create_brood_AMM}}.}

\item{connection}{Connection the SQL database.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from Ammersee, Germany.
}
