% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{create_capture_AMM}
\alias{create_capture_AMM}
\title{Create capture data table for Ammersee, Germany}
\usage{
create_capture_AMM(Brood_data, dir, species_filter)
}
\arguments{
\item{Brood_data}{Data frame. Output from \code{\link{create_brood_AMM}}.}

\item{dir}{Path to directory containing the relevant table exports from the AMM Access database.}

\item{species_filter}{Species six-letter codes from the standard protocol. Used to filter the data.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from Ammersee, Germany.
}
