% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_capture_PFN}
\alias{create_capture_PFN}
\title{Create capture data table for EastDartmoor.}
\usage{
create_capture_PFN(db)
}
\arguments{
\item{db}{Location of primary data from EastDartmoor.}
}
\value{
A data frame with Capture data
}
\description{
Create capture data table for EastDartmoor.
}
