% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{create_capture_UAN}
\alias{create_capture_UAN}
\title{Create capture data table for data from University of Antwerp, Belgium.}
\usage{
create_capture_UAN(data, species_filter, pop_filter)
}
\arguments{
\item{data}{Data frame. Primary data from University of Antwerp.}

\item{species_filter}{6 letter species codes for filtering data.}

\item{pop_filter}{Population three letter codes from the standard protocol.}
}
\value{
A data frame.
}
\description{
Create capture data table in standard format for data from University of
Antwerp, Belgium.
}
