% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dummy_data.R
\name{create_dummy_data}
\alias{create_dummy_data}
\title{Create quality check dummy data}
\usage{
create_dummy_data(overwrite = TRUE)
}
\arguments{
\item{overwrite}{Overwrite existing file. Default: TRUE}
}
\value{
List of 4 dataframes:
\item{Brood_data}{Dummy brood data.}
\item{Capture_data}{Dummy capture data.}
\item{Individual_data}{Dummy individual data.}
\item{Location_data}{Dummy location data.}
}
\description{
Create dummy pipeline output (\code{dummy_data}) to be tested in \code{\link{quality_check}}.
In each data frame, rows are specifically created to test single checks from \code{\link{quality_check}}.
See a detailed description of which rows correspond to which checks in \code{\link{dummy_data}}.
}
