% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BAN.R
\name{create_individual_BAN}
\alias{create_individual_BAN}
\title{Create individual data table for Bandon Valley.}
\usage{
create_individual_BAN(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Primary data from Bandon Valley.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Bandon Valley.
}
