% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{create_individual_CHO}
\alias{create_individual_CHO}
\title{Create individual data table for Choupal, Portugal.}
\usage{
create_individual_CHO(data, Capture_data, protocol_version)
}
\arguments{
\item{data}{Data frame. Primary data from Choupal.}

\item{Capture_data}{Data frame. Output from \code{\link{create_capture_CHO}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Choupal,
Portugal.
}
