% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{create_individual_CHO}
\alias{create_individual_CHO}
\title{Create individual data table for Choupal, Portugal.}
\usage{
create_individual_CHO(data, Capture_data, optional_variables = NULL)
}
\arguments{
\item{data}{Data frame. Primary data from Choupal.}

\item{Capture_data}{Data frame. Output from \code{\link{create_capture_CHO}}.}

\item{optional_variables}{A character vector of names of optional variables (generated by standard utility functions) to be included in the pipeline output.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Choupal,
Portugal.
}
