% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HAR.R
\name{create_individual_HAR}
\alias{create_individual_HAR}
\title{Create individual table for Harjavalta.}
\usage{
create_individual_HAR(Capture_data)
}
\arguments{
\item{Capture_data}{Output of \code{\link{create_capture_HAR}}.}
}
\value{
A data frame.
}
\description{
Create full individual data table in standard format for data from Harjavalta, Finland.
}
