% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{create_individual_MON}
\alias{create_individual_MON}
\title{Create individual data table for Montpellier}
\usage{
create_individual_MON(capture_data)
}
\arguments{
\item{capture_data}{Capture data generated by \code{\link{create_capture_MON}}.}
}
\value{
A data frame with Individual data
}
\description{
Create individual data table for Montpellier
}
