% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_individual_PFN}
\alias{create_individual_PFN}
\title{Create individual data table for PiedFlyNet populations}
\usage{
create_individual_PFN(
  Capture_data,
  species_filter,
  pop_filter,
  protocol_version
)
}
\arguments{
\item{Capture_data}{Capture data table generated by `create_capture_PFN`}

\item{species_filter}{Species six-letter codes from the standard protocol. Used to filter the data.}

\item{pop_filter}{Population three-letter codes from the standard protocol. Used to filter the data.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame with Individual data
}
\description{
Create individual data table for PiedFlyNet populations
}
