% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SSQ.R
\name{create_individual_SSQ}
\alias{create_individual_SSQ}
\title{Create individual data table for Santo Stefano Quisquina, Italy.}
\usage{
create_individual_SSQ(data, Capture_data, Brood_data, protocol_version)
}
\arguments{
\item{data}{Data frame. Primary data from Santo Stefano Quisquina.}

\item{Capture_data}{Data frame. Generate by \code{\link{create_capture_SSQ}}.}

\item{Brood_data}{Data frame. Generate by \code{\link{create_brood_SSQ}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for data from Santo Stefano
Quisquina, Italy
}
