% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VEL.R
\name{create_individual_VEL}
\alias{create_individual_VEL}
\title{Create individual data table for Velky Kosir.}
\usage{
create_individual_VEL(Capture_data)
}
\arguments{
\item{Capture_data}{Data frame. Combined data from
\code{\link{create_capture_VEL_FICALB}} and
\code{\link{create_capture_VEL_TIT}}.}
}
\value{
A data frame.
}
\description{
Create individual data table in standard format for both flycatcher and tit
data from Velky Kosir.
}
