% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WYT.R
\name{create_individual_WYT}
\alias{create_individual_WYT}
\title{Create individual data table for Wytham Woods}
\usage{
create_individual_WYT(Capture_data, protocol_version)
}
\arguments{
\item{Capture_data}{Capture data generated by \code{\link{create_capture_WYT}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame with Individual data
}
\description{
Create individual data table for Wytham Woods
}
