% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_AMM.R
\name{create_location_AMM}
\alias{create_location_AMM}
\title{Create location data table for Ammersee, Germany.}
\usage{
create_location_AMM(Capture_data, dir)
}
\arguments{
\item{Capture_data}{Data frame. Output from \code{\link{create_capture_AMM}}.}

\item{dir}{Path to directory containing the relevant table exports from the AMM Access database.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from Ammersee, Germany.
}
