% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HAR.R
\name{create_location_HAR}
\alias{create_location_HAR}
\title{Create location table for Harjavalta.}
\usage{
create_location_HAR(db)
}
\arguments{
\item{db}{Location of primary data from Harjavalta.}
}
\value{
A data frame.
}
\description{
Create full location data table in standard format for data from Harjavalta, Finland.
}
