% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KEV.R
\name{create_location_KEV}
\alias{create_location_KEV}
\title{Create location table for Kevo, Finland.}
\usage{
create_location_KEV(db, protocol_version)
}
\arguments{
\item{db}{Location of primary data from Kevo.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create full location data table in standard format for data from Kevo, Finland.
}
