% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{create_location_MON}
\alias{create_location_MON}
\title{Create location data table for Montpellier}
\usage{
create_location_MON(db, Capture_data, Brood_data, protocol_version)
}
\arguments{
\item{db}{Location of primary data from Montpellier.}

\item{Capture_data}{Capture data generated by \code{\link{create_capture_MON}}.}

\item{Brood_data}{Capture data generated by \code{\link{create_brood_MON}}.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame with Individual data
}
\description{
Create location data table for Montpellier
}
