% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_NIOO.R
\name{create_location_NIOO}
\alias{create_location_NIOO}
\title{Create location data table for NIOO pipeline.}
\usage{
create_location_NIOO(
  dir,
  location_data,
  species_filter,
  pop_filter,
  protocol_version
)
}
\arguments{
\item{dir}{Path to directory containing the relevant table exports from the NIOO Access database.}

\item{location_data}{Data frame with location codes and corresponding PopID.}

\item{species_filter}{Species six letter codes from the standard protocol.
Used to filter the data.}

\item{pop_filter}{Population three letter codes from the standard protocol.
Used to filter the data.}

\item{protocol_version}{Character string. The version of the standard protocol on which this pipeline is based.}
}
\value{
A data frame.
}
\description{
Create location data table in standard format for data from NIOO.
}
