% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{create_location_PFN}
\alias{create_location_PFN}
\title{Create location data table for EastDartmoor.}
\usage{
create_location_PFN(Brood_data)
}
\arguments{
\item{Brood_data}{Brood data table generated by create_brood_PFN}
}
\value{
A data frame with Location data
}
\description{
Create location data table for EastDartmoor.
}
