% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_WYT.R
\name{create_location_WYT}
\alias{create_location_WYT}
\title{Create location data table for Wytham Woods}
\usage{
create_location_WYT(Brood_data)
}
\arguments{
\item{Brood_data}{Brood data generated by \code{\link{create_brood_WYT}}.}
}
\value{
A data frame with Location data
}
\description{
Create location data table for Wytham Woods
}
