% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_values.R
\name{create_reference_values}
\alias{create_reference_values}
\title{Create biological reference values}
\format{An R data object containing three listst of reference values.
\describe{
  \item{brood_ref_values_list}{Reference values related to brood data.}
  \item{cap_adult_ref_values_list}{Reference values related to capture data of adults.}
  \item{cap_chick_ref_values_list}{Reference values related to capture data of chicks.}
}}
\usage{
create_reference_values(db = utils::choose.dir())
}
\arguments{
\item{db}{Directory to save reference values.}
}
\description{
Manually add & edit reference values in the code and run function to
create the necessary files to be used in \code{\link{quality_check}}.
}
\details{
For each species-specific variable a distinction is made between
improbable values and impossible values. For each, provide a minimum
and a maximum. Thus, in total, each species-specific variable
corresponds to 4 reference values.
}
