% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_reference_values.R
\name{create_reference_values}
\alias{create_reference_values}
\title{Create biological reference values}
\usage{
create_reference_values(db = utils::choose.dir())
}
\arguments{
\item{db}{Directory to save reference values.}
}
\value{
An R data object (reference_values.rda)
  with three lists of reference values:
  brood data, adult capture data and chick capture data.
}
\description{
Manually add & edit reference values in the code and run function to
create the necessary files to be used in \code{\link{quality_check}}.
}
\details{
For each species-specific variable a distinction is made between
improbable values and impossible values. For each, provide a minimum
and a maximum. Thus, in total, each species-specific variable
corresponds to 4 reference values.
}
