% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{dummy_data}
\alias{dummy_data}
\title{Quality check dummy data}
\format{
A list of 4 data frames:
\itemize{
   \item{Brood_data}
   \item{Capture_data}
   \item{Individual_data}
   \item{Location_data}
}
}
\description{
A dummy pipeline output specifically created to test single checks from \code{\link{quality_check}}. A CheckID column is added to each dataframe to mark which rows serve to test each check. See a detailed description of which rows correspond to which checks in 'Details'.
}
\details{
\strong{Brood data}:
\itemize{
  \item Rows 1-4 are records that test the functionality of check 'B2: Comparing clutch and brood sizes' (see \code{\link{compare_clutch_brood}}): a non-manipulated brood with larger clutch size than brood size, a non-manipulated brood with larger brood size than clutch size, a manipulated brood with larger clutch size than brood size, and a manipulated brood with larger brood size than clutch size.
  \item Rows 5-8 are records that test the functionality of check 'B3: Comparing brood sizes and fledgling numbers' (see \code{\link{compare_brood_fledglings}}): a non-manipulated brood with larger brood size than fledgling number, a non-manipulated brood with larger fledgling number than brood size, a manipulated brood with larger brood size than fledgling number, and a manipulated brood with larger fledgling number than brood size.
  \item Rows 9-10 are records that test the functionality of check 'B4: Comparing laying and hatching dates' (see \code{\link{compare_laying_hatching}}): a brood with an earlier laying than hatching date, and a brood with an earlier hatching than a laying date.
  \item Rows 11-12 are records that test the functionality of check 'B5: Comparing hatching and fledging dates' (see \code{\link{compare_hatching_fledging}}): a brood with an earlier hatching than fledging date, and a brood with an earlier fledging than a hatching date.
  \item Rows 13-629 are records that test the functionality of checks 'B6a-d: Checking brood variable values against reference values' (see \code{\link{check_values_brood}}): broods with probable, unlikely and impossible values in ClutchSize_observed (B6a), BroodSize_observed (B6b), NumberFledged_observed (B6c) and LayDate_observed (B6d).
  \item Rows 630-632 are records that test the functionality of check 'B7: Comparing brood size with number of chicks captured' (see \code{\link{compare_broodsize_chicknumber}}): a brood with a brood size equal to the number of chicks in Capture_data, a brood with a brood size larger than the number of chicks in Capture_data, and a brood with a brood size smaller than the number of chicks in Capture_data.
  \item Rows 633-635 are records that test the functionality of check 'B8: Checking that brood IDs are unique' (see \code{\link{check_unique_BroodID}}): a brood with a unique BroodID, and two broods with the same BroodID.
  \item Rows 636-641 are records that test the functionality of check 'B9: Checking that clutch type order is correct' (see \code{\link{check_clutch_type_order}}): two broods by one breeding pair in one season in the correct order of creation, two broods by one breeding pair in one season in the correct order of creation but "first" brood missing, two broods by one breeding pair in one season in the incorrect order of creation.
  \item Rows 642-643 are records that test the functionality of check 'B10: Comparing parent species' (see \code{\link{compare_species_parents}}): a brood with parents of the same species, and a brood with parents of different species.
  \item Rows 642-643 are records that are needed to test the functionality of check 'C4: Checking that adults caught on nest are listed are the parents (see \code{\link{check_adult_parent_nest}}): a brood with the same individuals as parents as the individuals caught on that nest, and a brood with individuals different from the individuals that are caught on this nest.
  \item Rows 646-647 are records that test the functionality of check 'B11: Comparing species of brood and of parents' (see \code{\link{compare_species_brood_parents}}): a brood of the same species as both of the parents, and a brood of a different species than either of the parents.
  \item Rows 648-649 are records that test the functionality of check 'B12: Comparing species of brood and of chicks' (see \code{\link{compare_species_brood_chicks}}): a brood of the same species as the chicks in that brood, and a brood of a different species than either of the chicks in that brood.
}

\strong{Capture data}:
\itemize{
  \item Rows 1-4 are records that are needed to test the functionality of check 'B7: Comparing brood size with number of chicks captured' (see \code{\link{compare_broodsize_chicknumber}}): chicks in equal, smaller or larger number than recorded as brood size in Brood_data.
  \item Rows 5-8 are records that are needed test the functionality of check 'B10: Comparing parent species' (see \code{\link{compare_species_parents}}): parents of the same or different species than recorded in Brood_data.
  \item Rows 9-986 are records that test the functionality of checks 'C2a-b: Checking capture variable values against reference values' (see \code{\link{check_values_capture}}): adults and chicks with probable, unlikely and impossible values in Mass (C2a) and Tarsus (C2b).
  \item Rows 987-988 are records that test the functionality of check 'C3: Checking chick age values' (see \code{\link{check_chick_age}}): a chick with a correct chick age, and a chick with an incorrect chick age.
  \item Rows 989-993 are records that are needed to test the functionality of check 'I2: Checking unique individual IDs' (see \code{\link{check_unique_IndvID}}): individuals with IDs that are unique, duplicated among populations and duplicated within populations.
  \item Rows 994-995 are records that are needed to test the functionality of check 'I3: Checking that chicks have BroodIDs' (see \code{\link{check_BroodID_chicks}}): a chick with a BroodID, and a chick without a BroodID.
  \item Rows 996-997 are records that are needed to test the functionality of check 'I4: Checking that individuals have no conflicting sex' (see \code{\link{check_conflicting_sex}}): an individual with certain sex, and an individual with uncertain sex (i.e. recorded as both male and female).
  \item Rows 998-999 are records that are needed to test the functionality of check 'I5:Checking that individuals have no conflicting species' (see \code{\link{check_conflicting_species}}): an individual with certain species, and an individual with uncertain species (i.e. recorded as two different species).
  \item Row 1000 is a record that is needed to test the functionality of check 'I6: Checking that individuals in Individual_data also appear in Capture_data' (see \code{\link{check_individuals_captures}}): an individual in Individual_data that is also recorded in Capture_data.
  \item Rows 1001-1003 are records that test the functionality of check 'C4: Checking that adults caught on nest are listed are the parents' (see \code{\link{check_adult_parent_nest}}): a parent caught on a nest and marked as the parent of that brood, a parent caught in a mistnet, and a parent caught on a nest but not marked as the parent of that brood.
}

\strong{Individual data}:
\itemize{
  \item Rows 1-4 are records that are needed to test the functionality of check 'B7: Comparing brood size with number of chicks captured' (see \code{\link{compare_broodsize_chicknumber}}): chicks in equal, smaller or larger number than recorded as brood size in Brood_data.
  \item Rows 5-8 are records that are needed test the functionality of check 'B10: Comparing parent species' (see \code{\link{compare_species_parents}}): parents of the same or different species than recorded in Brood_data.
  \item Rows 9-13 are records that test the functionality of check 'I2: Checking unique individual IDs' (see \code{\link{check_unique_IndvID}}): individuals with IDs that are unique, duplicated among populations and duplicated within populations.
  \item Rows 14-15 are records that test the functionality of check 'I3: Checking that chicks have BroodIDs' (see \code{\link{check_BroodID_chicks}}): a chick with a BroodID, and a chick without a BroodID.
  \item Rows 16-17 are records that test the functionality of check 'I4: Checking that individuals have no conflicting sex' (see \code{\link{check_conflicting_sex}}): an individual with certain sex, and an individual with uncertain sex (i.e. recorded as both male and female).
  \item Rows 18-19 are records that test the functionality of check 'I5:Checking that individuals have no conflicting species' (see \code{\link{check_conflicting_species}}): an individual with certain species, and an individual with uncertain species (i.e. recorded as two different species).
  \item Rows 20-21 are records that test the functionality of check 'I6: Checking that individuals in Individual_data also appear in Capture_data' (see \code{\link{check_individuals_captures}}): an individual that is also recorded in Capture_data, and an individual that is missing from Capture_data.
  \item Rows 22-25 are records that are needed to test the functionality of check 'B11: Comparing species of brood and of parents' (see \code{\link{compare_species_brood_parents}}): parents of the same species as their brood, and parents of different species than their brood.
  \item Rows 26-29 are records that are needed to test the functionality of check 'B12: Comparing species of brood and of chicks' (see \code{\link{compare_species_brood_chicks}}): chicks of the same species as the brood they are in, and chicks of different species than the brood they are in.
}

\strong{Location data}:
\itemize{
  \item Row 1 is a record that is needed to test the functionality of check 'I3: Checking that chicks have BroodIDs' (see \code{\link{check_BroodID_chicks}}): location information of a nestbox.
  \item Rows 2-4 are records that are needed to test the functionality of check 'C4: Checking that adults caught on nest are listed are the parents' (see \code{\link{check_adult_parent_nest}}): location information of two nestboxes and a mist net.
}
}
