% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_access_db.R
\name{export_access_db}
\alias{export_access_db}
\title{Export selected tables from Access database to csv files through Jackcess}
\usage{
export_access_db(
  dsn,
  table,
  output_dir,
  header = TRUE,
  delim = ",",
  quote = "\\"",
  filter = rjackcess::SimpleExportFilter()$INSTANCE
)
}
\arguments{
\item{dsn}{Character. File path to MS Access database source name.}

\item{table}{Character. Names of tables to export.}

\item{output_dir}{Character. File path to directory where the new files will be created.}

\item{header}{Logical. Include the column names in the header? Default: \code{TRUE}.}

\item{delim}{Character. The column delimiter in the output file. Default: ",".}

\item{quote}{Character. The quote character in the output file. Default: '"'.}

\item{filter}{a Jackcess ExportFilter object to filter columns or rows in the export. Default: \code{rjackcess::SimpleExportFilter()$INSTANCE}, which is the full, unfiltered instance (table).}
}
\value{
csv files of selected tables from Access database
}
\description{
Connect to MS Access database via Jackcess (a Java library) to export a selected tables as a  csv files.
}
\details{
Jackcess is a Java library for reading from and writing to MS Access databases.
Previously, we used ODBC (Open Database Connectivity) to connect to MS Access databases,
but this does not work on MacOS and Linux as straightforwardly as it does on Windows.
}
