% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_BAN.R
\name{format_BAN}
\alias{format_BAN}
\title{Construct standard format for data from Bandon Valley, Ireland.}
\usage{
format_BAN(db = utils::choose.dir(), path = ".", species = NULL,
  pop = NULL, output_type = "R")
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the nest box population in Bandon
Valley, Ireland, administered by John Quinn.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{Species}: There are records where species is uncertain (e.g. listed as
'GRETI?'). This uncertainty is ignored. We assume that the suggested species
is correct. We include blue tit, great tit, and coal tit. No other species are
recorded.

\strong{ClutchType_observed}: Clutch type is only listed as 'first' and
'second'. There is no distinction between 'second' and 'replacement'. We
categorise all these as 'second'. Possible distinction between 'second' and
'replacement' can be made with \strong{ClutchType_calculated}. There are a
small number of cases where nest attempt number is uncertain (e.g.
`2(MAYBE)`). This uncertainty is ignored.

\strong{LayDate, HatchDate, NumberFledged}: There are some cases where
values are given with uncertainty (e.g. 97+, 95?). We don't know how much
uncertainty is involved here, it is currently ignored, but we need to
discuss this with data owner.

\strong{ClutchSize}: Cases where clutch size is uncertain (e.g. nests were
predated before completion) are treated as NA because clutch size is unknown.

\strong{LocationID}: Box numbers are not unique, they are repeated between
plots. To generate LocationID, we therefore use Plot_BoxNumber.

\strong{BroodID}: Unique BroodID is currently made with
Year_Plot_BoxNumber_Day_Month.

\strong{BroodIDLaid/Fledged}: Currently, we have no information about the
brood where each individual was laid. Therefore, these are currently kept
blank.

\strong{Age_observed}: There is no recorded capture age. This is left as NA.

\strong{AvgEggMass}: Currently we only include records where the day of egg
weighing is <= LayDate + ClutchSize. This should be an estimate of the date
that incubation began. Once incubation starts, egg weight is not easily
comparable because it changes with embryonic development.

\strong{AvgChickMass/AvgTarsus}: Individual capture data is not included in
the data currently provided. These values are therefore left blank.

\strong{StartSeason}: Assume all boxes were placed in the first year of the study.
}
