% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{format_CHO}
\alias{format_CHO}
\title{Construct standard format for data from Choupal, Portugal.}
\usage{
format_CHO(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the great tit population in
Choupal, Portugal, administered by the University of Coimbra.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{NumberFledged}: This population has no estimation of actual fledgling
numbers. The last time nests are counted is 14 days post hatching. We use this
as an estimation of fledgling numbers. This also affects
\emph{ClutchType_calculated} as we use the estimate of fledgling numbers to
distinguish second/replacement clutches.

\strong{Age_observed}: Translation of age records: \itemize{ \item Any
individual caught as a chick was assumed to have a EURING code of 1: 'Pullus:
nestling or chick, unable to fly freely, still able to be caught by hand.'
\item Any individual listed as 'first year' was given a EURING code of 5: a
bird hatched last calendar year and now in its second calendar year. \item Any
individual listed as 'adult' was given a EURING code of 6: full-grown bird
hatched before last calendar year; year of hatching otherwise unknown.}

\strong{ClutchType_observed}: In the raw data, there is no distinction between
'second' and 'replacement' clutches. Any clutch recorded as '2nd' is assumed
to be a 'second' clutch under our definitions. 'ClutchType_calculated' may
give a more appropriate estimate of clutch type for this data.

\strong{ClutchSize, BroodSize, NumberFledged}: We currently only use records
of clutch, brood, and fledgling numbers that are recorded explicitly in the
data. This means that there are some nests where chicks have capture records,
but the \emph{Brood data} table will not give any value of NumberFledged.
(e.g. see BroodID 2004_NA). These capture records should be included, but we
need to determine the amount of potential uncertainty around these records.

\strong{LocationID}: For individuals captures in mist nets (specified by
trapping method column), a single LocationID "MN1" is used.

\strong{StartSeason}: Assume all boxes were placed in the first year of the study.
}
