% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_CHO.R
\name{format_CHO}
\alias{format_CHO}
\title{Construct standard format for data from Choupal, Portugal.}
\usage{
format_CHO(
  db = choose_directory(),
  species = NULL,
  site = NULL,
  optional_variables = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{site}{The three-letter code of the study site as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. For data owners with multiple study sites (e.g., NIOO, UAN) where a single
pipeline is used for many study sites, this argument is used to extract data from
individual study sites. For other pipelines that contain only one study site
this argument can be ignored.}

\item{optional_variables}{Optional variables (generated by standard utility functions) to be included in the pipeline output. Either a character vector of variable names or 'all' (if all optional variables should be included).}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 6 .csv files or 6 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the great tit population in
Choupal, Portugal, administered by the University of Coimbra.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{here}.

\strong{observedNumberFledged}: This population has no estimation of actual fledgling
numbers. The last time nests are counted is 14 days post hatching. We use this
as an estimation of fledgling numbers. This also affects
\emph{calculatedClutchType} as we use the estimate of fledgling numbers to
distinguish second/replacement clutches.

\strong{Age_observed}: Translation of age records: \itemize{ \item Any
individual caught as a chick was assumed to have a EURING code of 1: 'Pullus:
nestling or chick, unable to fly freely, still able to be caught by hand.'
\item Any individual listed as 'first year' was given a EURING code of 5: a
bird hatched last calendar year and now in its second calendar year. \item Any
individual listed as 'adult' was given a EURING code of 6: full-grown bird
hatched before last calendar year; year of hatching otherwise unknown.}

\strong{observedClutchType}: In the raw data, there is no distinction between
'second' and 'replacement' clutches. Any clutch recorded as '2nd' is assumed
to be a 'second' clutch under our definitions. 'calculatedClutchType' may
give a more appropriate estimate of clutch type for this data.

\strong{observedClutchSize, observedBroodSize, observedNumberFledged}: We currently only use records
of clutch, brood, and fledgling numbers that are recorded explicitly in the
data. This means that there are some nests where chicks have capture records,
but the \emph{Brood data} table will not give any value of observedNumberFledged
(e.g., see broodID 2004_NA). These capture records should be included, but we
need to determine the amount of potential uncertainty around these records.

\strong{individualID}: Individuals marked as "no ring", "not ringed" or "branco" are removed from Capture_data and Individual_data. Check with data owner on how to handle these.

\strong{captureAlive, releaseAlive}: Assume all individuals were alive when captured and released.

\strong{captureRingNumber}: First captures of all individuals are assumed to be ringing events, and thus captureRingNumber is set to NA.

\strong{locationID}: For individuals captured in mist nets (specified by
trapping method column), a single locationID "MN1" is used.

\strong{startYear}: Assume all boxes were placed in the first year of the study.

\strong{habitatID}: Assume that habitat type is 1.4: Forest - Temperate. Check with data owner.
}
