% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_GRO.R
\name{format_GRO}
\alias{format_GRO}
\title{Construct standard format for data from Grobla, Poland}
\usage{
format_GRO(
  db = choose_directory(),
  path = ".",
  species = NULL,
  pop = NULL,
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{path}{Location where output csv files will be saved.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the nest box population in Grobla, Poland, administered by Mariusz Cichoń
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

#'\strong{CaptureDate}: Adults are typically captured two weeks into the nestling period.
If hatch date is not known for the nest, CaptureDate is set to two weeks after May 14th, which is the average hatch date.
}
