% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_Harjavalta.R
\name{format_Harjavalta}
\alias{format_Harjavalta}
\title{Construct standard summary for data from Harjavalta, Finland.}
\usage{
format_Harjavalta(db = NULL, Species = NULL, path = ".")
}
\arguments{
\item{db}{Location of database file.}

\item{Species}{A numeric vector. Which species should be included (EUring codes)? If blank will return all major species (see details below).}

\item{path}{Location where output csv files will be saved.}
}
\value{
Generates 5 .csv files with data in a standard format.
}
\description{
Construct standard summary for data from Harjavalta, Finland.
}
