% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_HOC.R
\name{format_HOC}
\alias{format_HOC}
\title{Construct standard format for data from Hochstadt, Germany.}
\usage{
format_HOC(db = utils::choose.dir(), species = NULL, pop = NULL,
  path = ".", output_type = "R")
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the hole nesting bird
populations in Hochstadt, Germany administered by Max Plank Institute
for Ornithology, Seewiesen (Michaela Hau).
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard protocl please see
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{here}.
}
