% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_KEV.R
\name{format_KEV}
\alias{format_KEV}
\title{Construct standard format for data from Kevo, Finland.}
\usage{
format_KEV(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R",
  return_errors = FALSE
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the hole nesting bird population
in Kevo, Finland, administered by the University of Turku.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard protocl please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{Species:} We only include records for great tit, pied flycatcher,
Siberan tit (Parus cinctus), and Common redstart (Phoenicurus phoenicurus).
All these species have at least 350 capture records. There are also records of
blue tits, spotted flycatcher (Muscicapa striata), willow tit (Parus
montanus), and Eurasian wryneck (Jynx torquilla); however, these species are
captured at most 13 times over >30 years. Therefore, they are excluded.

\strong{Age}: Chick age is listed as: PP (nestling), PM (fledgling), FL(unknown),
1, +1, 2, +2. Translation from field notebook would translate PP/PM into
EURING 1. PP are chicks in the nest. PM are chick left the nest but caught by
hand (i.e. still not able to fly). 1 = 3 (i.e. fully grown in first year).
2 = 5 (i.e. fully grown and known to be born last year).
+1/+2 are 4/6 respectively, at least 1/2 years old.
FL is unknown. We can't even attribute it to chick or adult.
For this we use EURING 2 (able to fly freely but otherwise unknown).

\strong{LayDateError & HatchDateError}: Accuracy of laying and hatch date
are given as categories: 0-1; 1-2; 2-3; 'inaccurate'. Where error is a range,
the more conservative error is used (i.e. 0-1 is recorded as 1).
Cases listed as 'inaccurate' have an error of at least a week.
Therefore, these ones are given Lay/HatchDateError of 7. Dates in these
cases are highly inaccurate and shouldn't be considered for any phenology
analysis.

\strong{Capture data:} As with Harjavalta, linking nestling and adult capture
data can be difficult. There are eight different scenarios we need to
consider. Each of these is described below, with our solution: \itemize{ \item
#1. Individual captured as adult (Age is NA or not PM/PP) (e.g. HL-137181).

In this case, we assume that there is only information in Capture_data, and we
do not search for any information in Nestling_data.

\item #2. Record of chick (Age is PM/PP) with no information in the last ring
number column AND no corresponding record in Nestling_data (i.e. no record
where BroodID and last 2 digits of the ring number were the same) (e.g.
HL-568201).

In this case, we assume that the information in Capture_data is the correct
capture information.

\item #3.	Record of chick (Age is PM/PP) with no information in the last ring
number column, with corresponding record in Nestling_data, BUT where the
capture dates in Nestling_data and Capture_data do not match (e.g. V-478577)

In this case, we assume that the individual was captured multiple times and
the records in Nestling_data and Capture_data are both true capture records.
All information from both tables is kept.

\item #4.	Record of chick (Age is PM/PP) with no information in the last ring
number column, corresponding record in Nestling_data that was taken at exactly
the same capture date (e.g. JL-112052)

Where there are measurements (e.g. mass, tarsus) in both Nestling_data and
Capture_data, we give precedence to the data from Capture_data. We only use
data from Nestling_data when there is no data in Capture_data

\item #5.	Record of chick (Age is PM/PP) with information in the last ring
number column which matches data in Nestling_data (e.g. HL-26103 - 26107)

We assume that there is a continuous ring series used (e.g. 26103, 26104,
26105, 26106, 26107) and link all corresponding records from Nestling_data
where the BroodID and last 2 digits of the ring number match.

\item #6.	Record of chick (Age is PM/PP) with information in the last ring
number column but no matches to data in Nestling_data (e.g. HL-364911 -
364916).

We assume the individuals were captured and ringed, but there was no data
collected on them (e.g. Mass). We give them an 'empty' record in Capture_data
(e.g. only IndvID, CaptureDate, CaptureTime).

\item #7. Record in Nestling_data that has no ring number information (e.g. A,
B, C, D) (e.g. Brood 2011_0003_1)

We assume these were chicks captured before they were old enough to be ringed.
We keep these as a legitimate capture record as they can tell us something
about chick growth.

\item #8.	Record in Nestling_data that has ring number information, but there
is no corresponding record of chicks with those rings being captured from that
brood in Capture_data (e.g. Brood 1985_2081_1 with individuals 71, 72, 73,
74).

We cannot determine the identity of the captured individual without a matching
record in Capture_data because we only know the last 2 digits of the ring
number. Currently, we include them as captures with IndvID = NA. These can
still be used to estimate AvgChickMass, AvgTarsus for a given BroodID even
though the individual cannot be identified; however, we suspect these are
mistakes in the primary data.}
}
