% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{format_MAY}
\alias{format_MAY}
\title{Construct standard format for data from Mayachino, Russia.}
\usage{
format_MAY(
  db = choose_directory(),
  species = NULL,
  site = NULL,
  optional_variables = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{site}{The three-letter code of the study site as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. For data owners with multiple study sites (e.g., NIOO, UAN) where a single
pipeline is used for many study sites, this argument is used to extract data from
individual study sites. For other pipelines that contain only one study site
this argument can be ignored.}

\item{optional_variables}{Optional variables (generated by standard utility functions) to be included in the pipeline output. Either a character vector of variable names or 'all' (if all optional variables should be included).}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 6 .csv files or 6 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the study site at
Mayachino, Russia, administered by the Institute of Biology at the Karelian Research Centre.
}
\details{
This section provides details on data management choices that are unique to
these data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{here}.

\strong{plotID}: The "line of nest boxes" are interpreted as distinctive plots. Check with data owner.

\strong{speciesID}: No species identification in primary data. Check with data owner.

\strong{Lay dates & hatch dates}: Some lay dates and hatch dates contain special characters (e.g., "<", "?", "()"), which are ignored. Some lay dates and hatch dates in primary data are a range of values (e.g., "31-34"). These are converted into the minimum (minimumLayDate, minimumHatchDate) and maximum dates (maximumLayDate, maximumHatchDate); observed dates (observedLayDate, observedHatchDate) are given by the average of the range (rounded down; i.e. "32" in the example). Check with data owner.

\strong{Clutch size, brood size, fledgling number}: Some clutch sizes, brood sizes, and fledgling numbers contain special characters (e.g., ?", "()"), which are ignored. Some clutch sizes, brood sizes, and fledgling numbers are written as an arithmetic expression (e.g., "7+4"), and interpreted as such (i.e., the observed value equals 11 in the example).

\strong{captureAlive, releaseAlive}: All individuals are assumed to be captured and released alive.

\strong{captureRingNumber, releaseRingNumber}: First captures of all individuals are assumed to be ringing events, and thus captureRingNumber is set to NA.

Only when individual IDs start with two letters, followed by 5 or 6 digits, the ID is considered a ring number and stored in captureRingNumbber and/or releaseRingNumber. When, for example, the letters are missing, the ring number is considered incomplete and captureRingNumber and releaseRingNumber are set to NA.

\strong{startYear}: Assume all boxes were placed in the first year of the study.

\strong{habitatID}: Assume that habitat type is 1.1: Forest - Boreal Check with data owner.
}
