% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{format_MON}
\alias{format_MON}
\title{Construct standard format for data from Montpellier, France.}
\usage{
format_MON(db = utils::choose.dir(), species = NULL, pop = NULL,
  path = ".", debug = FALSE, output_type = "csv", verbose = FALSE)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This includes: a) Histogram of
continuous variables with mean/SD b) unique values of all categorical
variables.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the hole nesting bird population
in Montpellier, France, administered by CNRS Montpellier (Anne Charmantier and
colleagues).
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard protocl please see
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{IndvID:} In Capture_data, any individuals given unidentified numbers
(i.e. containing 'no-ident') are treated as NA. These records are still kept
as they contain information about broods (e.g. AvgChickMass) but are not
integrated into the individual data table.

\strong{Sex:} We ignore uncertainty in sex (e.g. category 4 = Male probable
but not certain)

\strong{LocationID:} For birds caught in boxes. Location is
Plot_NextboxNumber.

\strong{BroodID:} Broods are BreedingSeason_LocationID_ClutchTypeObserved =
BreedingSeason_Plot_NestboxNumber_ClutchTypeObserved.

\strong{ClutchTypeObserved:} No clutch type recorded, only calculated clutch
type is given.

\strong{Tarsus:} Left and right tarsus are measured. Right generally has more
data, so we use this as our measure of tarsus length. Currently, we assume
everything is in Svensson's alternative, but there is supposedly some change
from before 1989. Need to ask Anne about this.

\strong{Age:} We translate observed age codes into EURING codes as follows:
\itemize{

\item P0 (Poussin bagué au nichoir ou en cavité naturelle/chick banded in box
or natural cavity): We give EURING code 1: Nestling or chick unable to fly
freely.

\item J0 (Juvénile (oiseau de l’année) capturé entre le moment où il s’est
envolé du nichoir et le 31 décembre) (Juvenile (of this year) captured between
fledging and Dec 31st of that year) We give EURING code 3: First-year:
full-grown bird hatched in the breeding season of this calendar year.

\item PN (where N > 0) Individuals caught as chick and caught again N seasons
later We give EURING code 3 + 2*N (i.e. of known age because it was caught
when it could be accurately aged)

\item J1 (Juvénile (oiseau de l’année dernière) capturé après le 1er janvier
et avant la mue post reproduction (juillet-août)) (Juvenile (of last year)
capture between Jan 1st and July-Aug) We give EURING code 5: 2nd year: a bird
hatched last calendar year and now in its second calendar year.

\item JN (where N > 1) Individuals caught as juvenile and caught again N
season later We give EURING code 3 + 2*N: Nestling or chick unable to fly
freely.

\item AN (where N >= 0) An adult not captured as a chick or juvenile. We give
EURING code 4 + 2*N (i.e. of unknown age because it was not caught when it
could be accurately aged).

\item IN (where N > 1): An adult not captured as a chick or juvenile. We give
EURING code 4 + 2*1-N (i.e. of unknown age because it was not caught when it
could be accurately aged). }

\strong{Age_calculated:} We strictly say that we canot accurately know the age
of an individual unless they were caught in their first year. Therefore,
although birds classified as J1 can be aged very accurately, we still treat
these as having age uncertain because they weren't caught as chicks.

\strong{BroodIDLaid/Fledged:} BroodIDLaid is the brood in which an individual
was listed in one of the chick columns. The BroodIDFledged is the same as BroodIDLaid
unless a destination brood is listed. In any cases where multiple broods are associated
with an individual (e.g. they are listed as a chick twice) we currently select the
first record and return a warning. We assume the first nest recorded is more likely to
be the 'true' nest until these are corrected.
}
