% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MON.R
\name{format_MON}
\alias{format_MON}
\title{Construct standard format for data from Montpellier, France.}
\usage{
format_MON(db = utils::choose.dir(), species = NULL, pop = NULL,
  path = ".", output_type = "csv", verbose = FALSE)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}

\item{verbose}{Should messages be printed during the pipeline?}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the hole nesting bird
populations in Southern France (Rouviere, Montpellier, Corsica etc.),
administered by CNRS Montpellier (Anne Charmantier and colleagues).
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard protocl please see
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{IndvID:} In Capture_data, any individuals given unidentified numbers
(i.e. containing 'no-ident') are given an IndvID = NA. These records are still kept
as they contain information about broods (e.g. AvgChickMass) but are not
integrated into the individual data table.

\strong{Sex:} We ignore uncertainty in sex (e.g. category 4 = Male probable
but not certain is just treated as M).

\strong{LocationID:} For birds caught in boxes. Location is
Plot_NextboxNumber.

\strong{BroodID:} Broods are BreedingSeason_LocationID_ClutchType_observed =
BreedingSeason_Plot_NestboxNumber_ClutchType_observed.

\strong{Tarsus:} Left and right tarsus are measured. Right tarsus is
used where available. Left tarsus measures are only used where right
is missing. All tarsus is measured with Svensson's Alternative method;
however, be aware that there was a change in the technique used in 1989.
Pre-1989, tarsus was measured using a compass and ruler. 1989+ used
standard callipers. This can affect tarsus measurements and should be
accounted for in any analysis.

\strong{Age:} We translate observed age codes into EURING codes as follows:
\itemize{

\item P0 (Poussin bagué au nichoir ou en cavité naturelle/chick banded in box
or natural cavity): We give EURING code 1: Nestling or chick unable to fly
freely. Any individual in the chick capture table is given Age 1.

\item J0 (Juvénile (oiseau de l’année) capturé entre le moment où il s’est
envolé du nichoir et le 31 décembre) (Juvenile (of this year) captured between
fledging and Dec 31st of that year) We give EURING code 3: First-year:
full-grown bird hatched in the breeding season of this calendar year.

\item PN (where N > 0) Individuals caught first as chick and caught again N seasons
later We give EURING code 3 + 2*N (i.e. of known age because it was caught
when it could be accurately aged)

\item J1 (Juvénile (oiseau de l’année dernière) capturé après le 1er janvier
et avant la mue post reproduction (juillet-août)) (Juvenile (of last year)
capture between Jan 1st and July-Aug) We give EURING code 5: 2nd year: a bird
hatched last calendar year and now in its second calendar year.

\item JN (where N > 1) Individuals caught as juvenile and caught again N
season later We give EURING code 3 + 2*N, as above.

\item AN (where N >= 0) An adult not captured as a chick or juvenile. We give
EURING code 4 + 2*N (i.e. of unknown age because it was not caught when it
could be accurately aged).

\item IN (where N > 1): An adult not captured as a chick or juvenile. We give
EURING code 4 + 2*1-N (i.e. of unknown age because it was not caught when it
could be accurately aged). }

\strong{Age_calculated:} We strictly say that we canot accurately know the age
of an individual unless they were caught in their first year. Therefore,
although birds classified as J1 can be aged very accurately, we still treat
these as having age uncertain because they weren't caught as chicks.

\strong{BroodIDLaid/Fledged:} BroodIDLaid is the brood in which an individual
was listed as a chick in the brood data table. The BroodIDFledged is the same as
BroodIDLaid unless origin and/or destination information is listed. In this case,
cross-fostering has occurred. BroodIDFledged is then the brood active at the destination
nest box at the time of capture. In any cases where multiple
broods are associated with an individual we currently select the first record and return a warning. We assume
the first nest recorded is more likely to be the 'true' nest until these are
corrected.

\strong{ReleasePlot/PopID:} Individuals transferred to aviaries are given ReleasePlot/PopID
'aviary'.

\strong{ExperimentID:} Currently, broods are given an ExperimentID TRUE or FALSE.
Experiments will be expanded to include exact experimental details at a later stage.

\strong{LocationID:} When individual is captured as chick or in nest the LocationID
is Plot_BoxNumber_NB. For non-chick captures, if there is no BoxNumber listed
we return LocationID NA. Even if it wasn't a capture in a nestbox, we have no
way of defining the location. If a BoxNumber is provided the LocationID is
Plot_BoxNumber_X where X is NB for nest box/winter roost captures and MN for
mist net or trap cage and caller captures.
}
