% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_NIOO.R
\name{format_NIOO}
\alias{format_NIOO}
\title{Construct standard summary for Hoge Veluwe population}
\usage{
format_NIOO(db = NULL, Species = c(14620, 14640), path = ".")
}
\arguments{
\item{db}{Location of database}
}
\description{
Construct standard summary for Hoge Veluwe population
}
