% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_NIOO.R
\name{format_NIOO}
\alias{format_NIOO}
\title{Construct standard format for NIOO data.}
\usage{
format_NIOO(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates 4 .csv files or R data frames in the SPI-Birds standard format.
}
\description{
A pipeline to produce the standard format for 8 hole-nesting bird study
populations at the Netherlands Institute of Ecology (NIOO-KNAW).
}
\details{
This section provides details on data management choices that are unique to
the NIOO database. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{Species}: By default the pipeline will include great tit \emph{Parus
major}; blue tit \emph{Cyanistes caeruleus}; pied flycatcher \emph{Ficedula
hypoleuca}; Eurasian nuthatch \emph{Sitta europaea}; coal tit \emph{Periparus
ater}; and tree sparrow \emph{Passer montanus}. Other minority species are
excluded.

\strong{Populations}: This pipeline extracts data for 8 populations managed by
NIOO-KNAW: Buunderkamp, Lichtenbeek, Westerheide, Hoge Veluwe, Warnsborn,
Vlieland, Oosterhout, and Liesbosch.

\strong{Sex}: We condense sex information to only include groups M, F, and C
(conflicting). Uncertainty in sex was ignored (e.g.
'male?' or 'female?').

\strong{Measurement error}: For BroodSize, NumberFledged, and LayDate a best
estimate is provided. Best estimate is halfway between the minimum and maximum
possible value. \emph{N.B.:} In cases where the best estimate is not an integer,
we provide the lower value (e.g. 2.5 is recorded as 2).
Error is provided in BroodSizeError,
NumberFledgedError, and LayDateError this is the absolute error (+/-) around the best
estimate.

\strong{CapturePlot, ReleasePlot, LocationID}: NIOO data gives CaptureLocation
and ReleaseLocation of each capture.
Each location is within a plot, specified by AreaID. We use this AreaID
information as the Capture/ReleasePlot
information.

\strong{Capture_data}: Capture_data has information on the accuracy of capture dates.
Values >1 are inaccurate and are ignored. A value of 0 is unknown. These are likely
mistakes in the primary data (i.e. the accuracy should always be known). For now,
we include captures with accuracy of both 0 and 1 in the final data.

\strong{AvgEggMass} Egg measurements are included in the NIOO database, but these are a bit more difficult to include
because they aren't associated with a given brood (they can be weighed before and after a cross fostering). For now,
we don't include this data, but we hope to in the future. Therefore, AvgEggMass is currently just NA.

\strong{ChickAge:} For every capture, we estimate the age of a chick as the difference between the hatch date
taken from BroodIDFledged (in Individual_data) and the CaptureDate. We include chick ages for all individuals
up until 30 days post hatching to accomodate possible late fledging.
}
