% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PEW.R
\name{format_PEW}
\alias{format_PEW}
\title{Construct standard format data from Peerdsbos West, Belgium (PEW)}
\usage{
format_PEW(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for bird study population
at the Peerdsbos West, Belgium, administered by Arne Iserbyt.
}
\examples{

}
