% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_PFN.R
\name{format_PFN}
\alias{format_PFN}
\title{Construct standard format for PiedFlyNet data}
\usage{
format_PFN(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.2.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the hole-nesting bird populations
monitored/administered by PiedFlyNet.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard protocl please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{Species}: By default, pipeline outputs will include great tit \emph{Parus
major}; blue tit \emph{Cyanistes caeruleus}; pied flycatcher \emph{Ficedula
hypoleuca}; Eurasian nuthatch \emph{Sitta europaea}; coal tit \emph{Periparus
ater}; marsh tit \emph{Poecile palustris}, and common redstart \emph{Phoenicurus phoenicurus}.
Other minority species are excluded.

\strong{Primary data sources}: Brood data is extracted from data in a
brood table/nest monitoring format ("Nest data"). Capture data is extracted from both
Nest data and Malcolm Burgess' version of the BTO ringing database for birds marked within
the PiedFlyNet ("IPMR data"). The format of this ringing data is IPMR, the 'old' BTO ringing database format, and Malcolm
has his own pipeline to convert the new BTO ringing database format to the old one.
Nest data is available for all populations, IPMR data is not.
Some information on captures is present in Nest and IPMR data, while some information is
exclusive to either data source. See the documentation of `create_capture_PFN()` for
a description of how data from Nest and IPMR data is consolidated.

\strong{Unidentified individuals}: Some individuals have not been given unique IDs. The pipleline
recognizes these as improper IDs as they do not follow the standard format of ring
numbers for this data (defined as a sequences of 1-9 letters followed by a sequence of 1-9 integer numbers).
In the Brood data, we assume unknown identity for individuals with improper IDs and treat ID as NA.
From Capture data and Individual data, these individuals are omitted entirely.

\strong{Re-ringing events}: Some individuals carry multiple rings (up to 3 in this data) throughout
their lives, and re-ringing events are recorded in IPMR data. To correctly link data to
individuals (not just ring numbers), the pipeline assigns a unique individual identifier defined as
the number of the first ring ever given to this individual.

\strong{Resightings}: The IPMR primary data contains some resightings of birds. At present, these
records are excluded from the outputs because the standard format for capture data cannot distinguish
between recaptures and resightings in its current state. Functionality do do this (i.e. via a "capture type" variable)
may be added to a future update of the standard format.

\strong{Age calculation}: The age of each bird is estimated based on its history of ringing and recaptures.
Since some birds that breed in one of the focal populations may have been ringed in another study sites (= immigrants), the
pipeline calculates age for birds in any one population based on capture records in all monitored populations,
and not just ringing/captures in locations within the specific study site.

\strong{LayDate_observed and HatchDate_observed}: Information is provided as date of first egg (DFE) and
date of hatching (DH). These are given as integer number, and represent days after a set
starting date (day 0). Day 0 is defined as 31. March.

\strong{ChickAge}: For every capture, we estimate the age of a chick as the difference between the hatch date
taken from `BroodIDFledged` (in `Individual_data`) and the `CaptureDate`.

#'\strong{Location information}: This is generated for all nestboxes (from Nest data)
and all additional capture locations (IPMR data). For East Dartmoor, detailed
information on location (`Latitude`, `Longitude`) and the period a nestbox has
been deployed (´StartSeason`, `EndSeason`) is available and incorporated. For
All other populations, this data is not available at present. Locations for other
populations therefore have no coordinate information, and `StartSeason` and
`EndSeason` are calculated as the first and last year in which birds were
observed to be breeding within a box.
}
