% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SSQ.R
\name{format_SSQ}
\alias{format_SSQ}
\title{Construct standard format for data from Santo Stefano Quisquina, Italy.}
\usage{
format_SSQ(
  db = utils::choose.dir(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates either 4 .csv files or 4 data frames in the standard format.
}
\description{
A pipeline to produce the standard format for the great and blue tit population
in Santo Stefano Quisquina, Sicly, Italy, administered by Camillo Cusimano
and Daniela Campobello.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{BroodID}: Unique BroodID is constructed using:
BreedingSeason_LocationID_LayDate (April days)

\strong{Species}: In the individual data, there are some cases where an
IndvID is associated with >1 species. I assume these are just typos and I
will just take the first species.

\strong{CaptureDate}: No exact capture date is currently given. For adults we
use the laying date of the nest as a proxy for capture date. Chicks were only
ever captured on the nest, we used laying date + clutch size + 15 days
incubation + 12 days. This is because chicks were ringed at 12 days old at
the latest.

\strong{Age_calculated}: All ringed chicks were assumed to be ringed at EURING code
1 (i.e. pre-fledging).

\strong{Individual_data}: There are cases where chicks from different nests are
given the same ring number. Unsure if this is the rings being reused or a
typo. Currently, I leave it as is and assume this is a typo that needs to be
fixed in the primary data.

\strong{StartSeason}: Some nest boxes were replaced over the course of
the study; however, these replacements were not explicitly recorded.
Therefore, we list all nestboxes as functioning for the full study period.
}
