% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_SSQ.R
\name{format_SSQ}
\alias{format_SSQ}
\title{Construct standard summary for data from Santo Stefano Quisquina, Italy.}
\usage{
format_SSQ(db = file.choose(), Species = NULL, path = ".",
  debug = FALSE)
}
\arguments{
\item{db}{Location of database file.}

\item{Species}{A numeric vector. Which species should be included (EUring
codes)? If blank will return all major species (see details below).}

\item{path}{Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This includes: a) Histogram of
continuous variables with mean/SD b) unique values of all categorical
variables.}
}
\value{
Generates 5 .csv files with data in a standard format.
}
\description{
A pipeline to produce a standard output for the great and blue tit population
in Santo Stefano Quisquina, Sicly, Italy, administered by Camillo Cusimano
and Daniela Campobello.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see XXXXX
PLACE HOLDER!

\strong{BroodID}: For now we make BroodID as a combination of: year of nest,
nestID, layingdate. This is necessary because there are no other ways to
distinguish a two broods laid in the same nest and same year.

\strong{Species}: In the individual data, there are some cases where an
IndvID is associated with >1 species. I assume these are just typos and I
will just take the first species.

\strong{CaptureDate}: No exact capture date is currently given. For adults we
use the laying date of the nest as a proxy for capture date. Chick were only
ever captured on the nest, we used laying date + clutch size + 15 days
incubation + 12 days. This is because chicks were ringed at 12 days old at
the latest.

\strong{Age_calc}: All ringed chicks were assumed to be ringed at EURING code
1 (i.e. pre-fledging). For adults where no age was provided, we assumed that
first observation was 6 (i.e. at least 2 years old)

\strong{Indv_data}: There are cases where chicks from different nests are
given the same ring number. Unsure if this is the rings being reused or a
typo. Currently, I leave it as is and assume this is a typo that needs to be
fixed in the primary data.

\strong{Nestbox StartYear}: Some nest boxes were replaced over the course of
the study; however, these replacements were not explicitly recorded.
Therefore, we list all nestboxes as functioning for the full study period.
}
