% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{format_UAN}
\alias{format_UAN}
\title{Construct standard summary for University of Antwerp populations}
\usage{
format_UAN(db = choose.dir(), Species = NULL, path = ".",
  debug = FALSE)
}
\arguments{
\item{db}{Directory where raw data files are stored.}

\item{Species}{A numeric vector. Which species should be included (EUring codes)? If blank will return all major species (see details below).}

\item{path}{Location where output csv files will be saved.}

\item{debug}{For internal use when editing pipelines. If TRUE, pipeline
generates a summary of pipeline data. This
includes: a) Histogram of continuous variables with mean/SD b) unique
values of all categorical variables.}
}
\value{
Generates 5 .csv files with data in a standard format.
}
\description{
Construct standard summary for University of Antwerp populations
}
