% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{format_UAN}
\alias{format_UAN}
\title{Construct standard summary for University of Antwerp populations}
\usage{
format_UAN(data_folder = choose.dir(), path = ".")
}
\arguments{
\item{data_folder}{Folder location where raw .csv files are stored.}

\item{path}{Location where output files will be saved.}
}
\value{
Generates 5 .csv files with data in a standard format.
}
\description{
Construct standard summary for University of Antwerp populations
}
