% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_UAN.R
\name{format_UAN}
\alias{format_UAN}
\title{Construct standard format for data from the Univeristy of Antwerp.}
\usage{
format_UAN(
  db = choose_directory(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
4 data tables in the standard format (version 1.1.0). When `output_type = "R"`, a list of 4 data frames corresponding to the 4 standard data tables and 1 character vector indicating the protocol version on which the pipeline is based. When `output_type = "csv"`, 4 .csv files corresponding to the 4 standard data tables and 1 text file indicating the protocol version on which the pipeline is based.
}
\description{
A pipeline to produce the standard format for 2 hole-nesting bird study
populations administered by the University of Antwerp.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.1.0.pdf}{here}.

\strong{ClutchType_observed}: The raw data distinguishes second and third
nests and first, second, and third replacements. We group these all as 'second'
and 'replacement' respectively.

\strong{ClutchSize_min, ClutchSize_max}: The raw data includes a column to determine whether
clutch size was counted with or without a brooding female. The presence of a
brooding female can effect the uncertainty in the count. After discussions
with the data owner clutch size counted with a brooding female is given an error
of +/- 2, so that ClutchSize_min = ClutchSize_observed - 2, and ClutchSize_max = ClutchSize_observed + 2.

\strong{ExperimentID}: Experimental codes are provided in their original
format. These still need to be translated into our experimental groups.

\strong{Tarsus}: Tarsus is measured using Svennson's Standard in early years
and Svennson's Alternative in later years. When Svennson's Alternative is
available this is used, otherwise we use converted Svensson's Standard, using
\code{\link[pipelines]{convert_tarsus}}.

\strong{Age}:
For Age_observed: \itemize{
\item If a capture has a recorded
ChickAge or the Capture Type is listed as 'chick' it is given a EURING code 1:
nestling or chick, unable to fly freely, still able to be caught by hand.
\item Recorded value 1 (first calendar year) is given a EURING code 3:
full-grown bird hatched in the breeding season of this calendar year.
\item Recorded value 2 (second calendar year) is given a EURING code 5:
a bird hatched last calendar year and now in its second calendar year.
\item Recorded value 3 (>first calendar year) is given a EURING code 4:
full-grown bird hatched before this calendar year;
year of hatching otherwise unknown.
\item Recorded value 4 (>second calendar year) is given a EURING code 6:
full-grown bird hatched before last calendar year; year of hatching otherwise
unknown.
\item After discussing with data owners, recorded value 5 (full grown age unknown)
is given NA.
}

For Age_calculated \itemize{
\item Any capture record with EURING <= 3 is considered to have a known age
(i.e. EURING codes 5, 7, 9, etc.). We consider identification in the nest
or as fledglings to be reliable indicators of year of hatching.
\item Any capture record with EURING >3 is considered to have an uncertain age
(i.e. EURING codes 4, 6, 8, etc.). We consider aging of adults to be too
uncertain.
}

\strong{Sex_observed, Sex_calculated}: Any uncertainty in sex is ignored. For example, 'M?' is treated as male.

\strong{ReleaseAlive}: Individuals who were captured alive are assumed to be released alive.

\strong{LocationID}: LocationIDs are formatted as <PopID>_<original LocationID> to avoid duplicates.

\strong{Latitude and Longitude}: Location data is stored in Lambert72 CRS.
This has been converted to WGS84 to be standard with other systems.
}
