% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_VEL.R
\name{format_VEL}
\alias{format_VEL}
\title{Construct standard summary for data from Velky Kosir, Czechia.}
\usage{
format_VEL(
  db = utils::choose.dir(),
  species = NULL,
  pop = NULL,
  path = ".",
  output_type = "R"
)
}
\arguments{
\item{db}{Location of database file.}

\item{species}{Species of interest. The 6 letter codes of all the species of
interest as listed in the
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. If blank will return all major species.}

\item{pop}{The three-letter code of population as listed in the \href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{standard
protocol}. For data owners with multiple populations (e.g. NIOO, UAN) where a single
pipeline is used for many populations this argument is used to extract data from
individual populations. For other pipelines that contain only one population
this argument can be ignored.}

\item{path}{Location where output csv files will be saved.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R').}
}
\value{
Generates 4 .csv files with data in a standard format.
}
\description{
A pipeline to produce a standard output for the nest box population in Velky
Kosir, Czechia, administered by Milos Krist.
}
\details{
This section provides details on data management choices that are unique to
this data. For a general description of the standard format please see see
\href{https://github.com/SPI-Birds/documentation/blob/master/standard_protocol/SPI_Birds_Protocol_v1.0.0.pdf}{here}.

\strong{Species}: There are a small number of records for species other than
great tit, blue tit and collared flycatcher. These are currently excluded.
Blue tits have 93 recorded broods. This is less than our previous 100 broods
cut-off, but they are still included.

\strong{ClutchType_calc}: We assume that any NA records for number fledged
are true unknowns (rather than 0s). In which case, we can't always estimate
whether a clutch is a second clutch.

\strong{ExperimentID}: Currently, we just copy the text directly from the
tables. Need to go through an classify each experiment and check with Milos
that these classifications are reasonable.

\strong{BroodID}: BroodID is currently Year_nestbox_day_month. This accounts
for multiple clutches laid in the same nest box.

\strong{Tarsus}: Tarsus length is originally measured using Oxford maximum
method. We convert this to Svensson's Alternative method using formula
described in the standard format.

\strong{ClutchSize, BroodSize, NumberFledged:} Some records have uncertainty
(e.g. 11+). We're unsure how large this uncertainty is. It's currently
ignored, but will talk with data owner to incorporate this.

\strong{CaptureDate:} In the raw data for flycatchers there are columns for
measurements taken at 6 days and at 13 days. Sometimes, no measurements are
recorded (i.e. columns are NA) and we need to decide how to deal with these.
In discussion with the data owner we know that chicks are generally ringed at 6 days;
therefore, we assume that any ringed chick was captured at 6 days old, even if
they have no other measurements. We only include a further capture at 13 days
old if some measurements were taken. Those individuals with no mass/tarsus/wing length
measures recorded at 13 days were assumed to have not been captured at this time.
}
