% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{individual_data_template}
\alias{individual_data_template}
\title{individual_data_template template}
\format{
Tibble
}
\description{
Tibble giving names and associated classes for all the columns in the Individual data for the standard protocol v1.1.0
}
