% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{location_check}
\alias{location_check}
\title{Perform quality checks on location data}
\usage{
location_check(
  Location_data,
  Brood_data,
  Capture_data,
  approved_list,
  output,
  skip,
  map
)
}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}

\item{Brood_data}{Data frame. Brood data output from pipeline.}

\item{Capture_data}{Data frame. Capture data output from pipeline.}

\item{approved_list}{List object. List of approved records from brood_approved_list.csv,
capture_approved_list.csv, individual_approved_list.csv, location_approved_list.csv.}

\item{output}{Character. Run checks on potential errors ("errors"), warnings ("warnings"), or both ("both"; default).}

\item{skip}{Character. Identifiers of the individual quality checks (CheckID) that should be skipped.}

\item{map}{Logical. Produce map of locations? See \code{\link{check_coordinates}}.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{WarningRows}{A vector of rows with warnings.}
\item{ErrorRows}{A vector of rows with errors.}
\item{Warnings}{A list of row-by-row warnings.}
\item{Errors}{A list of row-by-row errors.}
}
\description{
A wrapper that runs all single checks related to \code{Location_data}.
}
\details{
The following location data checks are performed:
\itemize{
\item \strong{L1}: Check location coordinates using \code{\link{check_coordinates}}.
}
}
