% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location_check.R
\name{location_check}
\alias{location_check}
\title{A wrapper function to perform quality checks on location data}
\usage{
location_check(Location_data)
}
\arguments{
\item{Location_data}{Data frame. Location data output from pipeline.}
}
\value{
A list of 3 items: a summary data frame of all checks, a record-by-record list of warnings, and a record-by-record list of errors.
}
\description{
A wrapper that runs all single checks related to `Location_data`.
}
\details{
\strong{Location check 1}: Check if the formats of each column in `Location_data` match with the standard format using \code{\link{check_format_location}}.
}
