% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{location_data_template}
\alias{location_data_template}
\title{Location data template in standard format (version 1.1.0)}
\format{
Data frame with 1 row and 9 variables:
}
\description{
Location data template in standard format (version 1.1.0)
}
