% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pop_codes}
\alias{pop_codes}
\title{Population names and locations}
\format{
A data frame with 26 rows and 4 variables
\describe{
 \item{PopID}{Three-letter population code.}
 \item{PopName}{Name of population.}
 \item{Country}{Name of country where population is situated.}
 \item{Owner}{Letter code for data owner. Note, multiple populations can have
 the same data owner.}
 \item{OwnerName}{Name of owner.}}
}
\usage{
pop_codes
}
\description{
Information and identifying codes for all populations and data owners.
}
\keyword{datasets}
