% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{pop_species_combos}
\alias{pop_species_combos}
\title{Combination of all PopID and Species from all pipelines.}
\format{
A data frame with 97 rows and 2 variables.
\describe{
   \item{PopCode}{Three letter population ID.}
   \item{SpeciesCode}{Six letter species ID.}
}
}
\usage{
pop_species_combos
}
\description{
A data frame with each row containing a unique population
and species combination.
}
\keyword{datasets}
