% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{A function to perform a quality check on pipeline outputs}
\usage{
quality_check(db = utils::choose.dir(), pop, species,
  output_format = "both")
}
\arguments{
\item{db}{Location of standard .csv outputs of the pipeline.}

\item{pop}{Three-letter population ID.}

\item{species}{Six-letter species ID.}

\item{output_format}{Format of report: "html", "pdf", or "both" (default).}
}
\value{
A summary dataframe of test warnings and errors for use in testthat,
  and a report of the line-by-line list of warnings and errors.
}
\description{
A function to perform a quality check on pipeline outputs
}
