% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{Perform a quality check on pipeline outputs}
\usage{
quality_check(
  R_data,
  output = TRUE,
  output_format = "both",
  check_format = TRUE
)
}
\arguments{
\item{R_data}{Output of pipeline as an R object. Generated using
\code{output_type = R} in \code{\link{run_pipelines}}.}

\item{output}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, a report is produced. Default: \code{TRUE}.}

\item{output_format}{A character. Format of output report. Options: \code{"html"}, \code{"pdf"}, or \code{"both"}. Default: \code{"both"}.}

\item{check_format}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, the checks on variable format (i.e. \code{\link{check_format_brood}}, \code{\link{check_format_capture}}, \code{\link{check_format_individual}} and \code{\link{check_format_location}}) are included in the quality check. Default: \code{TRUE}.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{NumberChecks}{Number of checks performed.}
\item{NumberWarnings}{Number of checks resulted in warnings.}
\item{NumberErrors}{Number of checks resulted in errors.}
\item{ElapsedTime}{Elapsed time in seconds.}
\item{R_data}{Pipeline output (a list of 4 dataframes) with Warning & Error columns marking the rows with warnings and errors.}

and a report (pdf, html or both) of row-by-row list of warnings and errors if \code{output_format = TRUE}.
}
\description{
A wrapper function that performs a quality check and produces a report.
}
\details{
Combines the four data frame-specific wrappers: \code{\link{brood_check}}, \code{\link{capture_check}}, \code{\link{individual_check}} and \code{\link{location_check}}.
}
\examples{
\dontrun{

CHO <- run_pipelines(PopID = "CHO", output_type = "R")
quality_check(CHO)

}
}
