% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{Perform a quality check on pipeline outputs}
\usage{
quality_check(
  R_data,
  output = TRUE,
  output_format = "both",
  output_file = "output_report",
  latex_engine = "lualatex",
  test = FALSE,
  map = TRUE
)
}
\arguments{
\item{R_data}{Output of pipeline as an R object. Generated using
\code{output_type = R} in \code{\link{run_pipelines}}.}

\item{output}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, a report is produced. Default: \code{TRUE}.}

\item{output_format}{Character. Format of output report. Options: "html", "pdf", and "both" (default).}

\item{output_file}{Character. Name of the output file. Default: "output_report".}

\item{latex_engine}{Character. LaTeX engine for producing PDF output. Options are "pdflatex", "xelatex", and "lualatex" (default). NB: pdfLaTeX and XeLaTeX have memory limit restrictions, which can be problematic when generating large pdfs. LuaLaTeX has dynamic memory management which may help for generating large pdfs.}

\item{test}{Logical. Is `quality_check` being used inside package tests? If TRUE, `R_data` is ignored and
dummy data will be used instead.}

\item{map}{Logical. Produce map of capture locations? See \code{\link{check_coordinates}}.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and errors.}
\item{NumberChecks}{Number of checks performed.}
\item{NumberWarnings}{Number of checks resulted in warnings.}
\item{NumberErrors}{Number of checks resulted in errors.}
\item{ElapsedTime}{Elapsed time in seconds.}
\item{R_data}{Pipeline output (a list of 4 dataframes) with Warning & Error columns marking the rows with warnings and errors.}

and a report (pdf, html or both) of row-by-row list of warnings and errors if \code{output_format = TRUE}.
}
\description{
A wrapper function that performs a quality check and produces a report.
}
\details{
Combines the four data frame-specific wrappers: \code{\link{brood_check}}, \code{\link{capture_check}}, \code{\link{individual_check}} and \code{\link{location_check}}.
}
\examples{
\dontrun{

CHO <- run_pipelines(PopID = "CHO", output_type = "R")
quality_check(CHO)

}
}
