% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{A function to perform a quality check on pipeline outputs}
\usage{
quality_check(db = choose.dir(), report = NULL)
}
\arguments{
\item{db}{Location of standard .csv outputs of the pipeline.}

\item{report}{Produce report (pdf) for user.}
}
\value{
A text file of line-by-line warnings,
  a text file of line-by-line errors,
  a summary dataframe of test warnings and errors for use in testthat,
  and a user-friendly report of the line-by-line list of warnings and errors.
}
\description{
A function to perform a quality check on pipeline outputs
}
