% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{Perform a quality check on pipeline outputs}
\usage{
quality_check(
  R_data,
  output = "both",
  report = TRUE,
  report_format = "both",
  report_file = "quality-check-report",
  latex_engine = "lualatex",
  test = FALSE,
  map = TRUE
)
}
\arguments{
\item{R_data}{Output of pipeline as an R object. Generated using
\code{output_type = R} in \code{\link{run_pipelines}}.}

\item{output}{Character. Run checks on and produce report of: potential errors ("errors"), warnings and verified records ("warnings"), or both in separate files ("both"; default).}

\item{report}{\code{TRUE} or \code{FALSE}. If \code{TRUE}, a report is produced. Default: \code{TRUE}.}

\item{report_format}{Character. Format of output report: "html", "pdf", or "both" (default).}

\item{report_file}{Character. Name of the output reports. Default: "quality-check-report". Note that to the file name of the report on potential errors the suffix "_errors" is added (i.e. "quality-check-report_errors"), and to the report on warnings and verified records "_warnings" (i.e., "quality-check-report_warnings").}

\item{latex_engine}{Character. LaTeX engine for producing PDF output. Options are "pdflatex", "xelatex", and "lualatex" (default). NB: pdfLaTeX and XeLaTeX have memory limit restrictions, which can be problematic when generating large pdfs. LuaLaTeX has dynamic memory management which may help for generating large pdfs.}

\item{test}{Logical. Is `quality_check` being used inside package tests? If  \code{TRUE}, `R_data` is ignored and
dummy data will be used instead.}

\item{map}{Logical. If  \code{TRUE}, a map of locations is added in the report. See \code{\link{check_coordinates}}.}
}
\value{
A list of:
\item{CheckList}{A summary dataframe of check warnings and potential errors.}
\item{NumberChecks}{Number of checks performed.}
\item{NumberWarnings}{Number of checks resulted in warnings.}
\item{NumberErrors}{Number of checks resulted in potential errors.}
\item{ElapsedTime}{Elapsed time in seconds.}
\item{R_data}{Pipeline output (a list of 4 dataframes) with Warning & Error columns marking the rows with warnings and errors.}

and reports (pdf, html or both) of potential errors and/or warnings if \code{report = TRUE}.
}
\description{
A wrapper function that performs a quality check and produces a report.
}
\details{
Combines the four data frame-specific wrappers: \code{\link{brood_check}}, \code{\link{capture_check}}, \code{\link{individual_check}} and \code{\link{location_check}}.
}
\examples{
\dontrun{

CHO <- run_pipelines(site = "CHO", output_type = "R")
CHO_checked <- quality_check(R_data = CHO)

}
}
