% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality_check.R
\name{quality_check}
\alias{quality_check}
\title{A function to perform a quality check on the pipeline outputs}
\usage{
quality_check(db = choose.dir(), report = NULL)
}
\arguments{
\item{db}{Location of standard .csv outputs of pipeline.}

\item{report}{Produce report (pdf) for user.}
}
\value{
dataframe with row-by-row warnings and errors,
  dataframe with successful/failed checks/tests for use in testthat, user report.
}
\description{
A function to perform a quality check on the pipeline outputs
}
