% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_MAY.R
\name{retrieve_chickIDs_MAY}
\alias{retrieve_chickIDs_MAY}
\title{Retrieve chick IDs in MAY pipeline}
\usage{
retrieve_chickIDs_MAY(chickID)
}
\arguments{
\item{chickID}{Character. The series of partially incomplete chick IDs of a brood.}
}
\value{
A vector with the complete chick IDs of a brood, or NA.
}
\description{
In MAY primary data, the chick IDs in a brood are stored as series of partially incomplete character sequences (e.g., "856840,1,55-62", "099362-65"). This function extracts the full sequence of characters for each ID in the series. "-" are interpreted as a range; "," are interpreted as a regular separator. Values in other formats (e.g., "without a rings", "531094.95999999996") and sequences that lead to an excessive number of IDs (e.g. "54522-291") are set to NA.
}
\examples{

retrieve_chickIDs_MAY("856840,1,55-62")

}
