% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipelines.R
\name{run_pipelines}
\alias{run_pipelines}
\title{Run multiple data pipelines}
\usage{
run_pipelines(
  path = choose_directory(),
  site = NULL,
  species = NULL,
  output_type = "R",
  save = FALSE,
  save_path = NULL,
  filename = "standard_format"
)
}
\arguments{
\item{path}{File path. Location of all folders for data owners. Note, the folders for each data owner must include the unique code of the data owner as seen in `site_codes`.}

\item{site}{The three-letter code of sites to format as listed in `site_codes`.}

\item{species}{The six-letter code of species to include as listed in `species_codes`. Note, this argument takes precedence over argument siteID (i.e., if a site doesn't have the requested species it will not be formatted.)}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R'). Default: R.}

\item{save}{TRUE/FALSE. Should the output be saved locally? This is only relevant where
`output_type` is 'R'. If output_type is 'csv'
4 .csv files will be created in the save path (specified by `save_path` argument). If output_type is 'R'
and `save` is TRUE, an .RDS file will be created in the save path.}

\item{save_path}{Path where files will be saved if `save` is TRUE. By default, the save
path will be path/standard_format.}

\item{filename}{The file name of the saved file. No file extension is
needed, as this will differ depending on `output_type`. By default, file name is
"standard_format".}
}
\value{
Generate .csv files or return an R list object with 4 items
}
\description{
Run multiple data pipelines. Currently, this produces separate .csv files but will eventually produce combined .csv files for all sites.
}
\examples{
\dontrun{

#Create an R list of Harjavalta data
HAR_data <- run_pipelines(site = "HAR", species = "PARMAJ", output_type = "R")

}
}
