% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipelines.R
\name{run_pipelines}
\alias{run_pipelines}
\title{Run multiple data pipelines}
\usage{
run_pipelines(
  path = choose_directory(),
  siteID = NULL,
  speciesID = NULL,
  optional_variables = NULL,
  output_type = "R",
  save = FALSE,
  save_path = NULL,
  filename = "standard_format"
)
}
\arguments{
\item{path}{File path. Location of all folders for data owners. Note, the folders for each data owner must include the unique code of the data owner as seen in `site_codes`.}

\item{siteID}{The three-letter code of sites to format as listed in `site_codes`.}

\item{speciesID}{The six-letter code of species to include as listed in `species_codes`. Note, this argument takes precedence over argument siteID (i.e., if a site doesn't have the requested species it will not be formatted.)}

\item{optional_variables}{Optional variables (generated by standard utility functions) to be included in the pipeline output. Either a list of variable names or 'all' (if all optional variables should be included). See Details for a full list of optional variables.}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R'). Default: R.}

\item{save}{TRUE/FALSE. Should the output be saved locally? This is only relevant where
`output_type` is 'R'. If output_type is 'csv'
4 .csv files will be created in the save path (specified by `save_path` argument). If output_type is 'R'
and `save` is TRUE, an .RDS file will be created in the save path.}

\item{save_path}{Path where files will be saved if `save` is TRUE. By default, the save
path will be path/standard_format.}

\item{filename}{The file name of the saved file. No file extension is
needed, as this will differ depending on `output_type`. By default, file name is
"standard_format".}
}
\value{
Generate .csv files or return an R list object with 4 items (version <=1.1) or 6 (version >=1.2) items.
}
\description{
Run multiple data pipelines. Currently, this produces separate .csv files but will eventually produce combined .csv files for all sites.
}
\details{
Optional variables generated by standard utility functions:
\itemize{
\item \code{breedingSeason} [Brood_data]. The breeding season in which the brood was laid. See \code{\link{calc_season}} for more details.
\item \code{calculatedClutchType} [Brood_data]. The type of clutch as classified by strict decision rules. See \code{\link{calc_clutchtype}} for more details.
\item \code{nestAttemptNumber} [Brood_data]. The sequence of nesting attempts by a breeding pair in a season. See \code{\link{calc_nestattempt}} for more details.
\item \code{calculatedSex} [Individual_data]. The sex of an individual concluded from 'observedSex' in Capture_data. See \code{\link{calc_sex}} for more details.
\item \code{exactAge} [Capture_data]. The exact age of the bird in number of years since birth. Exact age can only be determined for individuals ringed as chicks. See \code{\link{calc_age}} for more details.
\item \code{minimumAge} [Capture_data]. The minimum age of the bird in number of years since birth/ringing. Minimum age can be determined for all individuals. See \code{\link{calc_age}} for more details.
}
}
\examples{
\dontrun{

#Create an R list of Harjavalta data
HAR_data <- run_pipelines(siteID = "HAR", speciesID = "PARMAJ", output_type = "R")

}
}
