% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pipelines.R
\name{run_pipelines}
\alias{run_pipelines}
\title{Run multiple data pipelines}
\usage{
run_pipelines(path = utils::choose.dir(), PopID = NULL,
  Species = NULL, output_type = "R")
}
\arguments{
\item{path}{File path. Location of all folders for data owners. Note, the
folders for each data owner must include the unique code of the data owner
as seen in pop_names.}

\item{PopID}{The three-letter code of populations to format as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
 protocol}.}

\item{Species}{The six-letter code of species to include as listed in the
\href{https://github.com/LiamDBailey/SPIBirds_Newsletter/blob/master/SPI_Birds_Protocol_v1.0.0.pdf}{standard
 protocol}. Note, this argument takes precedence over argument PopID (i.e.
if a population doesn't have the requested species it will not be
formatted.)}

\item{output_type}{Should the pipeline generate .csv files ('csv') or R objects ('R'). Default: R.}
}
\value{
Generate .csv files or return an R list object with 4 items
}
\description{
Run multiple data pipelines. Currently, this produces separate .csv files but
will eventually produce combined .csv files for all populations.
}
\examples{
\dontrun{

#Create an R list of Harjavalta data
HAR_data <- run_pipelines(PopID = "HAR", Species = "PARMAJ", output_type = "R")

}
}
