% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{Species EURING codes and letter codes}
\format{
A data frame with 10 rows and 4 variables
\describe{
  \item{SpeciesID}{Species EUring code.}
  \item{Species}{Species letter code.}
  \item{CommonName}{Common English name.}
  \item{BinomialName}{Genus and species name.}
  }
}
\usage{
species_codes
}
\description{
Letter codes to easily identify species (English names)
}
\keyword{datasets}
