% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{species_codes}
\alias{species_codes}
\title{Species taxonomic ranks and codes}
\format{
A data frame with 27 rows and 13 variables
\describe{
  \item{speciesCode}{SPI-Birds' internal persistent identifier for a species.}
  \item{speciesID}{SPI-Birds' 6-letter species identifier. First three letters indicate the generic name, last three letters indicate the specific epithet. Note that this identifier might change if a species' genus or species indicator change.}
  \item{speciesEURINGCode}{Species code used in EURING. NA for non-European species.}
  \item{kingdom}{Scientific name of the kingdom in which the species is identified.}
  \item{phylum}{Scientific name of the phylum in which the species is identified.}
  \item{class}{Scientific name of the class in which the species is identified.}
  \item{order}{Scientific name of the order in which the species is identified.}
  \item{family}{Scientific name of the family in which the species is identified.}
  \item{genus}{Scientific name of the genus in which the species is identified.}
  \item{specificEpithet}{Scientific name of the species epithet.}
  \item{infraspecificEpithet}{Scientific name of the infraspecific epithet (e.g. sub-species).}
  \item{scientificNameAuthorship}{Authorship information of the scientific name, including date information if known.}
  \item{vernacularName}{Common English name.}
  }
}
\usage{
species_codes
}
\description{
Species information, including various taxonomic ranks, internal and external codes.
}
\keyword{datasets}
