% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_datarqst.R
\name{subset_datarqst}
\alias{subset_datarqst}
\title{Create a subset of SPI-Birds standard data}
\usage{
subset_datarqst(
  file = file.choose(),
  PopID = unique(pop_names$code),
  Species = unique(Species_codes$Code),
  filter = NULL
)
}
\arguments{
\item{file}{The location of the file in the standard format (.RDS file). This file is created using `run_pipelines`.}

\item{PopID}{Character vector of three letter population codes. Include all populations that are requested.}

\item{Species}{Character vector of six letter species codes. Include all species that are requested.}

\item{filter}{Character vector of unique population species combinations (in the format PopID_Species).
Include all unique population species combinations requested.}
}
\value{
Create 4 .csv files at the location of the standard data. Files will be given a unique number
based on the date when the subset was performed.
}
\description{
Create a subset of data stored in the standard format.
}
\details{
The SPI-Birds team should use the function `run_pipelines`
to create a copy of all populations in the standard format,
which can be stored locally on the SPI-Birds computer.

When a request is made, `subset_datarqst` can be used
to take a subset of this larger file to cover the specific
populations/species requested by a user.
}
\examples{
\dontrun{

#Subset data on all species from two populations
subset_datarqst(PopID = c("HOG", "CHO"))

#Subset data on all population where two species occurs
subset_datarqst(Species = c("PARMAJ", "FICALB"))

#Create a specific subset of species and population combinations
subset_datarqst(filter = c("HOG_PARMAJ", "VEL_FICALB", "VLI_PARMAJ"))

}
}
