% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_general_format.R
\name{test_category_columns}
\alias{test_category_columns}
\title{Test category columns}
\usage{
test_category_columns(pipeline_output, table)
}
\arguments{
\item{pipeline_output}{A list of the 4 data frames returned from format_X
(where X is the three letter study site code).}

\item{table}{Which table should be checked? One of: "Brood", "Capture", "Individual", or "Location"}
}
\value{
Logical (TRUE/FALSE) for whether there are non-standard categories in the data frame.
}
\description{
Some columns should only include particular categorical values.
For example, 'observedClutchType' should only include any of "first", "second", "replacement" or NA.
This test makes sure that there are no unexpected values in these different category columns.
}
