% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_general_format.R
\name{test_ID_format}
\alias{test_ID_format}
\title{Test ID format}
\usage{
test_ID_format(pipeline_output, ID_col, ID_format)
}
\arguments{
\item{pipeline_output}{A list of the 4 data frames returned from format_X
(where X is the three letter study site code).}

\item{ID_col}{Character string. Which ID column should be tested? ("femaleID", "maleID", "C-individualID", "I-individualID").
The C and I in the final two options stand for Capture and Individual data, respectively. So, to test individualID in the Capture
data, ID_col should be set to "C-individualID".}

\item{ID_format}{Character string. Regular expression corresponding to the proper ID format.}
}
\value{
Logical (TRUE/FALSE) for whether all of IDs are formatted as expected.
}
\description{
Test that ID formats for each column are expected. There are 4 ID columns. "individualID" in the Capture and
Individual data, and "femaleID" and "maleID" in the Brood data. This function will take the regular expression that matches
the expected ID format for the pipeline and will make sure that the ID column does not have entries that do not match the expected format.
}
\details{
This function is intended to be used in the test for every pipeline in order to ensure consistency between pipelines.
}
