% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_general_format.R
\name{test_unique_values}
\alias{test_unique_values}
\title{Test unique values in key columns}
\usage{
test_unique_values(pipeline_output, column)
}
\arguments{
\item{pipeline_output}{A list of the 4 data frames returned from format_X
(where X is the three letter population code).}

\item{column}{Character string. Which column should be checked for duplicates? One of: "BroodID", "CaptureID", or "PopID-IndvID"}
}
\value{
Logical (TRUE/FALSE) for whether all values in the column are unique
}
\description{
Some columns are not allowed to have duplicated values. Namely, BroodID (Brood data), CaptureID (Capture data),
and the concatenation of PopID and IndvID (Individual data).
This function takes a column name and returns a logical value corresponding to whether all values in the column are unique (TRUE) or not (FALSE).
}
\details{
This function is intended to be used in the test for every pipeline in order to ensure consistency between pipelines.
}
